% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{add-score}
\alias{add-score}
\alias{AddMiscIntegrations}
\alias{AddMiscScores}
\title{Add integration(s) or score(s) slot(s) to the score tibble}
\usage{
AddMiscIntegrations(object, which)

AddMiscScores(object, which, class = "numeric")
}
\arguments{
\item{object}{a Seurat object}

\item{which}{the name of the new integration(s) or score(s)}

\item{class}{the class of the column, such as 'numeric' (default). For
instance, cell cycle conservation scores are 'list' and LISI scores are
'numeric_lisi'.}
}
\value{
the Seurat object with updated table of scores.
}
\description{
\code{AddMiscIntegrations}: Add integration(s) slot(s) to the score tibble
(or make sure they are in it). This enables to save their scores later.

\code{AddMiscScores}: Add score(s) slot(s) to the score tibble (or make sure
they are in it). This enables to save them later.
}
