% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Add two numeric vectors}
\usage{
add(x, y)
}
\arguments{
\item{x}{A numeric vector}

\item{y}{Another numeric vector}
}
\value{
A numeric vector of the same length of the longest among `x` and `y`.
}
\description{
Add two numeric vectors
}
\examples{
add(1, 2)
}
