% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SumFile.R
\name{AddNum}
\alias{AddNum}
\title{This function adds two numbers and returns the result}
\usage{
AddNum(x, y)
}
\arguments{
\item{x}{numeric variable}

\item{y}{numeric variable}
}
\value{
return numeric variable
}
\description{
This function adds two numbers and returns the result
}
