% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_package.R
\name{addPkgDepencies}
\alias{addPkgDepencies}
\title{function to add a list of packages dependencies}
\usage{
addPkgDepencies(dependencies = "data.table")
}
\arguments{
\item{dependencies}{a vector of packages on which the package to be built depends on}
}
\description{
function to add a list of packages dependencies
}
\examples{
addPkgDepencies(dependencies=c("data.table","dplyr"))
}
