% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sum.R
\name{addSum}
\alias{addSum}
\title{This function is used to add 2 numbers}
\usage{
addSum(x, y)
}
\arguments{
\item{x}{numeric variable}

\item{y}{numeric variable}
}
\value{
result numeric variable
}
\description{
This function is used to add 2 numbers
}
