% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_a_session_setup.R
\name{analysis_a_session_setup}
\alias{analysis_a_session_setup}
\title{analysis_a_session_setup}
\usage{
analysis_a_session_setup(id = "analysis_a_session_setup", user, is_admin)
}
\description{
analysis_a_session_setup
}
