% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_a_setup.R
\name{analysis_a_setup_server}
\alias{analysis_a_setup_server}
\title{analysis_a_setup_server}
\usage{
analysis_a_setup_server(input, output, session, user, is_admin, signal)
}
\description{
analysis_a_setup_server
}
