% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self_res_att.R
\name{apa_t_pair}
\alias{apa_t_pair}
\title{APA text for Paired-Samples T-Test}
\usage{
apa_t_pair(x, y, dv = "the DV", level1 = "level 1", level2 = "level 2")
}
\arguments{
\item{x}{A vector of the values for level 1.}

\item{y}{A vector of the values for level 2.}

\item{dv}{The text describing the DV in the output statement.}

\item{level1}{The text describing level 1 in the output statement.}

\item{level2}{The text describing level 2 in the output statement.}
}
\value{
A character string.
}
\description{
Create APA-formatted text for the results of an independent-samples t-test in the following format:

A paired-samples t-test was conducted to compare \{dv\} between \{level1\} (M = \{mean1\}, SD = \{sd1\}) and \{level2\} (M = \{mean2\}, SD = \{sd2\}). There was a \{non\}significant difference; t(\{df\}) = \{t_value\}, p = \{p_value\}.
}
\examples{
# use generic text
apa_t_pair(x = self_res_att$f_self,
           y = self_res_att$f_non)

# specify the text for dv and levels
apa_t_pair(x = self_res_att$f_self,
           y = self_res_att$f_non,
           dv = "preferences for female faces",
           level1 = "participants who resembled those faces",
           level2 = "non-self participants")
}
