% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atri_crf.R
\name{atri_crf}
\alias{atri_crf}
\title{Read ATRI CRF data}
\usage{
atri_crf(protocol, crf, datalake = TRUE, wide = TRUE, translated = TRUE)
}
\arguments{
\item{protocol}{The protocol name (e.g. 'a345-test-1').}

\item{crf}{The table name (e.g. 'registry').}

\item{datalake}{default is TRUE, which returns data from the latest data extract from the data lake s3 archive topic. 
With datalake=FALSE, returns the current data using the public API (/public/api/v1/crfs/data/<crf>).}

\item{wide}{default is TRUE, which returns data with one record per row (one column per field). With wide=FALSE, 
returns data in the raw, long format (one row per field).}

\item{translated}{default is TRUE, which returns translated data. Only applicable if wide=TRUE}
}
\value{
A data.frame of the crf data.
}
\description{
Read ATRI CRF data
}
\examples{
\donttest{
   registry_wide <- atri_crf("a345-test-1", "registry")
   registry_long <- atri_crf("a345-test-1", "registry",wide=F)
   registry_untranslated <- atri_crf("a345-test-1", "registry",translated=F)
   registry_live <- atri_crf("a345-test-1", "registry",datalake=F)
}
}
\seealso{
\code{\link{atri_api}}
}
