% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atri_query_data.R
\name{atri_query_data}
\alias{atri_query_data}
\title{Read ATRI query data}
\usage{
atri_query_data(protocol, collapsed = TRUE, datalake = TRUE)
}
\arguments{
\item{protocol}{The protocol name (e.g. 'a345-test-1').}

\item{collapsed}{default is TRUE, which returns the most recent transaction of the query ID (only 1 row per query ID). With collapsed=FALSE, returns all historical query transactions.}

\item{datalake}{default is TRUE, which returns data from the latest data extract from the data lake s3 archive topic. 
With datalake=FALSE, returns the current data using the public API (/public/api/v1/query/queries).}
}
\value{
A data.frame of the query data.
}
\description{
Read ATRI query data
}
\examples{
\donttest{
   query_data <- atri_query_data("a345-test-1")
   query_data_all_transactions <- atri_query_data("a345-test-1", collapsed=FALSE)
   query_data_live <- atri_query_data("a345-test-1",datalake=FALSE)
}
}
\seealso{
\code{\link{atri_api}}
}
