% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atri_study_data.R
\name{atri_study_data}
\alias{atri_study_data}
\title{Read ATRI study data}
\usage{
atri_study_data(protocol, data, datalake = TRUE)
}
\arguments{
\item{protocol}{The protocol name (e.g. 'a345-test-1').}

\item{data}{The study data source name. See below for examples for each data source. Additionally, allowable text input can be the data export label (e.g.'Participant list'), the relative public api pathway (e.g. '/subjects'), or the data lake filename (e.g. 'participant_list.csv' or 'participant_list').}

\item{datalake}{default is TRUE, which returns data from the latest data extract from the data lake s3 archive topic. 
With datalake=FALSE, returns the current data using the public API (/public/api/v1/<relative API path>).}
}
\value{
A data.frame of the study data.
}
\description{
Read ATRI study data
}
\examples{
\donttest{
#--Participant data--#
#Participant list
subjects<-atri_study_data(protocol="a345-test-1",data="subjects")
#Participant Attribute Value list
subject_attr<-atri_study_data(protocol="a345-test-1",data="subject_attr")
#Participant Attribute Datadic list
subject_attr_dd<-atri_study_data(protocol="a345-test-1",data="subject_attr_dd")

#--Site data--#
#Site list
sites<-atri_study_data("a345-test-1","sites")
#Site Attribute Value list
site_attr<-atri_study_data("a345-test-1","site_attr")
#Site Attribute Datadic list
site_attr_dd<-atri_study_data("a345-test-1","site_attr_dd")

#--Study attribute data--#
#Study Attribute Datadic list
study_attr_dd<-atri_study_data("a345-test-1","study_attr_dd")

#--CRF metadata--#
#eCRF list
crfs<-atri_study_data("a345-test-1","crfs")
#Monitor Review list
monitor_review<-atri_study_data("a345-test-1","monitor_review")
#Form Lock list
form_lock<-atri_study_data("a345-test-1","form_lock")

#--Event/Schedule of events data--#
#Event list
events<-atri_study_data("a345-test-1","events")
#Event Attribute Datadic list
event_attr_dd<-atri_study_data("a345-test-1","event_attr_dd")
#Schedule of Events
tracks<-atri_study_data("a345-test-1","tracks")
#Subject Track list
subject_track<-atri_study_data("a345-test-1","subject_track")

#--Schedule of events by track code--#
#If datalake=TRUE, must use datalake filelabel or filename
soe_a3_default<-atri_study_data("a345-test-1","schedule_of_events_a3_default")
soe_a4_default<-atri_study_data("a345-test-1","schedule_of_events_a4_default.csv")
#If datalake=FALSE, must use API relative path
soe_a3_default<-atri_study_data("a345-test-1","/tracks/a3_default")

#--Alternate examples for 'data' parameter input--#
#Data export label
subject_track.1<-atri_study_data("a345-test-1",data="Subject Track list")
#Relative API path
subject_track.2<-atri_study_data("a345-test-1",data="/subject/track/list")
#Data lake filename
subject_track.3<-atri_study_data("a345-test-1",data="subject_track_list.csv")
subject_track.4<-atri_study_data("a345-test-1",data="subject_track_list")

#--Live data pull--#
subject_track_live<-atri_study_data("a345-test-1","subject_track", datalake=F)
}
}
\seealso{
\code{\link{atri_api}}
}
