% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbknn.R
\name{bbknnIntegration}
\alias{bbknnIntegration}
\title{Run bbknn on Seurat's \link[SeuratObject]{Assay5} object through \code{\link[Seurat]{IntegrateLayers}}}
\usage{
bbknnIntegration(
  object,
  orig,
  groups = NULL,
  groups.name = NULL,
  layers = "data",
  scale.layer = "scale.data",
  conda_env = NULL,
  new.graph = "bbknn",
  new.reduction = "pca.bbknn",
  reduction.key = "bbknnPCA_",
  reconstructed.assay = "bbknn.ridge",
  ndims = 50L,
  ndims.use = 30L,
  ridge_regression = T,
  graph.use = c("connectivities", "distances"),
  verbose = TRUE,
  seed.use = 42L,
  ...
)
}
\arguments{
\item{object}{A \code{\link[SeuratObject]{Seurat}} object
(or an \code{\link[SeuratObject]{Assay5}} object if
not called by \code{\link[Seurat]{IntegrateLayers}})}

\item{orig}{\code{\link[SeuratObject]{DimReduc}} object. Not to be set
directly when called with \code{\link[Seurat]{IntegrateLayers}}, use
\code{orig.reduction} argument instead}

\item{groups}{A \bold{named} data frame with grouping information. Preferably
one-column when \code{groups.name = NULL}}

\item{groups.name}{Column name from \code{groups} data frame that stores
grouping information. If \code{groups.name = NULL}, the first column is used}

\item{layers}{Name of the layers to use in the integration}

\item{conda_env}{Path to conda environment to run bbknn (should also
contain the scipy python module).  By default, uses the conda environment
registered for bbknn in the conda environment manager}

\item{new.graph}{Name of the \link[SeuratObject:Graph-class]{Graph object}}

\item{new.reduction}{Name of the new integrated dimensional reduction}

\item{reduction.key}{Key for the new integrated dimensional reduction}

\item{reconstructed.assay}{Name for the \code{assay} containing the corrected
expression matrix}

\item{ndims}{Number of dimensions for the new PCA computed on first output of
bbknn. 50 by default. Ignored when \code{ridge_regression = FALSE}}

\item{ndims.use}{Number of dimensions from \code{orig} to use for bbknn, and
from newly computed PCA when \code{ridge_regression = TRUE}.}

\item{ridge_regression}{When set to \code{TRUE} (default), new clusters are
computed on the output of bbknn, then a ridge regression is performed to
remove technical variables while preserving biological variables. Then, a new
bbknn run is performed.}

\item{graph.use}{Which graph of bbknn to output. One of "\code{connectivities}"
(default, recommended) or "\code{distances}"}

\item{verbose}{Print messages. Set to \code{FALSE} to disable}

\item{seed.use}{An integer to generate reproducible outputs.
Set \code{seed.use = NULL} to disable}

\item{...}{Additional arguments to be passed to \code{bbknn.bbknn()}. When
\code{ridge_regression = TRUE}, also accepts arguments to pass to
\code{Seurat::FindClusters()}, \code{Seurat::RunPCA()} and
\code{bbknn.ridge_regression()}. See \strong{Details} section}
}
\value{
A list containing at least one of:
\itemize{
  \item a new Graph of name [\code{new_graph}]_scale.data corresponding to
  the output of the first run of \pkg{bbknn}
  \item a new Assay of name \code{reconstructed.assay} with corrected counts
  for each feature from \code{scale.layer}.
  \item a new DimReduc (PCA) of name \code{new.reduction} (key set to
  \code{reduction.key})
  \item a new Graph of name [\code{new_graph}]_ridge.residuals corresponding
  to the output of the first run of \pkg{bbknn}
}
When called via \code{\link[Seurat]{IntegrateLayers}}, a Seurat object with
the new reduction and/or assay is returned
}
\description{
A wrapper to run \code{bbknn} on multi-layered Seurat V5 object.
Requires a conda environment with \code{bbknn} and necessary dependencies
}
\details{
This wrappers calls three python functions through \pkg{reticulate}.
Find the \pkg{bbknn}-specific arguments there:
\itemize{
  \item{bbknn function:} {
  \href{https://bbknn.readthedocs.io/en/latest/bbknn.bbknn.html}{bbknn.bbknn}, which relies on
  \href{https://bbknn.readthedocs.io/en/latest/bbknn.matrix.bbknn.html}{bbknn.matrix.bbknn}}
  \item{ridge regression:} {
  \href{https://bbknn.readthedocs.io/en/latest/bbknn.ridge_regression.html}{bbknn.ridge_regression}, which relies on
  \href{https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.Ridge.html}{sklearn.linear_model.Ridge}}
}
}
\note{
This function requires the
\href{https://bbknn.readthedocs.io/en/latest/index.html}{\pkg{bbknn}} package
to be installed (along with \pkg{scipy})
}
\examples{
\dontrun{
# Preprocessing
obj <- SeuratData::LoadData("pbmcsca")
obj[["RNA"]] <- split(obj[["RNA"]], f = obj$Method)
obj <- NormalizeData(obj)
obj <- FindVariableFeatures(obj)
obj <- ScaleData(obj)
obj <- RunPCA(obj)

# After preprocessing, we integrate layers:
obj <- IntegrateLayers(object = obj, method = bbknnIntegration,
                       conda_env = 'bbknn', groups = obj[[]],
                       groups.name = 'Method')

# To disable the ridge regression and subsequent steps:
obj <- IntegrateLayers(object = obj, method = bbknnIntegration,
                       conda_env = 'bbknn', groups = obj[[]],
                       groups.name = 'Method', ridge_regression = FALSE)
}

}
\references{
Polański, K., Young, M. D., Miao, Z., Meyer, K. B., Teichmann,
S. A. & Park, J.-E. BBKNN: fast batch alignment of single cell transcriptomes.
Bioinformatics 36, 964–965 (2019).
\href{https://doi.org/10.1093/bioinformatics/btz625}{DOI}
}
\seealso{
\code{\link[Seurat]{IntegrateLayers}}, \code{\link[Seurat]{writing-integration}}
}
