% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testR.R
\name{benchmarks}
\alias{benchmarks}
\title{Benchmark R and Rcpp functions.}
\description{
Use R package \code{microbenchmark} to compare the performance of C functions (\code{rw.Metropolis}) and Cpp functions (\code{crw}).
}
\examples{
\dontrun{
tm1 <- microbenchmark::microbenchmark(
rw3 = rw.Metropolis(2,25,2000),
crw3 = crw(2,25,2000)
)
print(summary(tm1)[,c(1,3,5,6)])
}
}
