% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bird_function.R
\name{bird_function}
\alias{bird_function}
\title{A bird function}
\usage{
bird_function(love = TRUE)
}
\arguments{
\item{love}{Do you love birds? Defaults to TRUE.}
}
\description{
A bird function
}
\examples{
bird_function()
}
\keyword{cats}
