% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_package.R
\name{buildPkgdownWebsite}
\alias{buildPkgdownWebsite}
\title{function to build the pkgdown website}
\usage{
buildPkgdownWebsite()
}
\description{
function to build the pkgdown website
}
\examples{
buildPkgdownWebsite()
}
