% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_phylo_profile.R
\name{calc_pres_spec}
\alias{calc_pres_spec}
\title{Calculate percentage of present species in each ortholog group}
\usage{
calc_pres_spec(profile_with_tax, taxa_count)
}
\arguments{
\item{profile_with_tax}{long data frame of main phyloprofile input together
with their taxonomy info}

\item{taxa_count}{number of species occur in each supertaxon}
}
\value{
A data frame with % of present species for each seed protein in
each supertaxon
}
\description{
Calculate percentage of present species in each ortholog group
}
\examples{
# NOTE: for internal testing only - not recommended for outside using
data("profile_with_taxonomy", package="phyloprofile")
taxa_count <- plyr::count(profile_with_taxonomy, "supertaxon")
taxa_count$freq <- 1
calc_pres_spec(profile_with_taxonomy, taxa_count)
}
\seealso{
\code{\link{profile_with_taxonomy}} for a demo input data
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
