% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_taxa_groups.R
\name{calculate_p_value}
\alias{calculate_p_value}
\title{calculate the p_values}
\usage{
calculate_p_value(var_in, var_out, significance_level)
}
\arguments{
\item{var_in}{list of values for the variable concerning the in-group}

\item{var_out}{list of values for the variable concerning the out-group}

\item{significance_level}{significant cutoff for statistical test}
}
\value{
return the pvalues
}
\description{
calculate the p_values
}
\author{
Carla Mölbert (carla.moelbert@gmx.de)
}
