% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_main_input.R
\name{check_input_validity}
\alias{check_input_validity}
\title{Check the validity of the main input file}
\usage{
check_input_validity(filein)
}
\arguments{
\item{filein}{input file}
}
\value{
The format of the input file format, or type of error
}
\description{
Check if input file has one of the following format: orthoXML,
multiple FASTA, wide or long matrix, or a list of OMA IDs.
}
\examples{
filein <- system.file(
    "extdata", "test.main.wide", package = "phyloprofile", mustWork = TRUE
)
check_input_validity(filein)
}
\seealso{
\code{\link{check_oma_id}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
