% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oma_browser.R
\name{check_oma_id}
\alias{check_oma_id}
\title{Check OMA IDs}
\usage{
check_oma_id(ids)
}
\arguments{
\item{ids}{list of ids needs to be checked}
}
\value{
list of invalid IDs (not readable for OMA)
}
\description{
Check if input IDs are valid IDs for OMA Browser
(either OMA IDs or UniProt IDs)
}
\examples{
check_oma_id("HUMAN29398")
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
