% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_S4_methods.R
\name{checkCondaBin}
\alias{checkCondaBin}
\alias{checkCondaEnvPath}
\alias{checkCondaEnvName}
\alias{checkCondaEnv}
\title{Check the validity of conda environment's components}
\usage{
checkCondaBin(x, ..., verbose = getOption("verbose"))

checkCondaEnvPath(x, ..., verbose = getOption("verbose"))

checkCondaEnvName(x, ..., verbose = getOption("verbose"))

checkCondaEnv(x, ..., verbose = getOption("verbose"))
}
\arguments{
\item{x}{a \linkS4class{CondaEnv} object or a sub element of it of class
\code{CondaEnvSlot}. \code{checkCondaEnv()} only accepts a CondaEnv object}

\item{...}{ignored when \code{x} is a \linkS4class{CondaEnv} object.
Otherwise, additional arguments to \pkg{reticulate} functions in
\strong{Description} such as \code{conda = ...}}

\item{verbose}{ignored}
}
\value{
an updated object of same class as input \code{x}
}
\description{
Check the validity of conda environment's components. Best used on a
\linkS4class{CondaEnv} object. Internally call:
\itemize{
  \item \strong{checkCondaBin}: \link[reticulate]{conda_binary}
  \item \strong{checkCondaEnvPath}: \link[reticulate]{condaenv_exists}
  \item \strong{checkCondaEnvName}: \link[reticulate]{condaenv_exists}
  \item \strong{checkCondaEnv}: both
}
}
\note{
\code{checkCondaEnv} is a wrapper that calls all checks.
}
\seealso{
\link[=isValid]{isValid()}, \linkS4class{CondaEnv}
}
