% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkinput.R
\name{checkinput}
\alias{checkinput}
\title{This can be run to test if the command input of parameters contradict to each other
And also check the structure of the data
If something goes wrong warning messages are given}
\usage{
checkinput(
  X,
  Y,
  ML,
  DA,
  method,
  fitness,
  nInner,
  nOuter,
  varRatio,
  scale,
  modReturn,
  logg,
  parallel
)
}
\arguments{
\item{X}{The original data of X, not the result after onehotencoding}

\item{fitness}{}

\item{parallel}{}
}
\value{
correct_input: the original input(call) and the real input used in MUVR when you enter your input
}
\description{
This can be run to test if the command input of parameters contradict to each other
And also check the structure of the data
If something goes wrong warning messages are given
}
