% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_profile.R
\name{cluster_data_dend}
\alias{cluster_data_dend}
\title{Create a dendrogram tree from the distance matrix}
\usage{
cluster_data_dend(distance_matrix, cluster_method)
}
\arguments{
\item{distance_matrix}{calculated distance matrix}

\item{cluster_method}{clustering method ("single", "complete", 
"average" for UPGMA, "mcquitty" for WPGMA, "median" for WPGMC, 
or "centroid" for UPGMC)}
}
\value{
A dendrogram tree object
}
\description{
Create a dendrogram tree from the distance matrix
}
\examples{
data("full_processed_profile_large", package="phyloprofile")
data <- full_processed_profile_large
profile_type <- "binary"
profiles <- get_data_clustering(
    data, profile_type, var1_aggregate_by, var2_aggregate_by)
dist_method <- "mutual_information"
distance_matrix <- get_distance_matrix(profiles, dist_method)
cluster_method <- "complete"
cluster_data_dend(distance_matrix, cluster_method)
}
\seealso{
\code{\link{get_data_clustering}}, 
\code{\link{get_distance_matrix}}, \code{\link{hclust}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
