% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_S4_definitions.R
\docType{class}
\name{CondaEnv-class}
\alias{CondaEnv-class}
\title{Encapsulates information about a conda environment}
\description{
The \code{CondaEnv} class provides a basic structure to store information
about a conda environment for a given method. It is designed to be used by
the \link[=CondaEnvManager-class]{CondaEnvManager} class.
}
\section{Slots}{

\describe{
\item{\code{method}}{1-length character. Indicating the method this \code{CondaEnv}
refers to}

\item{\code{needs.conda}}{logical. Whether the \code{method} needs conda (yes if
python-based, no if R-based)}

\item{\code{conda.bin}}{single string. path to conda binary. If empty (\code{""}),
\code{NULL} or \code{"conda"}, \code{"auto"} is passed to
\link[reticulate]{conda_binary} to find path to conda binary in \code{PATH}}

\item{\code{conda.env.name}}{1-length character. Name of the conda environment}

\item{\code{conda.env.path}}{1-length character. Path to the conda environment}

\item{\code{is.valid}}{logical. Whether the environment is valid. Should not be set
by the user. Set by \link{checkCondaEnv}.}
}}

\seealso{
\link[=CondaEnv]{CondaEnv()}, \linkS4class{CondaEnvManager},
\link[reticulate:conda-tools]{conda_binary()}
}
\keyword{internal}
