% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_S4_helpers.R
\name{CondaEnv}
\alias{CondaEnv}
\title{Handy \code{CondaEnv} instance constructor}
\usage{
CondaEnv(
  method = known.methods,
  conda.bin = NULL,
  conda.env.name = NULL,
  conda.env.path = NULL
)
}
\arguments{
\item{method}{the method name. One of "combat", "harmony", "mnn", "bbknn",
"scvi", "scanvi", "scanorama"}

\item{conda.bin}{the path to the conda binary. If empty (\code{""}),
\code{NULL} or \code{"conda"}, \code{"auto"} is passed to
\link[reticulate]{conda_binary}() to find the path to the conda binary in
\code{PATH}}

\item{conda.env.name}{name of the conda environment}

\item{conda.env.path}{path to the conda environment}
}
\value{
a \linkS4class{CondaEnv} object
}
\description{
Wrapper to create a \linkS4class{CondaEnv} object meant to store
information about a single conda environment for a specific method. In most
cases, it is best that the user favors \code{CondaEnvManager}-related
functions.
}
\seealso{
\linkS4class{CondaEnv}
}
