% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_S4_definitions.R
\docType{class}
\name{CondaEnvManager-class}
\alias{CondaEnvManager-class}
\title{Manager of conda environments for python-based integration methods}
\description{
The \code{CondaEnvManager} class provides a handy way to set up, store and use
conda environments for python-based integration methods. It is designed to be
set up and modified via helper functions, not directly by the user.
}
\section{Slots}{

\describe{
\item{\code{combat}}{CondaEnv. For R-based combat method, nothing to set up.}

\item{\code{harmony}}{CondaEnv. For R-based harmony method, nothing to set up.}

\item{\code{mnn}}{CondaEnv. For R-based MNN method, nothing to set up.}

\item{\code{bbknn}}{CondaEnv. For python-based bbknn method.}

\item{\code{scvi}}{CondaEnv. For python-based SCVI method. Can be shared with SCANVI.}

\item{\code{scanvi}}{CondaEnv. For python-based SCANVI method. Can be shared with SCVI.}

\item{\code{scanorama}}{CondaEnv. For python-based Scanorama method}

\item{\code{trvae}}{CondaEnv. For python-based trVAE method}
}}

\seealso{
\linkS4class{CondaEnv}
}
\keyword{internal}
