% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_S4_helpers.R
\name{CondaManager}
\alias{CondaManager}
\title{Handy \code{CondaEnvManger} instance constructor}
\usage{
CondaManager(cache)
}
\arguments{
\item{cache}{a correctly formatted (see below) data.frame or a path to such a
data.frame stored inside a RDS file. If cache is missing, return an blank
object (i.e. in an uninitialized state).}
}
\value{
a \linkS4class{CondaEnvManager} object
}
\description{
Wrapper to create a \linkS4class{CondaEnvManager} object meant to store
information about conda environments for all known methods out of a data.frame
object (typically stored in the package's cache). In most cases, it is best
that the user favors the \link[=UpdateEnvCache]{UpdateEnvCache()} function
which enables to create/overwrite conda environments and to update the
package's internal cache.
}
\details{
The \code{cache} data.frame's expected column names are:
\itemize{
  \item method
  \item needs.conda
  \item conda.bin.value
  \item conda.bin.valid
  \item conda.env.name.value
  \item conda.env.name.valid
  \item conda.env.path.value
  \item conda.env.path.valid
  \item is.valid
}
}
\seealso{
\linkS4class{CondaEnvManager}
}
