% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{connection}
\alias{connection}
\title{Create the connection object used for all the database functions}
\usage{
connection(connectionString)
}
\arguments{
\item{connectionString}{- the string representing the connection string...}
}
\value{
The connection object
}
\description{
Takes a connection string and returns a connection object
}

