% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cortest.R
\name{cortest}
\alias{cortest}
\title{Calibrated tests for correlation between paired samples}
\usage{
cortest(x, y, ...)
}
\arguments{
\item{x, y}{the two continuous variables. Must be of same length.}

\item{alternative}{indicates the alternative hypothesis and must be one of "two.sided", "greater" or "less".}

\item{method}{a character string indicating which test to implement. Can be \code{"pearson"} or \code{"kendall"}.}
}
\value{
Returns the result of the test with its corresponding p-value, the value of the test statistic and the estimated value of the Pearson correlation coefficient
or Kendall's tau.
}
\description{
Tests the association/correlation for continuous paired samples using corrected versions of the Pearson correlation test and Kendall tau test. These two tests are asymptotically calibrated.
}
\details{
Two tests are implemented. The null hypothesis for the corrected Pearson test is: H0 Cor(X,Y)=0 where Cor represents the Pearson correlation coefficient.
The alternative is specified by the \code{alternative} argument. The null hypothesis for the corrected Kendall test is: H0 tau=0 where rho represents the Kendall's tau coefficient.
Both tests are asymptotically calibrated in the sense that the rejection probability under the null hypothesis is asymptotically equal to 5\%.
}
\examples{
n=100
M=10000 #number of replications
testone=function(n){
X=rnorm(n,0,1)
epsi=rnorm(n,0,1)
Y=X^2+0.3*epsi
list(test1=cortest(X,Y)$p.value,test2=cor.test(X,Y)$p.value) #cor.test is the standard Pearson test
}
res1=res2=rep(NA,M)
# Replications to check if the the corrected Pearson test and the standard test are well calibrated
for (i in 1:M)
{
result=testone(n)
res1[i]=result$test1
res2[i]=result$test2
}
mean(res1<0.05)  #0.0549
mean(res2<0.05)  #0.37

#Replications with Kendall test (takes long time to run)
M=1000
testone=function(n){
X=rnorm(n,0,1)
epsi=rnorm(n,0,1)
Y=X^2+0.3*epsi
list(test1=cortest(X,Y)$p.value,test2=cor.test(X,Y)$p.value,test3=cortest(X,Y,method="kendall")$p.value,test4=cor.test(X,Y,method="kendall")$p.value) #cor.test is the standard Pearson or Kendall correlation test
}
res1=res2=res3=res4=rep(NA,M)
# Replications to check if the the tests are well calibrated
for (i in 1:M)
{
result=testone(n)
res1[i]=result$test1
res2[i]=result$test2
#res3[i]=result$test3
#res4[i]=result$test4
}
mean(res1<0.05)  #0.039
mean(res2<0.05)  #0.346
#mean(res3<0.05) #0.044
#mean(res4<0.05) #0.154

}
\keyword{test}
