% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Counting.R
\name{count.variants}
\alias{count.variants}
\title{Counting variant types that occur in the individual genotypes}
\usage{
count.variants(vcf)
}
\arguments{
\item{x}{A number.}
}
\value{
Returns a list object, call function to variable
}
\description{
This function is used to calculate how many times each variant type
occurs in the genotype among all individuals in the input dataset.
}
\details{
The information is saved to a list, which has names for different variant types
And then string of the number for counts how many times each variation occurs
in the data
}
\section{Warning}{

The function is very slow, big datasets take a lot of time

The function requires a tidyvcfR file as an input
}

\examples{
count.variants(data_subset)

}
