% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_domain_plot.R
\name{create_archi_plot}
\alias{create_archi_plot}
\title{Create protein's domain architecure plot}
\usage{
create_archi_plot(info, domain_df, label_archi_size, title_archi_size)
}
\arguments{
\item{info}{a list contains seed and orthologs IDs}

\item{domain_df}{dataframe contains domain info}

\item{label_archi_size}{lable size (in px)}

\item{title_archi_size}{title size (in px)}
}
\value{
A plot as arrangeGrob object. Use grid::grid.draw(plot) to render.
}
\description{
Create architecture plot for both seed and orthologous protein.
If domains of ortholog are missing, only architecture of seed protein will
be plotted. NOTE: seed protein ID is the one being shown in the profile plot,
which normally is also the orthologous group ID.
}
\examples{
seedID <- "OG_1009"
orthoID <- "A.thaliana@3702@241736"
info <- c(seedID, orthoID)
domain_file <- system.file(
    "extdata", "domain_files/OG_1009.domains",
    package = "phyloprofile", mustWork = TRUE
)
domain_df <- parse_domain_input(seedID, domain_file, "file")
plot <- create_archi_plot(info, domain_df, 9, 9)
grid::grid.draw(plot)
}
\seealso{
\code{\link{domain_plotting}}, \code{\link{sort_domains}},
\code{\link{parse_domain_input}}, \code{\link{get_qual_col_for_vector}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
