% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_gene_age.R
\name{create_gene_age_plot}
\alias{create_gene_age_plot}
\title{Create gene age plot}
\usage{
create_gene_age_plot(count_df, gene_age_text)
}
\arguments{
\item{count_df}{data for plotting gene age}

\item{gene_age_text}{text size}
}
\value{
A gene age distribution plot as a ggplot2 object
}
\description{
Create gene age plot
}
\examples{
count_df <- data.frame(
    age = "OG_1017",
    frea = 2,
    percentage = 100,
    pos = 50
)
gene_age_text <- 1
create_gene_age_plot(count_df, gene_age_text)
}
\seealso{
\code{\link{estimate_gene_age}} and \code{\link{gene_age_plotDf}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
