% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_distribution.R
\name{create_percentage_distribution_data}
\alias{create_percentage_distribution_data}
\title{Create data for percentage present taxa distribution}
\usage{
create_percentage_distribution_data(input_data, rank_name)
}
\arguments{
\item{input_data}{dataframe contains raw input data in long format}

\item{rank_name}{name of the working taxonomy rank (e.g. "species", "family")}
}
\value{
A dataframe ready for analysing the distribution of the percentage of
species in the selected supertaxa
}
\description{
Create data for percentage present taxa distribution
}
\examples{
data("main_long_raw", package="phyloprofile")
create_percentage_distribution_data(main_long_raw, "class")
}
\seealso{
\code{\link{main_long_raw}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
