% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_phylo_profile.R
\name{create_profile_data}
\alias{create_profile_data}
\title{Create data for plotting the phylogentic profiles}
\usage{
create_profile_data(superTaxon_data, ref_taxon, percent_cutoff,
    coortholog_cutoff_max, var1_cutoff, var2_cutoff, var1_relation,
    var2_relation, group_by_cat, cat_dt)
}
\arguments{
\item{superTaxon_data}{a reduced dataframe contains info for all profiles in
the selected taxonomy rank.}

\item{ref_taxon}{selected reference taxon}

\item{percent_cutoff}{min and max cutoffs for percentage of species present
in a supertaxon}

\item{coortholog_cutoff_max}{maximum number of co-orthologs allowed}

\item{var1_cutoff}{min and max cutoffs for var1}

\item{var2_cutoff}{min anc max cutoffs for var2}

\item{var1_relation}{relation of var1 ("protein" for protein-protein or
"species" for protein-species)}

\item{var2_relation}{relation of var2 ("protein" for protein-protein or
"species" for protein-species)}

\item{group_by_cat}{group genes by their categories (TRUE or FALSE)}

\item{cat_dt}{dataframe contains gene categories
(optional, NULL if group_by_cat = FALSE or no info provided)}
}
\value{
A dataframe ready for generating profile plot.
}
\description{
Create data for plotting the phylogentic profiles
}
\examples{
data("full_processed_profile", package="phyloprofile")
superTaxon_data <- reduce_profile(full_processed_profile)
ref_taxon <- "Mammalia"
percent_cutoff <- c(0.0, 1.0)
coortholog_cutoff_max <- 10
var1_cutoff <- c(0.75, 1.0)
var2_cutoff <- c(0.5, 1.0)
var1_relation <- "protein"
var2_relation <- "species"
group_by_cat <- FALSE
cat_dt <- NULL
create_profile_data(
    superTaxon_data,
    ref_taxon,
    percent_cutoff,
    coortholog_cutoff_max,
    var1_cutoff,
    var2_cutoff,
    var1_relation,
    var2_relation,
    group_by_cat,
    cat_dt
)
}
\seealso{
\code{\link{parse_info_profile}} and \code{\link{reduce_profile}}
for generating input dataframe, \code{\link{full_processed_profile}} for a
demo full processed profile dataframe
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
