% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_phylotree.R
\name{create_rooted_tree}
\alias{create_rooted_tree}
\title{Create rooted tree from a taxonomy matrix}
\usage{
create_rooted_tree(df, root_taxon)
}
\arguments{
\item{df}{data frame contains taxonomy matrix used for creating tree}

\item{root_taxon}{taxon used for rooting tree}
}
\value{
rooted tree based on root_taxon
}
\description{
Create rooted tree from a taxonomy matrix
}
\examples{
data("pp_taxonomy_matrix", package = "phyloprofile")
# prepare matrix for calculating distances
distDf <- subset(pp_taxonomy_matrix, select = -c(ncbiID, fullName))
row.names(distDf) <- distDf$abbrName
distDf <- distDf[, -1]
# create taxonomy tree rooted by ncbi10090
create_rooted_tree(distDf, "ncbi10090")
}
\seealso{
\code{\link{taxa2dist}} for distance matrix generation from a
taxonomy matrix, \code{\link{get_taxonomy_matrix}} for getting taxonomy
matrix, \code{\link{pp_taxonomy_matrix}} for a demo taxonomy matrix data
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
