% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_distribution.R
\name{create_var_dist_plot}
\alias{create_var_dist_plot}
\title{Create distribution plot}
\usage{
create_var_dist_plot(data, var_name, var_type, percent, dist_text_size)
}
\arguments{
\item{data}{dataframe contains data for plotting}

\item{var_name}{name of the variable that need to be analyzed (either name of
variable 1 or variable 2 or "percentage of present taxa")}

\item{var_type}{type of variable (either "var1", "var2" or "presSpec")}

\item{percent}{range of percentage cutoff}

\item{dist_text_size}{text size of the distribution plot}
}
\value{
A distribution plot as a ggplot object
}
\description{
Create distribution plot for one of the additional variable or
the percentage of the species present in the supertaxa.
}
\examples{
data("main_long_raw", package="phyloprofile")
data <- create_variable_distribution_data(
    main_long_raw, 0, 1, 0.5, 1
)
var_name <- "Variable abc"
var_type <- "var1"
percent <- c(0,1)
dist_text_size <- 12
create_var_dist_plot(
    data,
    var_name,
    var_type,
    percent,
    dist_text_size
)
}
\seealso{
\code{\link{main_long_raw}},
\code{\link{create_variable_distribution_data}},
\code{\link{create_variable_distribution_data_subset}},
\code{\link{create_percentage_distribution_data}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
