% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_distribution.R
\name{create_variable_distribution_data}
\alias{create_variable_distribution_data}
\title{Create data for additional variable distribution}
\usage{
create_variable_distribution_data(input_data, var1_cutoff_min,
    var1_cutoff_max, var2_cutoff_min, var2_cutoff_max)
}
\arguments{
\item{input_data}{dataframe contains raw input data in long format}

\item{var1_cutoff_min}{min cutoff for var1}

\item{var1_cutoff_max}{max cutoff for var1}

\item{var2_cutoff_min}{min cutoff for var2}

\item{var2_cutoff_max}{max cutoff for var2}
}
\value{
A dataframe ready for analysing the distribution of the additional
variable(s)
}
\description{
Create data for additional variable distribution
}
\examples{
data("main_long_raw", package="phyloprofile")
create_variable_distribution_data(
    main_long_raw, 0, 1, 0.5, 1
)
}
\seealso{
\code{\link{main_long_raw}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
