% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_distribution.R
\name{create_variable_distribution_data_subset}
\alias{create_variable_distribution_data_subset}
\title{Create data for additional variable distribution (for a subset data)}
\usage{
create_variable_distribution_data_subset(full_profile_data,
    distribution_data, selected_genes, selected_taxa)
}
\arguments{
\item{full_profile_data}{dataframe contains the full processed profiles}

\item{distribution_data}{dataframe contains the full distribution data}

\item{selected_genes}{list of genes of interst}

\item{selected_taxa}{list of taxa of interest}
}
\value{
A dataframe ready for analysing the distribution of the additional
variable(s) for a subset of genes and/or taxa.
}
\description{
Create data for additional variable distribution (for a subset data)
}
\examples{
data("full_processed_profile", package="phyloprofile")
data("main_long_raw", package="phyloprofile")
distribution_data <- create_variable_distribution_data(
    main_long_raw, 0, 1, 0.5, 1
)
selected_genes <- "OG_1019"
selected_taxa <- c("Mammalia", "Echinoidea", "Gunneridae", "Mucorales",
"Alphaproteobacteria")
create_variable_distribution_data_subset(
    full_processed_profile,
    distribution_data,
    selected_genes,
    selected_taxa
)
}
\seealso{
\code{\link{parse_info_profile}},
\code{\link{create_variable_distribution_data}},
\code{\link{full_processed_profile}}, \code{\link{main_long_raw}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
