% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_package.R
\name{createPackage}
\alias{createPackage}
\title{function to a repository and R package}
\usage{
createPackage(package.name, where, organisation.name = NULL)
}
\arguments{
\item{package.name}{the name of the package that you need to create}

\item{where}{the path to the folder where you want to create the package}

\item{organisation.name}{the name of the organisation in which the package should be created. For Epiverse TRACE, this is "epiverse-trace"}
}
\description{
function to a repository and R package
}
\examples{
createPackage(package.name="mypackage", where="/Users/km28/Documents/Karim/Karim/LSHTM/Codes", organisation.name="Karim-Mane")
}
