% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{csv}
\alias{csv}
\title{This raw data has been included for test 
purpose
Two fields are present in data
First csv file read into R object
Then save(mydata, file="data/mydata.RData") can be used to 
save data to ./data folder
Alternatevely devtools::use_data(objname) cab be used
Save .R file using @TAG to create manual in ./man folder}
\format{A data frame with columns:
\describe{
	\item{Standing}{ranking in numeric format}
	\item{Team}{Country Name}
}}
\usage{
data(csv)
}
\description{
This raw data has been included for test 
purpose
Two fields are present in data
First csv file read into R object
Then save(mydata, file="data/mydata.RData") can be used to 
save data to ./data folder
Alternatevely devtools::use_data(objname) cab be used
Save .R file using @TAG to create manual in ./man folder
}
\examples{
data(csv)
}
\keyword{datasets}
