% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CutKnn}
\alias{CutKnn}
\title{Remove excessive number of neighbours in a knn graph}
\usage{
CutKnn(
  object,
  graph.name,
  new.graph = NULL,
  k.max,
  assay = NULL,
  verbose = TRUE
)
}
\arguments{
\item{object}{A \code{\link[SeuratObject]{Seurat}} object}

\item{graph.name}{name of a \code{Graph} or \code{Neighbor} instance stored in
the \code{Seurat object}.}

\item{new.graph}{name of the trimmed graph to save in the \code{Seurat object}}

\item{k.max}{maximum number of neigbours allowed per cell}

\item{assay}{name of the assay to use. Ignored when graph object is a
\code{Neighbor} object. If not specified (default), the default assay is used}

\item{verbose}{whether to print messages. Set to \code{FALSE} to disable}
}
\value{
the Seurat object with a new \code{Graph} or \code{Neighbor} instance
}
\description{
Ensure that each cell's number of neighbours does not exceed a cutoff
}
