% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.R
\name{CV}
\alias{CV}
\title{Cross Validation for finding RKHS smoothing mean parameters}
\usage{
CV(grid = "FALSE", Data, alpha = 1, beta = 0.1, kernel = "Exp",
  phi = 0.01, ro = 0.2, fold = 10, cv.penalty = "Regular", Rep = 1000,
  print.cv = TRUE, print.estimated.time = TRUE, col.drop = TRUE)
}
\arguments{
\item{grid}{grid (x-axis) for each curve, default is equally espaced between 0 and 1.}

\item{Data}{a matrix which the of interest curves are located in columns}

\item{alpha, beta}{Privacy parameters, real numbers}

\item{kernel}{kernel function, can be "Exp" (Exponential kernel),
"M3/2" (Matern precess with \eqn{\nu}=3/2) "M5/2" (Matern precess with \eqn{\nu}=5/2)
"Gau" (Gaussian kernel) and "Sob" (Sobolev kernel)
else define it as a bivariate kernel function with parameters "t" and "s" and a range parameter "ro".}

\item{phi}{a real vector of penalty parameters, It will be done a grid search on them
to find the minimum Cross Validation}

\item{ro}{a real vector of kernel range parameters, It will be done a grid search on them
to find the minimum Cross Validation}

\item{fold}{number of fold using in Cross Validation}

\item{cv.penalty}{"Regular" or "Irregular"}

\item{Rep}{number of replications, (just) for "Irregular" method}
}
\value{
par: the optimum penalty (\eqn{\phi}) and kernel range parameter (\eqn{\rho})
in an RKHS smoothing mean which gives the minimum Cross Validation

time: estimatation of remaining time
}
\description{
This function apply two "Regular" and "Irregular" methods to find the penalty (\eqn{\phi}) and
kernel range parameter (\eqn{\rho}) in an RKHS smoothing mean
}

