% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradingview.R
\docType{data}
\name{data_all_stock}
\alias{data_all_stock}
\title{All the stock data}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 5602 rows and 41 columns.}
\usage{
data_all_stock
}
\description{
All the stock data
}
\keyword{datasets}
