% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_profile_plot.R
\name{data_customized_plot}
\alias{data_customized_plot}
\title{Create data for customized profile plot}
\usage{
data_customized_plot(data_heat, selected_taxa, selected_seq)
}
\arguments{
\item{data_heat}{a data frame contains processed profiles}

\item{selected_taxa}{subset of taxa}

\item{selected_seq}{subset of sequences}
}
\value{
A data frame contains data for customized profile plot.
}
\description{
Create data for customized profile plot based on a selected 
list of genes and/or taxa.
}
\examples{
data("full_processed_profile", package="phyloprofile")
selected_taxa <- c("Mammalia", "Echinoidea", "Gunneridae")
selected_seq <- "all"
data_customized_plot(full_processed_profile, selected_taxa, selected_seq)
}
\seealso{
\code{\link{from_input_to_profile}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
