% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_profile_plot.R
\name{data_main_plot}
\alias{data_main_plot}
\title{Create data for main profile plot}
\usage{
data_main_plot(data_heat)
}
\arguments{
\item{data_heat}{a data frame contains processed profiles}
}
\value{
A data frame contains data for main profile plot.
}
\description{
Create data for main profile plot
}
\examples{
data("full_processed_profile", package="phyloprofile")
data_main_plot(full_processed_profile)
}
\seealso{
\code{\link{from_input_to_profile}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
