% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAccess.R
\name{DataAccess.connection_status}
\alias{DataAccess.connection_status}
\title{returns the connection status for current da_obj connection string}
\usage{
DataAccess.connection_status(da_obj)
}
\arguments{
\item{da_obj}{- The Solvas dataAccess object}
}
\value{
connection status string "success" if connection is valid otherwise an error message is returned
}
\description{
Takes a DataAccess da_obj and returns "success" if the connection is 
working otherwise the error message is returned.
}

