% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAccess.R
\name{DataAccess.ConnectionStatus}
\alias{DataAccess.ConnectionStatus}
\title{Returns the connection status}
\usage{
DataAccess.ConnectionStatus(da_obj)
}
\arguments{
\item{da_obj}{- Current instance of Solvas|Capital's DataAccess class.}
}
\value{
connection status string "success" if connection is valid otherwise an error message is returned
}
\description{
Takes a DataAccess da_obj and returns "success" if the connection is 
working otherwise the error message is returned.
}

