% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAccess.R
\name{DataAccess.fi_instrument_as_of_get}
\alias{DataAccess.fi_instrument_as_of_get}
\title{Get fincancial instruments for the context}
\usage{
DataAccess.fi_instrument_as_of_get(object, as_of_date = NULL)
}
\arguments{
\item{object}{The Solvas dataAccess object}

\item{as_of_date}{The as_of_date to use for schedule data types}
}
\value{
Data frame containing the financial instrument data
}
\description{
Takes a DataAccess object and as_of_ and returns a datatable
with the instruments.  Properties that are schedules are coalesced 
to a slingle value based on the as_of_
}

