% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAccess.R
\name{DataAccess.fi_instrument_get}
\alias{DataAccess.fi_instrument_get}
\title{Get fincancial instruments for the context}
\usage{
DataAccess.fi_instrument_get(da_obj, effective_date = NULL,
  effective_period = NULL)
}
\arguments{
\item{da_obj}{- The Solvas dataAccess object}

\item{effective_date}{= Used for to resolve schedules column values to a single scalar value}

\item{effective_period}{= Used for to resolve schedules column values to a single scalar value (1=first period)}
}
\value{
Data frame containing the financial instrument data
}
\description{
Takes a DataAccess da_obj and effective_date or effective_period  and returns a datatable
with the instruments.  Properties that are schedules are coalesced 
to a single value based on the effective_date or effective_period. 
NOTE: EITHER effective_date or effective_period must be poplulate the other one must be NULL
}

