% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAccess.R
\name{DataAccess.fs_assumptions_get}
\alias{DataAccess.fs_assumptions_get}
\title{Get economic assumptions for the context}
\usage{
DataAccess.fs_assumptions_get(da_obj, transformation_sequence_order,
  criteria_sequence_order)
}
\arguments{
\item{da_obj}{- The Solvas dataAccess object}

\item{transformation_sequence_order}{- an integer value found on the 'Transformations' screen}

\item{criteria_sequence_order}{- an integer value found on the 'Transformation Entries for Sequence' screen}
}
\description{
Get economic assumptions for the context
}

