% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAccess.R
\name{DataAccess.FSAssumptionsGet}
\alias{DataAccess.FSAssumptionsGet}
\title{Get economic assumptions}
\usage{
DataAccess.FSAssumptionsGet(da_obj, transformation_sequence_order,
  criteria_sequence_order)
}
\arguments{
\item{da_obj}{- Current instance of Solvas|Capital's DataAccess class.}

\item{transformation_sequence_order}{- An integer value found on the 'Transformations' screen}

\item{criteria_sequence_order}{- An integer value found on the 'Transformation Entries for Sequence' screen}
}
\description{
Get economic assumptions
}

