% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAccess.R
\name{DataAccess.PackageVersionCompatible}
\alias{DataAccess.PackageVersionCompatible}
\title{Returns the connection status}
\usage{
DataAccess.PackageVersionCompatible(da_obj)
}
\arguments{
\item{da_obj}{- Current instance of Solvas|Capital's DataAccess class.}
}
\value{
"success" if the version is compatible with the Solvas|Capital database server
}
\description{
Takes a DataAccess da_obj and returns "success" if the connection is 
working otherwise the error message is returned.
}
