% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAccess_RC.R
\docType{class}
\name{DataAccess_RC}
\alias{DataAccess_RC}
\title{A DataAccess Reference class example}
\description{
A DataAccess Reference class example
}
\section{Fields}{

\describe{
\item{\code{connection_param}}{connection string...}

\item{\code{fs_id_param}}{fs id....}
}}
\section{Methods}{

\describe{
\item{\code{fi_instrument_get(as_of_date)}}{returns a datatable with the instruments for the passed as_of_date}

\item{\code{fi_instrument_initialize()}}{initializes the database }
}}
\examples{
da_rc <- new ("DataAccess_RC", connection_param=connection_string, fs_id_param=1)
da_rc$fi_instrument_get('1/1/2014')
}

