% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAccess_S3.R
\docType{methods}
\name{DataAccess_S3}
\alias{DataAccess_S3}
\title{DataAccess_S3: Object to connect and interact with the Capital database}
\usage{
DataAccess_S3(connection_param = "", fs_id_param = NULL)
}
\arguments{
\item{connection_param}{The SQL Server connection string}

\item{fs_id_param}{The scenario ID}
}
\details{
text describing parameter inputs in more detail.
\itemize{
 \item{"parameter 1"}{Stuff}
 \item{"parameter 2"}{Stuff}
 \item da <- Solvas.Capital.SQLUtility::DataAccess_S3(connectionString, fs_id)
 \item da$fi_instrument_get()
}
}
\examples{
da <- Solvas.Capital.SQLUtility::DataAccess_S3(connectionString, fs_id)
da$fi_instrument_get()
}

