% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAccess_S3_V2.R
\docType{methods}
\name{DataAccess_S3V2}
\alias{DataAccess_S3V2}
\title{DataAccess_S3V2: Object to connect and interact with the Capital database}
\usage{
DataAccess_S3V2(connection_param = "", fs_id_param = NULL)
}
\arguments{
\item{connection_param}{The SQL Server connection string}

\item{fs_id_param}{The scenario ID}
}
\examples{
da_s3v2 <- DataAccess_S3V2(connection_param=connectionString, fs_id_param=1)
DataAccess_S3V2.fi_get_instrument(da_s3v2)
}

