% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAccess_S3_V2.R
\name{DataAccess_S3V2.fi_get_instrument}
\alias{DataAccess_S3V2.fi_get_instrument}
\title{Get fincancial instruments for the context}
\usage{
DataAccess_S3V2.fi_get_instrument(object, as_of_date = NULL)
}
\arguments{
\item{object}{The Solvas dataAccess object}

\item{as_of_date}{The as_of_date to use for schedule data types}
}
\value{
Data frame containing the financial instrument data
}
\description{
Takes a connection string (connectionString) and scenario id (fs_id) and returns a datatable
with the instruments
}

