% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAccess_S4.R
\docType{class}
\name{DataAccess_S4-class}
\alias{DataAccess_S4}
\alias{DataAccess_S4-class}
\title{DataAccessS4: Object to connect and interact with the Capital database
DataAccessS4(connection_param, fs_id)}

