% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_statistics.R
\name{deconvolve_transformed_list}
\alias{deconvolve_transformed_list}
\title{Batch EDec deconvolution on transformed data}
\usage{
deconvolve_transformed_list(transformed_list, cell_proportion)
}
\arguments{
\item{transformed_list}{List of transformed expression matrices}

\item{cell_proportion}{Cell type proportion matrix}
}
\value{
Named list of deconvolution results
}
\description{
Batch EDec deconvolution on transformed data
}
