% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_df.R
\name{describe_df}
\alias{describe_df}
\title{Descriptive statistics}
\usage{
describe_df(df)
}
\arguments{
\item{df}{Dataframe for which the description is applied.}
}
\value{
Dataframe with the descriptive statistics.
}
\description{
Calculates the following descriptive statistics to return them in a DF: mean,
standard deviation, variance, minimum, quartile 1, quartile 2, quartile 3,
maximum, kurtosis, skewness, count of missing data, IQR and count
}
\examples{
describe_df(iris)
}
