% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{divide}
\alias{divide}
\title{divide two vectors}
\usage{
divide(x, y)
}
\arguments{
\item{x,y}{Numeric vectors}
}
\value{
division of x and y
}
\description{
divide two vectors or numbers
}
\examples{
divide(4,2)
divide(c(2,1),c(10,5))
}

