% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_pantheria_data.R
\name{dl_pantheria_data}
\alias{dl_pantheria_data}
\title{Opens penthera dataset from github}
\usage{
dl_pantheria_data()
}
\arguments{
\item{x}{nothing}

\item{y}{nothing}
}
\value{
a txt dataset.
}
\description{
Opens penthera dataset from github
}
