% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_wildfinder_data.R
\name{dl_wildfinder_data}
\alias{dl_wildfinder_data}
\title{Opens wildfinder dataset from github}
\usage{
dl_wildfinder_data()
}
\arguments{
\item{x}{nothing}

\item{y}{nothing}
}
\value{
a txt dataset.
}
\description{
This function downloads the wildfinder datasets (.csv files) hosted on the
GitHub repository \url{https://github.com/rdatatoolbox/datarepo/}
}
