% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_integrate.R
\name{DoIntegrate}
\alias{DoIntegrate}
\title{Integrate layers using one or multiple integration method(s)}
\usage{
DoIntegrate(
  object,
  ...,
  use.hvg = TRUE,
  use.future = TRUE,
  future.globals.size = getOption("future.globals.maxSize")
)
}
\arguments{
\item{object}{A Seurat object}

\item{...}{one or more integration method \strong{call}(s) to correct batch
effects with. Must be of the form \code{package::MethodIntegration()}. It is
recommended to use \code{::} because it is safer in case of namespace
collision or if the package is not attached nor loaded. \strong{Don't forget
the parentheses}.}

\item{use.hvg}{whether to use highly variable genes. \code{FALSE} causes all
the features present in the assay to be used.}

\item{use.future}{whether to use \pkg{future} to run integrations in
a background session. Useful when python-based algorithms are invoked.}

\item{future.globals.size}{maximum allowed size (in bytes) of global
variables to export. By default, uses the value of the option
"future.globals.maxSize". If it is \code{NULL}, thrice the size of the Seurat
object is used. Inoperative when \code{use.future = FALSE}}
}
\value{
the updated Seurat object enriched with the integration methods'
outputs.
}
\description{
Integrate layers of a Seurat object using one or more integration methods.

Available integration methods are listed at the bottom of this page.
\code{DoIntegrate()} works best with \pkg{SeuratIntegrate}'s methods.
}
\details{
Each call to an integration method \strong{require parentheses}. Parameter
values specific to each method can be enclosed between these brackets,
although the defaults arguments should be good enough for most cases. Note
that for each integration method, \strong{the argument values specified in
its call supersede \code{DoIntegrate}'s internal computations}. For instance,

\preformatted{
DoIntegrate(seu,
            SeuratIntegrate::CombatIntegration(features = Features(seu)),
            Seurat::CCAIntegration(),
            use.hvg = TRUE)
}

forces ComBat but not CCA to use all features instead of the variable ones.
}
\note{
The desired value of parameters \code{use.hvg}, \code{use.future} and
\code{future.globals.size} can be different for each method called. Hence,
they accept vectors with more than one element. They must be in the same
order as the integration method calls.

This unconventional way of calling methods presents a few advantages:
With a single call to \code{DoIntegrate}, you can perform multiple
integrations at once, while preserving the flexibility of an individual
method call by fine-tuning integration-specific parameters.
}
\section{Integration Method Functions}{


The following integration method functions are available:
\Sexpr[stage=render,results=rd]{Seurat:::.rd_methods("integration")}

}

