% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_domain_plot.R
\name{domain_plotting}
\alias{domain_plotting}
\title{Create architecure plot for a single protein}
\usage{
domain_plotting(df, geneID, sep, label_size, title_size, min_start, 
    max_end, color_scheme)
}
\arguments{
\item{df}{domain dataframe for ploting}

\item{geneID}{ID of seed or orthologous protein}

\item{sep}{separate indicator for title}

\item{label_size}{lable size}

\item{title_size}{title size}

\item{min_start}{the smallest start position of all domains}

\item{max_end}{the highest stop position of all domains}

\item{color_scheme}{color scheme for all domain types}
}
\value{
A ggplot object
}
\description{
Create architecure plot for a single protein
}
\examples{
\dontrun{
# get domain data
domain_file <- system.file(
    "extdata", "domain_files/OG_1009.domains",
    package = "phyloprofile", mustWork = TRUE
)
domain_df <- parse_domain_input(seedID, domain_file, "file")
df <- domain_df[domain_df$orthoID == "A.thaliana@3702@241736",]
# create color scheme for all domain types
all_features <- levels(as.factor(df$feature))
all_colors <- get_qual_col_for_vector(all_features)
color_scheme <- structure(
    all_colors,
    .Names = all_features
)
# other parameters
geneID <- "A.thaliana@3702@241736"
sep <- "|"
label_size <- 9
title_size <- 9
min_start <- min(df$start)
max_end <- max(df$end)
# do plotting
domain_plotting(
    df,
    geneID,
    sep,
    label_size, title_size,
    min_start, max_end,
    color_scheme
)
}
}
\seealso{
\code{\link{get_qual_col_for_vector}},
\code{\link{parse_domain_input}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
