% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double.test.R
\name{double}
\alias{double}
\title{This function doubles x}
\usage{
double(x)
}
\arguments{
\item{x, }{the number to be doubled}
}
\value{
a doubled number
}
\description{
This function doubles x
}
\examples{
double(2)
}
\author{
Jenessa Lemon
}

