% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP.RKHS.R
\name{DP.RKHS}
\alias{DP.RKHS}
\title{Releasing Differential Private RKHS smoothing mean of a dataset}
\usage{
DP.RKHS(grid = "Default", Data, alpha = 1, beta = 0.1, kernel = "Gau",
  phi = 0.01, ro = 0.2, col.drop = TRUE)
}
\arguments{
\item{grid}{grid (x-axis) for each curve, default is equally espaced between 0 and 1.}

\item{Data}{a matrix which the of interest curves are located in columns}

\item{alpha, beta}{Privacy parameters, real numbers}

\item{phi}{real number, penalty parameter}

\item{ro}{kernel range parameter in kernel, real number}

\item{drop.col}{TRUE/FALSE for cleaning the Data, deleting Columns/Rows including missing values}
}
\value{
f.tilda: DP RKHS smoothing mean

delta: the coefficient of the noise, real number

f: RKHS smoothing mean

X: original data (X=Data)
}
\description{
This function create a DP RKHS smoothing mean from an existing data set
with known eigenvalues and eigenvectors
}

