% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DP.Sim.R
\name{DP.Sim}
\alias{DP.Sim}
\title{Releasing Differential Private RKHS smoothing mean of a simulated dataset}
\usage{
DP.Sim(alpha = 1, beta = 0.1, kernel = "Gau", phi = 0.01, ro = 0.2,
  n = 100, N = 25, tau = 0.4, case = 2, pow = 4, mu = rep(0, n))
}
\arguments{
\item{alpha, beta}{Privacy parameters, real numbers}

\item{phi}{real number, penalty parameter}

\item{ro}{range parameter in kernel, real number}

\item{n}{real number, number of grid points}

\item{N}{real number, number of observations}

\item{tau}{range of the uniform distribution in KL expansion}

\item{pow}{smoothing parameter, e.val.x_{i}=i^{-pow}}

\item{mu}{real vector n*1, initial mean vector}

\item{e.val.x}{real vector n*1, eigenvalues}

\item{e.vec.x}{real valued matrix n*N, eigenvectors}

\item{e.val.z}{real valued matrix n*N, eigenvectors of noise}
}
\value{
f.tilda: DP RKHS smoothing mean,  n*1 real valued vector

delta: the coefficient of the noise, real number

f: RKHS smoothing mean, n*1 real valued vector

X: original data generated by the selected noise covariance operator C by KL epansion
when e.val.x_{i}=i^{-\code{pow}} and e.vec.x=e.vec.z ...
}
\description{
This function create a DP RKHS smoothing mean from an existing data set
with known eigenvalues and eigenvectors
}

