% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDP.Sim.R
\name{DP4.Sim}
\alias{DP4.Sim}
\title{Releasing 4 Differential Private RKHS smoothing means of a simulated dataset}
\usage{
DP4.Sim(alpha = rep(1, 4), beta = rep(0.1, 4), kernel = c("Exp", "M3/2",
  "M5/2", "Gau"), n = rep(100, 4), N = rep(25, 4), phi = rep(0.01, 4),
  ro = rep(0.2, 4), tau = rep(0.4, 4), case = rep(2, 4), pow = rep(4,
  4), mu = matrix(rep(0, 4 * 100), ncol = 4))
}
\arguments{
\item{alpha, beta}{Privacy parameters, real numbers}

\item{n}{real vector 4*1, number of grid points}

\item{N}{real vector 4*1, number of observations}

\item{phi}{real number, penalty parameter}

\item{ro}{range parameter in kernel, real number}

\item{tau}{range of the uniform distribution in KL expansion}

\item{pow}{smoothing parameter, e.val.x_{i}=i^{-pow}}

\item{mu}{real vector n*1, initial mean vector}

\item{e.val.x}{real vector n*1, eigenvalues}

\item{e.vec.x}{real valued matrix n*N, eigenvectors}

\item{e.val.z}{real valued matrix n*N, eigenvectors of noise}
}
\value{
four \link{DP_Sim} output each of them includin "f.tilda","delta","f" and "X"
}
\description{
This function create 4 DP RKHS smoothing means from an existing data set
with known eigenvalues and eigenvectors
}

