% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_mixes.R
\name{energy_mixes}
\alias{energy_mixes}
\title{Function to randomly generate a mix of energy from different sources to meet the global energy demand}
\usage{
energy_mixes(energy, demand, probs)
}
\arguments{
\item{energy}{a vector of the names of different energy sources (i.e. s = solar, w = wind etc)}

\item{demand}{the amount of energy (in TWh) that you want to draw (i.e. the global demand)}

\item{probs}{the probability of drawing each type of energy based on global energy potential for each source}
}
\value{
out a data frame consisting of the following items
\describe{
\item{energy_type}{The different energy sources}
\item{TWh}{The number of TWh provided by each energy source}
\item{percent_of_demand}{The percent of total demand that comes from each energy source}
}
}
\description{
Function to randomly generate a mix of energy from different sources to meet the global energy demand
}
