% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_gene_age.R
\name{estimate_gene_age}
\alias{estimate_gene_age}
\title{Calculate the phylogenetic gene age from profiles}
\usage{
estimate_gene_age(processed_profile_data, rank_name, ref_taxon,
    var1_cutoff, var2_cutoff, percent_cutoff)
}
\arguments{
\item{processed_profile_data}{dataframe contains the full processed
phylogenetic profiles}

\item{rank_name}{taxonomy rank (e.g. "species", "genus", "family")}

\item{ref_taxon}{reference taxon (e.g. "Homo sapiens", "Homo", "Hominidae")}

\item{var1_cutoff}{cutoff for var1}

\item{var2_cutoff}{cutoff for var2}

\item{percent_cutoff}{cutoff for percentage of species present in each
supertaxon}
}
\value{
A dataframe contains estimated gene ages
}
\description{
Calculate the phylogenetic gene age from profiles
}
\examples{
data("full_processed_profile", package="phyloprofile")
rank_name <- "class"
ref_taxon <- "Mammalia"
processed_profile_data <- full_processed_profile
var1_cutoff <- c(0, 1)
var2_cutoff <- c(0, 1)
percent_cutoff <- c(0, 1)
estimate_gene_age(
    processed_profile_data,
    rank_name,
    ref_taxon,
    var1_cutoff, var2_cutoff, percent_cutoff
)
}
\seealso{
\code{\link{parse_info_profile}} for creating a full processed
profile dataframe; \code{\link{get_name_list}} and
\code{\link{get_taxonomy_matrix}} for getting taxonomy info,
\code{\link{full_processed_profile}} for a demo input dataframe
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
