% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.r
\name{euclidean}
\alias{euclidean}
\title{Euclidean}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A numeric scalar}

\item{b}{A numeric scalar}
}
\value{
Return the greatest common divisor of the parameters a and b
}
\description{
Euclidean
}
\examples{
euclidean(123612, 13892347912)
euclidean(13892347912,123612)
euclidean(100, 1000)
}
