% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mF1.R
\name{evaluate}
\alias{evaluate}
\title{evaluate.}
\usage{
evaluate(true_lab, pred_lab, Indices = NULL)
}
\arguments{
\item{pred_lab}{predicated label}

\item{Indices}{a index set}

\item{true_labe}{true label.}
}
\value{
a list.
}
\description{
calculate the mF1 between true label and predicated label
}
\details{
calculate the mF1 between true label and predicated label
}
