\name{extent}

\Rdversion{1.1}

\alias{extent}
\alias{extent,VectorLayer-method}

\title{
Exent
}

\description{
Get the extent (bounding box) of an object
}

\usage{
extent(x, ...)
}

\arguments{
  \item{x}{A GeoVector object}
  \item{\dots}{Additional arguments (none implemented)}
}

\value{
Extent object
}

\author{
Robert J. Hijmans
}


\keyword{ spatial  }
