% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_map_state.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Draw map
using datafile of a specific year, yearly accident in a state is drawn in a map.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{State number}

\item{year}{The year of the accidents recorded}
}
\value{
This function returns yearly accident map
}
\description{
Draw map
using datafile of a specific year, yearly accident in a state is drawn in a map.
}
