% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{fars_read}
\alias{fars_read}
\title{Read data file.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name of the file to be called.}
}
\value{
This function returns a part of file as a printed table.
}
\description{
This reads the bz2 file to be used in other functions of this package.
Users do not have to see the result of this function, thus it is not exported.
}
\examples{
fars_read("accident_2014.csv.bz2")
}
