% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read_years.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Select data
This function harvests data from the file for summary or drawing maps.
Users don't need to do this separately, thus this function is not exported.}
\usage{
fars_read_years(years)
}
\arguments{
\item{year}{The year of the accidents recorded}
}
\value{
This function returns month data of the accidents in the accident file
}
\description{
Select data
This function harvests data from the file for summary or drawing maps.
Users don't need to do this separately, thus this function is not exported.
}
