% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_summarize_years.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summary
using datafile of a specific year, the number of accidents per month are presented.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{year}{The year of the accidents recorded}
}
\value{
This function returns the number of accidents per month in the accident file
}
\description{
Summary
using datafile of a specific year, the number of accidents per month are presented.
}
