% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_fbdocs.R
\name{fbdocs_server}
\alias{fbdocs_server}
\title{Server for HiDAP documents}
\usage{
fbdocs_server(input, output, session, values)
}
\arguments{
\item{input}{shinyserver input}

\item{output}{shinyserver output}

\item{session}{shinyserver session}

\item{values}{reactive values}
}
\description{
Server for HiDAP documents
}
\author{
Omar Benites
}

