% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_fbdocs.R
\name{fbdocs_ui}
\alias{fbdocs_ui}
\title{UI for HIDAP documentation
Returns user friendly ui}
\usage{
fbdocs_ui(type = "tab", title = "HIDAP Documents",
  name = "docshidap_fieldbooks")
}
\arguments{
\item{type}{type of UI element, deault is a tab in a shinydashboard}

\item{title}{diaply title name}

\item{name}{UI TabName}
}
\description{
UI for HIDAP documentation
Returns user friendly ui
}
\author{
Omar Benites
}

