% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{fi_instrument_get}
\alias{fi_instrument_get}
\title{Get fincancial instruments for a specific screnario ID and}
\usage{
fi_instrument_get(connectionString, fs_id)
}
\arguments{
\item{connectionString}{- the string representing the connection string...}

\item{fs_id}{The scenario ID}
}
\value{
Data frame containing the financial instrument data
}
\description{
Takes a connection string (connectionString) and scenario id (fs_id) and returns a datatable
with the instruments
}

