% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{fi_instrument_init_nc}
\alias{fi_instrument_init_nc}
\title{Initalizes the financial instrument records - must be done on scenarios
that have not been run}
\usage{
fi_instrument_init_nc(connectionString, fs_id)
}
\arguments{
\item{connectionString}{The string representing the connection string...}

\item{fs_id}{The scenario ID}
}
\value{
Data frame containing the financial instrument data
}
\description{
Takes a connection string (connectionString) and scenario id (fs_id) and 
initializes the financial instruments on the server.  This *should* not 
be executed on a scenario that has already been run as it will wipe out the
snapshot.
}

