% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterCirBed.R
\name{filter_cirBed}
\alias{filter_cirBed}
\title{Filter the bed file generated by CircleMap}
\usage{
filter_cirBed(cirBed, scores = 50, discord = 4, splitreads = 4, length = 3e+07)
}
\arguments{
\item{cirBed}{get by cirBed = read.table("my_unknown_circleMap.bed", sep = '\t', stringsAsFactors = FALSE)}

\item{scores}{Threshold of Circle score. score below would be removed}

\item{discord}{Threshold of number of discordant read pairs supporting the detected circular DNA. Below the threshold would be reomved}

\item{splitreads}{Threshold of number of split reads supporting the detected circular DNA. Below the threshold would be removed}

\item{length}{Maximum length of the circular DNA detected.}
}
\value{
a data frame with :
c("Chromosome","start_coord",
  "end_coord",
  "dis_cord",
  "split_reads",
  "circle_score",
  "Mean_coverage",
  "SD",
  "CISC",
  "CIEC",
  "coverage_continuity")
}
\description{
This function performs filtration of the raw bed file generated by CircleMap
}
