% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{FindOptimalClusters}
\alias{FindOptimalClusters}
\title{Find a clustering that maximises NMI or ARI}
\usage{
FindOptimalClusters(
  object,
  graph.name = NULL,
  cell.var = NULL,
  cluster.name = "{graph.name}_{cell.var}_{metric}",
  modularity.fxn = 1,
  initial.membership = NULL,
  node.sizes = NULL,
  resolutions = seq(from = 0.1, to = 2, by = 0.1),
  optimisation.metric = c("nmi", "ari"),
  algorithm = 1,
  n.start = 10,
  n.iter = 10,
  random.seed = 0,
  group.singletons = TRUE,
  temp.file.location = NULL,
  edge.file.name = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{a Seurat object}

\item{graph.name}{the name of the knn graph to score.}

\item{cell.var}{The name(s) of the column(s) with cell type label variable
(must be in the object metadata). Multiple column names are accepted}

\item{cluster.name}{a (optionally 'glue') string used as the new metadata
column name (see \strong{Details} section)}

\item{modularity.fxn}{Modularity function (1 = standard; 2 = alternative).}

\item{initial.membership}{Passed to the `initial_membership` parameter
of `leidenbase::leiden_find_partition`.}

\item{node.sizes}{Passed to the `node_sizes` parameter of
`leidenbase::leiden_find_partition`.}

\item{resolutions}{the resolutions to compute clusters for}

\item{optimisation.metric}{one of "nmi" or "ari" or both (default).
The metric(s) to use to check clustering results against the \code{cell.var}.}

\item{algorithm}{Algorithm for modularity optimization (1 = original Louvain
algorithm; 2 = Louvain algorithm with multilevel refinement; 3 = SLM
algorithm; 4 = Leiden algorithm).}

\item{n.start}{Number of random starts.}

\item{n.iter}{Maximal number of iterations per random start.}

\item{random.seed}{Seed of the random number generator.}

\item{group.singletons}{Group singletons into nearest cluster. If FALSE, assign all singletons to
a "singleton" group}

\item{temp.file.location}{Directory where intermediate files will be written.
Specify the ABSOLUTE path.}

\item{edge.file.name}{Edge file to use as input for modularity optimizer jar.}

\item{verbose}{Print output}

\item{...}{Arguments passed to other methods}
}
\value{
the updated seurat object with the new metadata column(s)
}
\description{
Compute clusters for multiple resolutions and saves in the metadata the
clustering result that reaches the maximum NMI and/or ARI value for a given
cell-type label variable.
}
\details{
\code{cluster.name} can use the 'glue' syntax to avoid overwriting metadata
columns when multiple metrics and/or cell-type label variables are provided.
But it can be ignored otherwise. Injectable variables are "graph.name",
"cell.var" and "metric" (\strong{not} "\strong{optimisation.metric}"). They
must be \strong{flanked by single curly brackets} ("\{" and "\}").

For instance, if you prefer to name the clusters with an integration instead
of the \code{graph.name}, don't use the glue syntax for it (e.g.
"combat_\{cell.var\}_\{metric\}" would work, but any of
"\{combat\}_\{cell.var\}_\{metric\}" or
"\{integration\}_\{cell.var\}_\{metric\}" would throw an error)
}
\references{
Luecken, M. D., Büttner, M., Chaichoompu, K., Danese, A.,
Interlandi, M., Mueller, M. F., Strobl, D. C., Zappia, L., Dugas, M.,
Colomé-Tatché, M. & Theis, F. J. Benchmarking atlas-level data integration in
single-cell genomics. Nat Methods 19, 41–50 (2021).
\href{https://doi.org/10.1038/s41592-021-01336-8}{DOI}
}
\seealso{
\code{\link[Seurat]{FindClusters}}, \code{\link{ScoreARI}} and
\code{\link{ScoreNMI}}
}
