% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_phylo_profile.R
\name{from_input_to_profile}
\alias{from_input_to_profile}
\title{Create data for plotting profiles (from raw input to final dataframe)}
\usage{
from_input_to_profile(raw_input, rank_name, ref_taxon, taxa_tree,
    var1_aggregate_by, var2_aggregate_by, percent_cutoff,
    coortholog_cutoff_max, var1_cutoff, var2_cutoff, var1_relation,
    var2_relation, group_by_cat, cat_dt)
}
\arguments{
\item{raw_input}{input file (in long, wide, multi-fasta or orthoxml format)}

\item{rank_name}{taxonomy rank (e.g. "species","phylum",...)}

\item{ref_taxon}{selected reference taxon}

\item{taxa_tree}{input taxonomy tree (optional)}

\item{var1_aggregate_by}{aggregate method for VAR1 (min, max, mean or median)}

\item{var2_aggregate_by}{aggregate method for VAR2 (min, max, mean or median)}

\item{percent_cutoff}{min and max cutoffs for percentage of species present
in a supertaxon}

\item{coortholog_cutoff_max}{maximum number of co-orthologs allowed}

\item{var1_cutoff}{min and max cutoffs for var1}

\item{var2_cutoff}{min and max cutoffs for var2}

\item{var1_relation}{relation of var1 ("protein" for protein-protein or
"species" for protein-species)}

\item{var2_relation}{relation of var2 ("protein" for protein-protein or
"species" for protein-species)}

\item{group_by_cat}{group genes by their categories (TRUE or FALSE)}

\item{cat_dt}{dataframe contains gene categories}
}
\value{
dataframe for generating profile plot
}
\description{
Create data needed for plotting phylogenetic profile
from raw input file.
}
\examples{
raw_input <- system.file(
    "extdata", "test.main.long", package = "phyloprofile", mustWork = TRUE
)
rank_name <- "class"
ref_taxon <- "Mammalia"
taxa_tree <- NULL
var1_aggregate_by <- "max"
var2_aggregate_by <- "mean"
percent_cutoff <- c(0.0, 1.0)
coortholog_cutoff_max <- 10
var1_cutoff <- c(0.75, 1.0)
var2_cutoff <- c(0.5, 1.0)
var1_relation <- "protein"
var2_relation <- "species"
group_by_cat <- FALSE
cat_dt <- NULL
from_input_to_profile(
    raw_input,
    rank_name,
    ref_taxon,
    taxa_tree,
    var1_aggregate_by,
    var2_aggregate_by,
    percent_cutoff,
    coortholog_cutoff_max,
    var1_cutoff,
    var2_cutoff,
    var1_relation,
    var2_relation,
    group_by_cat,
    cat_dt
)
}
\seealso{
\code{\link{create_long_matrix}}, \code{\link{get_input_taxa_id}},
\code{\link{get_input_taxa_name}}, \code{\link{sort_input_taxa}},
\code{\link{parse_info_profile}}, \code{\link{reduce_profile}},
\code{\link{create_profile_data}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
