% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{fs_economic_factors_get}
\alias{fs_economic_factors_get}
\title{Get economic factor schedules for the specified connection string, scenario and transformation (by sequence order)}
\usage{
fs_economic_factors_get(connectionString, fs_id, transformation_sequence_order,
  criteria_squence_order)
}
\arguments{
\item{connectionString}{- the string representing the connection string...}

\item{fs_id}{The scenario ID}

\item{transformation_sequence_order}{Way to identify the the transformation entry to get the criteria from}

\item{criteria_sequence_order}{Way to identify the the criteria entry to get the characteristics from}
}
\value{
Data frame containing the financial instrument data.
}
\description{
Takes a connection string (connectionString) and scenario id (fs_id) and returns a datatable
with the instruments
}

