% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{fs_economic_factors_get_nc}
\alias{fs_economic_factors_get_nc}
\title{Gets Capital assumptions schedules for the specified connection string, scenario and transformation (by sequence order)
TODO: make this work on Description rather than transformation sequence! 
Takes a connection string (connectionString) and scenario id (fs_id) and returns a datatable
with the instruments}
\usage{
fs_economic_factors_get_nc(connectionString, fs_id,
  transformation_sequence_order, criteria_squence_order)
}
\arguments{
\item{connectionString}{- the string representing the connection string...}

\item{fs_id}{The scenario ID}

\item{transformation_sequence_order}{Way to identify the the transformation entry to get the criteria from}

\item{criteria_sequence_order}{Way to identify the the criteria entry to get the characteristics from}
}
\value{
Data frame containing the financial instrument data.
}
\description{
Gets Capital assumptions schedules for the specified connection string, scenario and transformation (by sequence order)
TODO: make this work on Description rather than transformation sequence! 
Takes a connection string (connectionString) and scenario id (fs_id) and returns a datatable
with the instruments
}

