% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.X.R
\name{g.X}
\alias{g.X}
\title{g.X: Generating data for simulation}
\usage{
g.X(N, n, e.val.x, e.vec.x, tau, mu, m)
}
\arguments{
\item{N}{real number, number of observations}

\item{n}{real number, number of grid points}

\item{e.val.x}{real vector n*1, eigenvalues}

\item{e.vec.x}{real valued matrix n*N, eigenvectors}

\item{tau}{range of the uniform distribution in KL expansion}

\item{mu}{real vector n*1, initial mean vector}

\item{m}{positive integer, number of eigenvectors are going to be used}
}
\value{
a real valued matrix n*N
}
\description{
This function simulate a data set by KL expansion
with using a known values for eigenvalues and
eigenvectors
}
\examples{
g.X(N=100,n=50,e.val.x=seq(1,0.1,length=50))
}

