% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.X.RKHS.R
\name{g.X.RKHS}
\alias{g.X.RKHS}
\title{Generating RKHS smoothing mean of a dataset}
\usage{
g.X.RKHS(N, n, e.val.x, e.vec.x, tau, phi, mu, m, e.val.z)
}
\arguments{
\item{N}{real number, number of observations}

\item{n}{real number, number of grid points}

\item{e.val.x}{real vector n*1, eigenvalues}

\item{e.vec.x}{real valued matrix n*N, eigenvectors}

\item{tau}{range of the uniform distribution in KL expansion}

\item{phi}{real number, penalty parameter}

\item{mu}{real vector n*1, initial mean vector}

\item{m}{positive integer, number of eigenvectors are going to be used}

\item{e.val.z}{real valued matrix n*N, eigenvectors of noise}
}
\value{
a real valued matrix n*N
}
\description{
This function create a RKHS smoothing mean from an existing data set
with known eigenvalues and eigenvectors
}

