% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_gene_age.R
\name{gene_age_plotDf}
\alias{gene_age_plotDf}
\title{Create data for plotting gene ages}
\usage{
gene_age_plotDf(gene_ageDf)
}
\arguments{
\item{gene_ageDf}{data of estimated gene age}
}
\value{
A dataframe for plotting gene age plot
}
\description{
Create data for plotting gene ages
}
\examples{
gene_ageDf <- data.frame(
geneID = c("OG_1017", "OG_1019"),
cat = c("0000001", "0000001"),
age = c("07_LUCA", "07_LUCA")
)
gene_age_plotDf(gene_ageDf)
}
\seealso{
\code{\link{estimate_gene_age}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
