\name{geoPolygons}

\Rdversion{1.1}

\alias{geoPolygons}
\alias{geoPolygons,list,data.frame-method}
\alias{geoPolygons,list,missing-method}

\title{
Create GeoPolygon object
}

\description{
Create a geoPolygons object
}

\usage{
geoPolygons(x, d, ...)
}

\arguments{
  \item{x}{A GeoVector object}
  \item{d}{A GeoVector object}
  \item{\dots}{Additional arguments }
}

\value{
geoPolygons object
}

\author{
Robert J. Hijmans
}


\keyword{ spatial  }
