% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{GetMiscIntegrations}
\alias{GetMiscIntegrations}
\alias{GetMiscScores}
\alias{IntegrationScores}
\title{Retrieve integration scores from a Seurat object}
\usage{
GetMiscIntegrations(object, search = NULL)

GetMiscScores(object, search = NULL)

IntegrationScores(object, scaled = FALSE)
}
\arguments{
\item{object}{a Seurat object}

\item{search}{a character vector of names to search for through
case-insensitive exact match. \code{NULL} disables the search (default)}

\item{scaled}{whether to query the unscaled (default) or scaled scores (if
scaling has been performed)}
}
\value{
\code{GetMiscIntegrations}: a character vector of integration names,
or \code{NULL} when the object doesn't have 'si_scores' in Misc or when
search did not return any result.

\code{GetMiscScores}: a character vector of score names, or \code{NULL} when
the object doesn't have 'si_scores' in Misc or when search did not return any
result.

\code{IntegrationScores}: either \code{NULL} if the requested object does not
exist, otherwise a tibble. The first column contains the name of the
integrations, and each following column corresponds to a score.
}
\description{
Scores are stored as tibble objects in Misc. Slot for raw scores is named
'si_scores' while scaled scores are found under 'si_scaled.scores'.

\code{GetMiscIntegrations}: Get (or search for) integration names in the
score tibble

\code{GetMiscScores}:  Get (or search for) score names in the score tibble

\code{IntegrationScores}: Get the tibble with scaled or unscaled scores
}
