% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oma_browser.R
\name{get_all_domains_oma}
\alias{get_all_domains_oma}
\title{Create domain annotation dataframe from a raw OMA dataframe}
\usage{
get_all_domains_oma(final_oma_df)
}
\arguments{
\item{final_oma_df}{raw OMA data for a list of input IDs}
}
\value{
domain annotation in a dataframe to input into PhyloProfile, which
contains seed ID, ortholog ID, ortholog length, annotated feature, start
and end position of that feature.
}
\description{
Create domain annotation dataframe from a raw OMA dataframe
}
\examples{
oma_data <- get_data_for_one_oma("HUMAN29397", "OG")
get_all_domains_oma(oma_data)
}
\seealso{
\code{\link{get_data_for_one_oma}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
