% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oma_browser.R
\name{get_all_fasta_oma}
\alias{get_all_fasta_oma}
\title{Get all fasta sequences from a raw OMA dataframe}
\usage{
get_all_fasta_oma(final_oma_df)
}
\arguments{
\item{final_oma_df}{raw OMA data for a list of input IDs}
}
\value{
dataframe contains all fasta sequences
}
\description{
Get all fasta sequences from a raw OMA dataframe
}
\examples{
oma_data <- get_data_for_one_oma("HUMAN29397", "OG")
get_all_fasta_oma(oma_data)
}
\seealso{
\code{\link{get_data_for_one_oma}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
