% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_taxa_groups.R
\name{get_common_ancestor}
\alias{get_common_ancestor}
\title{Generate the in_group}
\usage{
get_common_ancestor(in_group, rank, name_list, selected_in_group, 
    ncbi_id_list)
}
\arguments{
\item{in_group}{list of taxa}

\item{rank}{selected rank}

\item{name_list}{contains "ncbiID", "fullName", "rank", "parentID"}

\item{selected_in_group}{contains "abbrName, "ncbiID", fullName", "strain",
"genus"..}

\item{ncbi_id_list}{list of input taxon IDs}
}
\value{
common anchestor
}
\description{
Generate the in_group
}
\examples{
in_group <- c("Aeropyrum pernix", "Agrobacterium fabrum")
rank <- "species"
name_list_file <- system.file(
    "extdata", "data/taxonNamesFull.txt",
    package = "phyloprofile", mustWork = TRUE
)
name_list <- as.data.frame(data.table::fread(name_list_file))
selected_in_group <- get_taxonomy_matrix(FALSE, NULL)
ncbi_id_list <- c(56636, 1176649)
get_common_ancestor(
    in_group,
    rank,
    name_list,
    selected_in_group,
    ncbi_id_list
)
}
\author{
Carla Mölbert (carla.moelbert@gmx.de)
}
