% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{get_confusion_matrix}
\alias{get_confusion_matrix}
\title{Get Confusion Matrix}
\usage{
get_confusion_matrix(lr_result, cut_off = 0.5)
}
\arguments{
\item{lr_result}{A \code{list} value of class 'logistic_regression', found from running the 'get_beta_estimate' function in this package. The list contains the following objects:
\describe{
 \item{initial_beta}{The initial values of coefficients.}
 \item{beta_estimate}{The estimated coefficients of the logistic regression.}
 \item{response}{The \code{double} vector containing the response used for the estimation.}
 \item{predictors}{The \eqn{n \times p} \code{double} value of the matrix containing the values of the predictors used for the estimation.}
}}

\item{cut_off}{A \code{double} value of the cut_off used to set the predicted values to 0 or 1.}
}
\value{
A \code{list} containing the following objects:
\describe{
 \item{tp}{The \code{double} value of True Positives. Count of data points for which actual and predicted value is 1.}
 \item{tn}{The \code{double} value of True Negatives. Count of data points for which actual and predicted value is 0.}
 \item{fp}{The \code{double} value of False Positives. Count of data points for which actual value is 0 and predicted value is 1.}
 \item{fn}{The \code{double} value of False Negatives. Count of data points for which actual value is 1 and predicted value is 0.}
 \item{total}{The \code{double} value of total number of data points.}
}
}
\description{
This function computes Confusion Matrix for the found coefficients of logistic regression.
}
\author{
Saksham Goel
}
