% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_core_gene.R
\name{get_core_gene}
\alias{get_core_gene}
\title{Identify core genes for a list of selected taxa}
\usage{
get_core_gene(rank_name, taxa_core, processed_profile_data,
    var1_cutoff, var2_cutoff, percent_cutoff, core_coverage)
}
\arguments{
\item{rank_name}{taxonomy rank (e.g. "species", "genus", "family")}

\item{taxa_core}{name list of selected taxa}

\item{processed_profile_data}{dataframe contains the full processed
phylogenetic profiles}

\item{var1_cutoff}{cutoff for var1}

\item{var2_cutoff}{cutoff for var2}

\item{percent_cutoff}{cutoff for percentage of species present in each
supertaxon}

\item{core_coverage}{the least number of selected taxa should be considered}
}
\value{
A list of core genes
}
\description{
Identify core genes for a list of selected taxa
}
\examples{
data("full_processed_profile", package="phyloprofile")
rank_name <- "class"
taxa_core <- c("Mammalia", "Mucorales", "Alphaproteobacteria")
processed_profile_data <- full_processed_profile
var1_cutoff <- c(0.75, 1.0)
var2_cutoff <- c(0.75, 1.0)
percent_cutoff <- c(0.0, 1.0)
core_coverage <- 1
get_core_gene(
    rank_name,
    taxa_core,
    processed_profile_data,
    var1_cutoff, var2_cutoff,
    percent_cutoff, core_coverage
)
}
\seealso{
\code{\link{parse_info_profile}} for creating a full processed
profile dataframe
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
