% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_profile.R
\name{get_data_clustering}
\alias{get_data_clustering}
\title{Get data for calculating the distance matrix}
\usage{
get_data_clustering(data, profile_type, var1_aggregate_by, 
    var2_aggregate_by)
}
\arguments{
\item{data}{processed profile data}

\item{profile_type}{type of data used for calculating the distance matrix. 
Either "binary" (consider only the presence/absence status of orthlogs), or
"var1"/"var2" for taking values of the additional variables into account.}

\item{var1_aggregate_by}{aggregate method for VAR1 (min, max, mean or median)}

\item{var2_aggregate_by}{aggregate method for VAR2 (min, max, mean or median)}
}
\value{
A wide dataframe contains values for calculating distance matrix.
}
\description{
Get data for calculating the distance matrix
}
\note{
Documented by Vinh Tran (tran@bio.uni-frankfurt.de)
}
\examples{
data("full_processed_profile", package="phyloprofile")
data <- full_processed_profile
profile_type <- "binary"
var1_aggregate_by <- "max"
var2_aggregate_by <- "mean"
get_data_clustering(data, profile_type, var1_aggregate_by, var2_aggregate_by)
}
\seealso{
\code{\link{from_input_to_profile}}
}
\author{
Carla Mölbert (carla.moelbert@gmx.de)
}
