% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oma_browser.R
\name{get_data_for_one_oma}
\alias{get_data_for_one_oma}
\title{Get OMA info for a query protein and its orthologs}
\usage{
get_data_for_one_oma(seed_id, ortho_type)
}
\arguments{
\item{seed_id}{protein query ID in OMA or UniProt format}

\item{ortho_type}{type of OMA orthologs}
}
\value{
data frame contains info for all sequences of that OMA group (or HOG)
}
\description{
Get taxonomy IDs, sequences and annotations for an OMA
orthologous group (or OMA HOG).
}
\examples{
get_data_for_one_oma("HUMAN29397", "OG")
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
