% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oma_browser.R
\name{get_data_for_one_ortholog}
\alias{get_data_for_one_ortholog}
\title{Get taxonomy ID, sequence and annotation for one OMA sequence}
\usage{
get_data_for_one_ortholog(id)
}
\arguments{
\item{id}{oma ID of a ortholog}
}
\value{
data frame
}
\description{
Get taxonomy ID, sequence and annotation for one OMA sequence
}
\examples{
\dontrun{
get_data_for_one_ortholog(HUMAN29397)
}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
