% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_profile.R
\name{get_dendrogram}
\alias{get_dendrogram}
\title{Plot dendrogram tree}
\usage{
get_dendrogram(dd)
}
\arguments{
\item{dd}{dendrogram object}
}
\value{
A dendrogram plot
}
\description{
Plot dendrogram tree
}
\examples{
data("full_processed_profile_large", package="phyloprofile")
data <- full_processed_profile_large
profile_type <- "binary"
profiles <- get_data_clustering(
    data, profile_type, var1_aggregate_by, var2_aggregate_by)
dist_method <- "mutual_information"
distance_matrix <- get_distance_matrix(profiles, dist_method)
cluster_method <- "complete"
dd <- cluster_data_dend(distance_matrix, cluster_method)
get_dendrogram(dd)
}
\seealso{
\code{\link{cluster_data_dend}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
