% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_profile.R
\name{get_distance_matrix}
\alias{get_distance_matrix}
\title{Calculate the distance matrix}
\usage{
get_distance_matrix(profiles, method)
}
\arguments{
\item{profiles}{profile data for distance calculating}

\item{method}{distance calculation method ("euclidean", "maximum", 
"manhattan", "canberra", "binary", "distance_correlation", 
"mutual_information" or "pearson" for binary data; "distance_correlation" or
"mutual_information" for non-binary data)}
}
\value{
A distance matrix for input phylogenetic profiles.
}
\description{
Calculate the distance matrix
}
\note{
Documented by Vinh Tran (tran@bio.uni-frankfurt.de)
}
\examples{
data("full_processed_profile_large", package="phyloprofile")
data <- full_processed_profile_large
profile_type <- "binary"
profiles <- get_data_clustering(
    data, profile_type, var1_aggregate_by, var2_aggregate_by)
method <- "mutual_information"
get_distance_matrix(profiles, method)
}
\seealso{
\code{\link{get_data_clustering}}
}
\author{
Carla Mölbert (carla.moelbert@gmx.de)
}
