% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_domain_input.R
\name{get_domain_file}
\alias{get_domain_file}
\title{Get domain file from a single (concatenate) file}
\usage{
get_domain_file(input_file)
}
\arguments{
\item{input_file}{concatenate domain file}
}
\value{
Domain file and its complete directory path
}
\description{
Get domain file from a single (concatenate) file
}
\examples{
\dontrun{
input_file <- system.file(
    "extdata", "domain_files/OG_1009.domains",
    package = "phyloprofile", mustWork = TRUE
)
get_domain_file("lca-micors")
}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
