% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_domain_input.R
\name{get_domain_folder}
\alias{get_domain_folder}
\title{Get domain files from a folder}
\usage{
get_domain_folder(seed, domain_path)
}
\arguments{
\item{seed}{seed ID}

\item{domain_path}{path to domain folder}
}
\value{
Domain file and its complete directory path
}
\description{
Get domain files from a folder
}
\examples{
\dontrun{
domain_path <- paste0(
    path.package("phyloprofile", quiet = FALSE), "extdata/domain_files"
)
get_domain_online("OG_1009", domain_path)
}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
