% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oma_browser.R
\name{get_domain_from_url}
\alias{get_domain_from_url}
\title{Get domain annotation from OMA Browser}
\usage{
get_domain_from_url(domainURL)
}
\arguments{
\item{domainURL}{URL address for domain annotation of ONE OMA id}
}
\value{
data frame contains feature names, start and end positions
}
\description{
Get domain annotation from OMA Browser based on a URL or a
raw data frame contains annotation info from OMA
}
\examples{
\dontrun{
get_domain_from_url("https://omabrowser.org/api/protein/7916808/domains/")
}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
