% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_domain_input.R
\name{get_domain_online}
\alias{get_domain_online}
\title{Get domain file(s) for online data set}
\usage{
get_domain_online(demo_data)
}
\arguments{
\item{demo_data}{demo data name (either lca-micros or ampk-tor)}
}
\value{
Domain file and its complete directory path
}
\description{
Get domain file(s) for online data set
}
\examples{
\dontrun{
get_domain_online("lca-micors")
}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
