% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fasta_seqs.R
\name{get_fasta_seqs}
\alias{get_fasta_seqs}
\title{Get fasta sequences}
\usage{
get_fasta_seqs(data_in, filein, demo_data, input_type_fasta,
    concat_fasta, path, dir_format, file_ext, id_format)
}
\arguments{
\item{data_in}{a dataframe of the input phylogenetic profiles, that contains
at least 3 columns: geneIDs, orthoIDs and ncbiIDs}

\item{filein}{main input file (for checking input type)}

\item{demo_data}{name of demo data set (either "ampk-tor", "lca-micros",
or "none")}

\item{input_type_fasta}{source of fasta sequences ("Concatenated fasta file"
or "Fasta folder")}

\item{concat_fasta}{input concatenated fasta file}

\item{path}{path to fasta folder}

\item{dir_format}{directory format (either "path/speciesID.fa*" or
"path/speciesID/speciesID.fa*")}

\item{file_ext}{fasta file extension ("fa", "fasta", "fas" or "txt")}

\item{id_format}{fasta header format (">speciesID:seqID",
">speciesID@seqID", ">speciesID|seqID" or only "seqID")}
}
\value{
A dataframe contains fasta sequences
}
\description{
Get fasta sequences
}
\examples{
data("main_long_raw", package="phyloprofile")
data_in <- main_long_raw
filein <- system.file(
    "extdata", "test.main.long", package = "phyloprofile", mustWork = TRUE
)
demo_data <- "none"
input_type_fasta <- "Concatenated fasta file"
concat_fasta <- system.file(
    "extdata", "fasta_files/concatenatedFile.fa",
    package = "phyloprofile", mustWork = TRUE
)
# the following parameters are conly required if upload a fasta folder!
path <- NULL
dir_format <- NULL
file_ext <- NULL
id_format <- NULL
# get fasta sequences
get_fasta_seqs(
    data_in, filein, demo_data,
    input_type_fasta,
    concat_fasta,
    path,
    dir_format,
    file_ext,
    id_format
)
}
\seealso{
\code{\link{check_input_validity}}, \code{\link{str_reverse}},
\code{\link{main_long_raw}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
