% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_taxa_groups.R
\name{get_features}
\alias{get_features}
\title{get the list with all the features in the gene}
\usage{
get_features(selected_gene, domains)
}
\arguments{
\item{selected_gene}{gene to get the feartures for}

\item{domains}{contains "seedID", "orthoID", "feature", "start",   "end"}
}
\value{
dataframe for the specific gene containing "seedID",  "orthoID",
"feature", "start",   "end"
}
\description{
get the list with all the features in the gene
}
\author{
Carla Mölbert (carla.moelbert@gmx.de)
}
