% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_taxonomy_matrix.R
\name{get_ids_rank}
\alias{get_ids_rank}
\title{Get taxonomy info for a list of taxa}
\usage{
get_ids_rank(input_taxa, currentNCBIinfo)
}
\arguments{
\item{input_taxa}{input taxon list}

\item{currentNCBIinfo}{table/dataframe of the current taxonomy data
(taxonNamesFull.txt)}
}
\value{
A list of 3 dataframes: idList, rankList and reducedInfoList
}
\description{
Get NCBI taxonomy IDs, ranks and names for an input
taxon list
}
\examples{
input_taxa <- c("272557", "176299")
ncbi_filein <- system.file(
    "extdata", "data/taxonNamesFull.txt",
    package = "phyloprofile", mustWork = TRUE
)
currentNCBIinfo <- as.data.frame(data.table::fread(ncbi_filein))
get_ids_rank(input_taxa, currentNCBIinfo)
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
