% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_phylo_profile.R
\name{get_input_taxa_id}
\alias{get_input_taxa_id}
\title{Get ID list of input taxa from the main input}
\usage{
get_input_taxa_id(raw_profile)
}
\arguments{
\item{raw_profile}{long dataframe of input profile}
}
\value{
List of all input taxon IDs
}
\description{
Get ID list of input taxa from the main input
}
\examples{
data("main_long_raw", package="phyloprofile")
get_input_taxa_id(main_long_raw)
}
\seealso{
\code{\link{create_long_matrix}}, \code{\link{main_long_raw}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
