% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_phylo_profile.R
\name{get_input_taxa_name}
\alias{get_input_taxa_name}
\title{Get NCBI taxon names for a selected list of taxa}
\usage{
get_input_taxa_name(rank_name, taxon_IDs)
}
\arguments{
\item{rank_name}{taxonomy rank (e.g. "species","phylum",...)}

\item{taxon_IDs}{list of taxon IDs (check get_input_taxa_id())}
}
\value{
List of full names, taxonomy ranks and parent IDs for the input taxa
}
\description{
Get NCBI taxon names from "data/taxonNamesReduced.txt" for
a selected list of taxon
}
\examples{
taxon_ids <- c("ncbi10020", "ncbi10090")
get_input_taxa_name("species", taxon_ids)
}
\seealso{
\code{\link{get_input_taxa_id}} for getting input taxon IDs,
\code{\link{get_name_list}} for getting the full taxon name list
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
