% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_phylo_profile.R
\name{get_name_list}
\alias{get_name_list}
\title{Get list of pre-installed NCBI taxon names}
\usage{
get_name_list()
}
\value{
List of taxon IDs, their full names, taxonomy ranks and parent IDs
}
\description{
Get all NCBI taxon names from "data/taxonNamesReduced.txt"
}
\examples{
get_name_list()
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
