% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oma_browser.R
\name{get_oma_members}
\alias{get_oma_members}
\title{Get OMA members}
\usage{
get_oma_members(id, ortho_type)
}
\arguments{
\item{id}{ID of the seed protein (OMA or UniProt ID)}

\item{ortho_type}{type of OMA orthologs: either "HOG", "OG"
(orthologous group) or "PAIR" (orthologous pair - CURRENTLY NOT WORKING)}
}
\value{
list of ortholog members
}
\description{
Get OMA orthologs for a seed protein from OMA Browser
}
\examples{
get_oma_members("HUMAN29397", "OG")
}
\author{
Carla Mölbert {carla.moelbert@gmx.de}
}
