% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_taxa_groups.R
\name{get_p_values}
\alias{get_p_values}
\title{Decide if the gene is significant}
\usage{
get_p_values(in_group, out_group, variable, gene, parameters)
}
\arguments{
\item{in_group}{contains "supertaxon", "geneID", "ncbiID", "orthoID",
"var1", "var2", "paralog", "abbrName", "taxonID", "fullname",
"supertaxonID", "rank", "category", "presSpec", "mVar1", "mVar2"}

\item{out_group}{as in-group but with information containing the out-group}

\item{variable}{variable(s) to claculate the plots for}

\item{gene}{gene to calculate the p-values for}

\item{parameters}{contains "show_p_value","highlight_significant",
"significance", "var1_id", "var2_id", "x_size_gc", "y_size_gc",
"interesting_features", "angle_gc", "legend_gc", "legend_size_gc"}
}
\value{
return the pvalues
}
\description{
Decide if the gene is significant
}
\author{
Carla Mölbert (carla.moelbert@gmx.de)
}
