% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_taxa_groups.R
\name{get_prefix_features}
\alias{get_prefix_features}
\title{Get the database for each feature in a specific gene}
\usage{
get_prefix_features(data)
}
\arguments{
\item{data}{contains "seedID", "orthoID", "feature", "start", "end"}
}
\value{
list of prefixes for the features
}
\description{
Get the database for each feature in a specific gene
}
\author{
Carla Mölbert (carla.moelbert@gmx.de)
}
