% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_color_scheme.R
\name{get_qual_col_for_vector}
\alias{get_qual_col_for_vector}
\title{Get color for a list of items}
\usage{
get_qual_col_for_vector(x)
}
\arguments{
\item{x}{input list}
}
\value{
list of colors for each element (same elements will have the same
color)
}
\description{
Get color for a list of items
}
\examples{
items <- c("a", "b", "c")
get_qual_col_for_vector(items)
}
\seealso{
\code{\link{qualitative_colours}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
