% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oma_browser.R
\name{get_selected_fasta_oma}
\alias{get_selected_fasta_oma}
\title{Get selected fasta sequences from a raw OMA dataframe}
\usage{
get_selected_fasta_oma(final_oma_df, seq_id)
}
\arguments{
\item{final_oma_df}{raw OMA data for a list of input IDs}

\item{seq_id}{sequence need to be returned}
}
\value{
required sequence in fasta format
}
\description{
Get selected fasta sequences from a raw OMA dataframe
}
\examples{
oma_data <- get_data_for_one_oma("HUMAN29397", "OG")
get_selected_fasta_oma(oma_data, "HUMAN29397")
}
\seealso{
\code{\link{get_data_for_one_oma}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
