% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_taxa_groups.R
\name{get_selected_subset}
\alias{get_selected_subset}
\title{Get the subset depending on the choosen rank}
\usage{
get_selected_subset(rank, in_group, name_list, taxa_list)
}
\arguments{
\item{rank}{selected rank}

\item{in_group}{list of taxa}

\item{name_list}{contains "ncbiID", "fullName", "rank", "parentID"}

\item{taxa_list}{contains "abbrName, "ncbiID", fullName", "strain", "genus"}
}
\value{
list of prefixes for the features
}
\description{
Get the subset depending on the choosen rank
}
\author{
Carla Mölbert (carla.moelbert@gmx.de)
}
