% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_taxa_groups.R
\name{get_significant_genes}
\alias{get_significant_genes}
\title{Get the dataframe with the significant genes}
\usage{
get_significant_genes(in_group, selected_genes_list, rank, var,
    use_common_ancestor, reference_taxon, parameters, ncbi_id_list, 
    data_full, right_format_features, domains)
}
\arguments{
\item{in_group}{list of taxa}

\item{selected_genes_list}{list of genes}

\item{rank}{selected taxonamy rank}

\item{var}{variable for which to  calculate the significance}

\item{use_common_ancestor}{boolean if the common anchestor should be used}

\item{reference_taxon}{taxon which is used as reference}

\item{parameters}{contains "show_p_value","highlight_significant",
"significance", "var1_id", "var2_id", "x_size_gc", "y_size_gc",
"interesting_features", "angle_gc", "legend_gc", "legend_size_gc"}

\item{ncbi_id_list}{list of ncbi ids}

\item{data_full}{full processed main data}

\item{right_format_features}{boolean if the features have the right format}

\item{domains}{dataframe holding the domain input}
}
\value{
dataframe with the significant genes
}
\description{
Get the dataframe with the significant genes
}
\examples{
data("main_long_raw", package="phyloprofile")
data("full_processed_profile", package="phyloprofile")
in_group <- c("Aeropyrum pernix", "Agrobacterium fabrum")
selected_genes_list <- "OG_1017"
rank <- "species"
var <- ""
use_common_ancestor <- FALSE
reference_taxon <- "Homo sapiens"
ncbi_id_list <- get_input_taxa_id(main_long_raw)
data_full <- full_processed_profile
right_format_features <- TRUE
domain_file <- system.file(
    "extdata", "domain_files/OG_1017.domains",
    package = "phyloprofile", mustWork = TRUE
)
domains <- parse_domain_input("OG_1017", domain_file, "file")
parameters <- list(
    "show_p_value" = TRUE,
    "highlight_significant" = FALSE,
    "significance" = 0.05,
    "var1_id" = "1st variable",
    "var2_id" = "2nd variable",
    "x_size_gc" = 9,
    "y_size_gc" = 9,
    "interesting_features" = NULL,
    "angle_gc" = 90,
    "legend_gc" = "none",
    "legend_size_gc" = 9,
    "p_values_size" = 9
)
get_significant_genes(
    in_group,
    selected_genes_list,
    rank,
    var,
    use_common_ancestor,
    reference_taxon,
    parameters,
    ncbi_id_list,
    data_full,
    right_format_features,
    domains
)
}
\author{
Carla Mölbert (carla.moelbert@gmx.de)
}
