% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_phylo_profile.R
\name{get_taxonomy_matrix}
\alias{get_taxonomy_matrix}
\title{Get taxonomy matrix}
\usage{
get_taxonomy_matrix(subset_taxa_check, taxon_IDs)
}
\arguments{
\item{subset_taxa_check}{subset taxonomy matrix based on input taxon IDs
(TRUE/FALSE)}

\item{taxon_IDs}{list of input taxon IDs}
}
\value{
Data frame contains the (subset of) taxonomy matrix
}
\description{
Get full taxonomy matrix from "data/taxonomyMatrix.txt" or
only a subset of matrix based on an input taxon list
}
\examples{
# get full pre-installed taxonomy matrix
get_taxonomy_matrix(FALSE, NULL)
# get taxonomy matrix for a list of taxon IDs
taxon_ids <- c("ncbi10020", "ncbi10090")
get_taxonomy_matrix(TRUE, taxon_ids)
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
