% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_taxa_groups.R
\name{get_taxonomy_ranks}
\alias{get_taxonomy_ranks}
\title{Create a list with all main taxanomy ranks}
\usage{
get_taxonomy_ranks()
}
\value{
A list of all main ranks (from strain to superkingdom)
}
\description{
Create a list with all main taxanomy ranks
}
\examples{
get_taxonomy_ranks()
}
\author{
Carla Mölbert (carla.moelbert@gmx.de)
}
