% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBER.R
\name{getBER}
\alias{getBER}
\title{get Balanced Error Rate from classification analysis}
\usage{
getBER(actual, predicted)
}
\arguments{
\item{actual}{Vector of actual classifications of samples}

\item{predicted}{Vector of predicted classifications of samples}
}
\value{
Balanced Error Rate (BER)
}
\description{
get Balanced Error Rate from classification analysis
}
