% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_S4_helpers.R
\name{getCache}
\alias{getCache}
\title{Get current cache of conda environments}
\usage{
getCache(from = c("env", "cache"))
}
\arguments{
\item{from}{Either "env" or "cache". Where to load cache from.}
}
\value{
a \linkS4class{CondaEnvManager} object
}
\description{
Get current cache of conda environments from package environment or cache on disk
}
