% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_S4_helpers.R
\name{getCachePath}
\alias{getCachePath}
\title{Get path to package config cache}
\usage{
getCachePath(include.file = TRUE)
}
\arguments{
\item{include.file}{Set to \code{FALSE} to retrieve the directory containing
the cache file. Return the path to the file by default.}
}
\value{
A single character, the path to the cache file or directory
}
\description{
Get path to package config cache file or directory
}
