% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GetNeighborsPerBatch}
\alias{GetNeighborsPerBatch}
\alias{GetNeighboursPerBatch}
\alias{GetNeighborsPerBatch,Seurat,character-method}
\alias{GetNeighborsPerBatch,Seurat,data.frame-method}
\alias{GetNeighborsPerBatch,Graph,ANY-method}
\alias{GetNeighborsPerBatch,matrix,ANY-method}
\alias{GetNeighborsPerBatch,Neighbor,ANY-method}
\alias{GetNeighborsPerBatch,Matrix,ANY-method}
\title{Calculate number of nearest neighbours between batches}
\usage{
GetNeighborsPerBatch(
  object,
  batch.var,
  graph.name = "RNA_nn",
  count.self = TRUE
)

\S4method{GetNeighborsPerBatch}{Seurat,character}(
  object,
  batch.var,
  graph.name = "RNA_nn",
  count.self = TRUE
)
}
\arguments{
\item{object}{a \code{Seurat} object}

\item{batch.var}{name of a column in the \code{Seurat} object's metadata
containing batch information}

\item{graph.name}{name of a \code{Graph} or \code{Neighbor} instance stored in
\code{object}. When available, prefer the distance based network to the
connectivities graph (especially when computed with the UMAP method).}

\item{count.self}{whether to include self-to-self vertices in the calculation}
}
\value{
a square count matrix of size number of batches (see \strong{Details}
section)
}
\description{
Calculate number of nearest neighbours between batches out of a knn graph
}
\details{
The output matrix will likely not be symmetrical. This is due to the approximate
nature of the nearest neighbour search algorithm used to compute the knn
graph. It must be read by \strong{row}. For instance, the number of times
cells of batch 1 have cells of batch 3 in their nn is matrix[1,3]
}
\seealso{
The classes \code{\link[SeuratObject:Graph]{Graph}} and
\code{\link[SeuratObject:Neighbor]{Neighbor}}
}
\keyword{internal}
