% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GetPropIntraBatch}
\alias{GetPropIntraBatch}
\title{Calculate proportion of nearest neighbours within batches}
\usage{
GetPropIntraBatch(
  object,
  batch.var,
  graph.name = "RNA_nn",
  count.self = TRUE,
  per.batch = TRUE
)
}
\arguments{
\item{object}{a \code{Seurat} object}

\item{batch.var}{name of a column in the \code{Seurat} object's metadata
containing batch information}

\item{graph.name}{name of a \code{Graph} or \code{Neighbor} instance stored in
\code{object}. When available, prefer the distance based network to the
connectivities graph (especially when computed with the UMAP method).}

\item{count.self}{whether to include self-to-self vertices in the calculation}

\item{per.batch}{whether to keep proportions per batch or to aggregate
everything}
}
\value{
a vector of length 1 or number of batches, depending on
\code{per.batch} argument value
}
\description{
Calculate the proportion of nearest neighbours within batches out of a knn
graph
}
\seealso{
\code{\link{GetPropInterBatch}} and \code{\link{GetNeighborsPerBatch}}
}
