% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVIRanks.R
\name{getVIRank}
\alias{getVIRank}
\title{Extract autoselected variables from MUVR model object}
\usage{
getVIRank(MUVRclassObject, model = "mid")
}
\arguments{
\item{MUVRclassObject}{An object of MUVR class}

\item{model}{Which model to use ("min", "mid" (default), or "max")}
}
\value{
Data frame with order, name and average rank of variables (`order`, `name` & `rank`)
}
\description{
Extract autoselected variables from MUVR model object
}
