% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab_token.R
\name{grab_token}
\alias{grab_token}
\title{Grab protocol token}
\usage{
grab_token(protocol)
}
\arguments{
\item{protocol}{The protocol name (e.g. 'test-adni3').}
}
\value{
a string of protocol access token
}
\description{
grab access token from Rauth file for a protocol
}
\examples{
\donttest{
  grab_token("test-adni3")
}
}
