% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_cat.R
\name{graph_cat}
\alias{graph_cat}
\title{Charts for categorical variables}
\usage{
graph_cat(data, var1, var2 = NULL)
}
\arguments{
\item{data}{Dataset to which the variables belong.}

\item{var1}{Variable 1 for which the bar and pie charts are created.}

\item{var2}{Variable 2 for which the stacked bar chart is created with variable 1.}
}
\value{
Bar chart and pie chart, as applicable.
}
\description{
Creates a bar chart and a pie chart for a categorical variable, if the user wishes
it also creates a stacked bar chart between the given variables.
}
\examples{
diagrams_edacat <- graph_cat(iris, "Species");
diagram_ab <- graph_cat(farm, "animals", "food");
}
