% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr.cv.f.par.R
\name{gwr.cv.f.par}
\alias{gwr.cv.f.par}
\title{for any given bandwidth, gwr.cv.f.par compute the cv score. Done by running the n GLMs accross multiple cores/nodes}
\usage{
gwr.cv.f.par(bandwidth, y, x, coords, gweight, verbose = TRUE,
  longlat = FALSE, RMSE = FALSE, weights, show.error.messages = TRUE,
  ncores)
}
\description{
for any given bandwidth, gwr.cv.f.par compute the cv score. Done by running the n GLMs accross multiple cores/nodes
}
