% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr_par.R
\name{gwr_par}
\alias{gwr_par}
\title{GWR computation}
\usage{
gwr_par(formula, data, coords, bandwidth, weights = NULL,
  kernel = "gaussian", longlat = NULL, se.fit = FALSE,
  diagnostic = FALSE, adapt = F, ncores = NULL)
}
\arguments{
\item{formula}{Formula of the GWR}

\item{data}{Dataset (either data.frame of Spatial*Dataframe object)}

\item{coords}{A two-columns matrix with the coordinates as X-Y if data is a data.frame}

\item{kernel}{Character string. Weight kernel. Either "gaussian", "bisquare" or "boxcar"}

\item{longlat}{TRUE if coordinates are longitude-latitude in decimal degrees, in which case, distances are measured in kilometers}

\item{se.fit}{Logical. TRUE if standard errors of the fit should be assessed.}

\item{diagnostic}{Logical. If TRUE, the following metrics are displayed: AIC, AICc, RSS, Effective numbers of parameters and of degrees of freedom.}

\item{adapt}{Logical. TRUE if Adaptative bandwith, FALSE if fixed}

\item{ncores}{Number of cores in which the computation should be spanned}

\item{bandwith}{Bandwidth. Can be computed with gwr.sel.par()}
}
\value{
A fitted GWR
}
\description{
This function computes a GWR (Geographically Weighted Regression). It spans the computation across several nodes in a cluster.
}
