% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H0_reference.R
\name{H0_reference}
\alias{H0_reference}
\title{make referenc edistribution}
\usage{
H0_reference(Y, n = 1000, fitness = c("Q2", "BER", "MISS", "AUROC"))
}
\arguments{
\item{Y}{the taget variable}

\item{n}{number of permutations to run}

\item{fitness}{number of repetitions for each permutation (defaults to value of actual model)}

\item{...}{some more argument for sampling frm distribution}
}
\value{
a histogram of reference distribution
}
\description{
make referenc edistribution
}
