% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/height.R
\name{height}
\alias{height}
\title{height}
\usage{
height(b, u)
}
\arguments{
\item{b}{height}

\item{u}{units}
}
\value{
text including height with units
}
\description{
height
}
\examples{
height(5, "ft")
height(100, "cm")

}
