% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Print "Hello world!" and return TRUE invisibly.}
\usage{
hello()
}
\value{
Always returns `TRUE` invisibly.
}
\description{
Print "Hello world!" and return TRUE invisibly.
}
\examples{
hello()
}
