% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{helloR}
\alias{helloR}
\title{Prints "Hello" for any name!}
\usage{
helloR(name)
}
\arguments{
\item{name}{A name}
}
\value{
It returns "Hello name!"
}
\description{
Prints "Hello" for any name!
}
\examples{
helloR("RStudio")
}
