% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_profile_plot.R
\name{highlight_profile_plot}
\alias{highlight_profile_plot}
\title{Highlight gene and/or taxon of interest on the profile plot}
\usage{
highlight_profile_plot(data, plot_parameter, taxon_highlight, 
    rank_name, gene_highlight)
}
\arguments{
\item{data}{data for heatmap plot}

\item{plot_parameter}{plot parameters (type of x-axis "taxa" or "genes";
names of 2 variables; colors for lowest and highest value of variable 1; 
colors for lowest and highest value of variable 2; color of co-orthologs; 
text sizes for x, y axis and legend; legend position "top", "bottom", 
"right", "left" or "none"; zoom ratio of the co-ortholog dots from -1 to 3;
angle of x-axis from 0 to 90; 
show/hide separate line for reference taxon 1/0; 
enable/disable coloring gene categories TRUE/FALSE)}

\item{taxon_highlight}{taxon of interst}

\item{rank_name}{working taxonomy rank}

\item{gene_highlight}{gene of interest}
}
\value{
A profile heatmap plot with highlighted gene and/or taxon of interest
as ggplot object.
}
\description{
Highlight gene and/or taxon of interest on the profile plot
}
\examples{
data("full_processed_profile", package="phyloprofile")
plot_df <- data_main_plot(full_processed_profile)
plot_parameter <- list(
    "x_axis" = "taxa",
    "var1_id" = "FAS",
    "var2_id"  = "Traceability",
    "low_color_var1" =  "#FF8C00",
    "high_color_var1" = "#4682B4",
    "low_color_var2" = "#FFFFFF",
    "high_color_var2" = "#F0E68C",
    "para_color" = "#07D000",
    "x_size" = 8,
    "y_size" = 8,
    "legend_size" = 8,
    "main_legend" = "top",
    "dot_zoom" = 0,
    "x_angle" = 60,
    "guideline" = 0,
    "color_by_group" = FALSE
)
taxon_highlight <- "Mammalia"
rank_name <- "class"
gene_highlight <- "OG_1019"
highlight_profile_plot(
    plot_df, plot_parameter, taxon_highlight, rank_name, gene_highlight
)
}
\seealso{
\code{\link{data_main_plot}}, \code{\link{data_customized_plot}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
