\name{hole}

\Rdversion{1.1}

\alias{hole}
\alias{hole<-}


\title{
hole
}

\description{
Is a piece a hole? Or set it to be a hole or not
}

\usage{
hole(x, i, j)
hole(x, i, j) <- value

}

\arguments{
  \item{x}{GeoVector object}
  \item{i}{Integer (part) }
  \item{j}{Integer (hole) }
  \item{value}{Logical }
}

\value{
GeoVector object
}

\author{
Robert J. Hijmans
}

\keyword{ spatial  }
