% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{IntegrationScores}
\alias{IntegrationScores}
\title{Retrieve the tibble containing integration scores from a Seurat object}
\usage{
IntegrationScores(object, scaled = FALSE)
}
\arguments{
\item{object}{a Seurat object}

\item{scaled}{whether to query the unscaled (default) or scaled scores (if
scaling has been performed)}
}
\value{
either \code{NULL} if the requested object does not exist, otherwise
a tibble with the first column containing the name of the integrations, and
each other column corresponding to one score.
}
\description{
Get the scaled or unscaled scores as a tibble
}
