% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irisaverages.R
\name{irisaverages}
\alias{irisaverages}
\title{dplyr example}
\usage{
irisaverages(
  col = c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width")
)
}
\arguments{
\item{col}{which column to average}
}
\description{
This adds a new function that needs {dplyr}
}
