% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_S4_methods.R
\name{isValid}
\alias{isValid}
\title{Check the validity of conda environment's components}
\usage{
isValid(x, ...)
}
\arguments{
\item{x}{a \linkS4class{CondaEnvManager}, a
\linkS4class{CondaEnv} object or a sub element of it of class
\code{CondaEnvSlot}.}

\item{...}{ignored when \code{x} is a \code{CondaEnvSlot} object. Otherwise,
can be used to pass \code{do.check} argument.}

\item{do.check}{whether to call \link[=checkCondaEnv]{check functions}. Does
not apply when \code{x} is a \code{CondaEnvSlot} object.}

\item{verbose}{ignored}
}
\value{
\code{TRUE} or \code{FALSE} or a logical named vector when \code{x}
is a \linkS4class{CondaEnvManager} object (one element per
method)
}
\description{
Get the validity of conda environment's components. Best used on a
\linkS4class{CondaEnvManager} or a
\linkS4class{CondaEnv} object.
}
\seealso{
\linkS4class{CondaEnvManager}, \linkS4class{CondaEnv},
\link[=checkCondaEnv]{check conda functions}
}
