% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/imputation.R
\name{knnImpute}
\alias{knnImpute}
\title{NA Imputation for data.frames via knn}
\usage{
knnImpute(x)
}
\arguments{
\item{x}{A data.frame with missing values to impute}
}
\description{
This function takes a data.frame and replaces NA values with the column predictions for a knn model
}

