% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linReg.R
\name{linReg}
\alias{linReg}
\title{Linear regression}
\usage{
linReg(data, save.result = NULL)
}
\arguments{
\item{data}{Data frame containing Y and X variables.}

\item{save.result}{Logical parameter indicating if results will be saved.}
}
\value{
Fit a linear regression.
}
\description{
This function fit a linear regression.
}
