% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{make filename by giving the year as parameter only.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year of the accidents recorded}
}
\value{
This function returns a file name that corresponds to the \code{year}
inputted as parameter
}
\description{
Because this is used by other functions of this package,
users don't have to see the result of this function, thus it is not exported.
}
