% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialAnno.R
\name{marker_list_to_mat}
\alias{marker_list_to_mat}
\title{marker_list_to_mat.}
\usage{
marker_list_to_mat(marker_list, include_other = TRUE)
}
\arguments{
\item{marker_list}{a marker list.}

\item{include_other}{is logical value specifying whether or not include unknown}
}
\value{
a list contains.
}
\description{
marker_list_to_mat, a function that transform list to cell type gene matrix.
}
\details{
marker_list_to_mat, a function that transform list to cell type gene matrix.
}
