% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matrix-indexing}
\alias{matrix-indexing}
\alias{rowcol2idx}
\alias{idx2col}
\alias{idx2row}
\title{Matrix indexing}
\usage{
rowcol2idx(rows.idx, cols.idx, height)

idx2col(idx, height)

idx2row(idx, height)
}
\arguments{
\item{rows.idx}{indices of rows}

\item{cols.idx}{indices of columns}

\item{height}{height of the matrix (number of rows)}

\item{idx}{vector-like, matrix-wide indices}
}
\value{
a vector of indices
}
\description{
Switch between rows & columns indices and vector-like indices of matrices.\cr
\cr
\code{rowcol2idx} get indices out of rows and columns\cr
\code{idx2col} get columns out of indices\cr
\code{idx2row} get rows out of indices
}
