% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matrix-sorting}
\alias{matrix-sorting}
\alias{rowSort}
\alias{colSort}
\alias{rowSorted}
\alias{colSorted}
\title{Matrix sorting}
\usage{
rowSort(mat, by = NULL, ncol = NULL, decreasing = FALSE)

colSort(mat, by = NULL, ncol = NULL, decreasing = FALSE)

rowSorted(mat, decreasing = FALSE)

colSorted(mat, decreasing = FALSE)
}
\arguments{
\item{mat}{matrix to sort or to check}

\item{by}{matrix to sort by. If \code{NULL}, sort \code{mat} by itself (default)}

\item{ncol}{number of desired columns for the sorted \code{mat}. If
\code{NULL}, keep original dimensions of \code{mat} (default)}

\item{decreasing}{logical.  Should the sort be increasing or decreasing?
    Not available for partial sorting.}
}
\value{
a sorted matrix of \code{ncol} columns
}
\description{
Sort a matrix in a row- or column-wise manner, or check that it is\cr
\cr
\code{rowSort} sort each row independently\cr
\code{colSort} sort each column independently\cr\cr
\code{rowSorted} check that each row is independently sorted\cr
\code{colSorted} check that each column is independently sorted
}
\note{
If \code{mat} and \code{by} have different dimensions, results might be
incorrect
}
