% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDP.RKHS.R
\name{MDP.RKHS}
\alias{MDP.RKHS}
\title{Releasing four Differential Private RKHS smoothing means of a dataset}
\usage{
MDP.RKHS(grid = "FALSE", Data, alpha = rep(1, 1), beta = rep(0.1, 1),
  kernel = c("Gau"), phi = rep(0.01, 1), ro = rep(0.2, 1),
  col.drop = TRUE)
}
\arguments{
\item{grid}{grid (x-axis) for each curve, default is equally espaced between 0 and 1.}

\item{Data}{a matrix which the of interest curves are located in columns}

\item{alpha, beta}{vector of privacy parameters}

\item{phi}{vector of penalty parameters}

\item{ro}{vector of kernel range parameters}

\item{drop.col}{TRUE/FALSE for cleaning the Data, deleting Columns/Rows including missing values}
}
\value{
four \link{DP.RKHS} output each of them includin "f.tilda","delta","f" and "X"
}
\description{
This function create four DP RKHS smoothing means from an existing data set
based on any arbitrary 4 by 1 vectors for \code{alpha}, \code{beta}, \code{kernel},
\code{phi} and \code{ro}.
}

