% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianCI.R
\name{medianCI}
\alias{medianCI}
\title{Confidence interval for the median in the one sample and in the paired two sample}
\usage{
medianCI(x, y = NULL, paired = FALSE, conf.level = 0.95)
}
\arguments{
\item{x, y}{two continuous variables.}

\item{paired}{a logical value. If it equals TRUE, you must provide values for \code{x} and \code{y}
and the paired test is implemented. If it equals FALSE, only \code{x} must be provided.}

\item{conf.level}{confidence level for the confidence interval.}
}
\value{
Returns the confidence interval along with the estimate of the median.
}
\description{
In the one sample case, compute the confidence interval for the median of the random variable. In the paired two sample case, compute the confidence interval for the median of the
difference between the two random variables.
}
\details{
Provide the confidence interval for \code{Med(X)} in the one sample case and for \code{Med(X-Y)} in the two sample case.
The confidence interval is based on the rank statistic.
}
\note{
The paired median confidence interval can be implemented by providing the variables \code{x} and \code{y} or by just providing
one vector equal to the difference between \code{x} and \code{y}.
}
\examples{
#Simulations
n=100
M=2000 #number of replications
res1=res2=res3=rep(NA,M)
testone=function(n){
D=rchisq(n,df=4)-qchisq(df=4, p=0.5)
result=medianCI(D)
list(test1=mediantest(D)$p.value,test2=c(result$CI))
}
for (i in 1:M)
{
result=testone(n)
res1[i]=result$test1
res2[i]=result$test2[1]
res3[i]=result$test2[2]
}
mean(res1<0.05) #0.049
(sum(res2>0)+sum(res3<0))/M #0.0525
}
\seealso{
mediantest
}
\keyword{test}
