% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediantest.R
\name{mediantest}
\alias{mediantest}
\title{Test for the median in the one and two sample paired tests}
\usage{
mediantest(x, y = NULL, alternative = "two.sided", paired = FALSE)
}
\arguments{
\item{x, y}{two continuous variables.}

\item{alternative}{indicates the alternative hypothesis and must be one of "two.sided", "greater" or "less".}

\item{paired}{a logical value. If it equals TRUE, you must provide values for \code{x} and \code{y}
and the paired test is implemented. If it equals FALSE, only \code{x} must be provided.}
}
\value{
Returns the result of the test with its corresponding p-value and the value of the test statistic.
}
\description{
In the one sample case, test if the median of the random variable is equal to 0. In the paired two sample case, test if the median of the
difference between the two random variables is equal to 0.
}
\details{
The null hypothesis for the one sample median test is: H0 Med(X)=0 where Med represents the median.
The alternative is specified by the \code{alternative} argument: "\code{greater}" means that Med(X)>0 and "\code{less}"
 means that Med(X)<0. The null hypothesis for the paired median test is: H0 Med(X-Y)=0. Both tests are asymptotically
 calibrated in the sense that the rejection probability under the null hypothesis is asymptotically equal to the level of the test. The
 test is based on the asymptotic law of the empirical median and uses a kernel estimator to estimate the density of \code{X} (in the one sample case)
 or of \code{X-Y} in the two sample case at 0.
}
\note{
The paired median test can be implemented by providing the variables \code{x} and \code{y} or by just providing
one vector equal to the difference between \code{x} and \code{y}.
}
\examples{
#Simulations
n=100
M=1000 #number of replications
res1=res2=rep(NA,M)
testone=function(n){
D=rchisq(n,df=4)-qchisq(df=4, p=0.5)
list(test1=mediantest(D)$p.value,test2=binom.test(sum(D>0),n)$p.value)
} #test2 is the sign test.
for (i in 1:M)
{
result=testone(n)
res1[i]=result$test1
res2[i]=result$test2
}
mean(res1<0.05) #0.048
mean(res2<0.05) # 0.04
}
\seealso{
\code{\link{cortest}}, \code{\link{robustest}}, \code{\link{vartest}}, \code{\link{wilcoxtest}}.
}
\keyword{test}
