% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{minus}
\alias{minus}
\title{subtract two vectors.}
\usage{
minus(x, y)
}
\arguments{
\item{x,y}{Numeric vectors}
}
\value{
difference of x and y
}
\description{
subtract two vectors or numbers
}
\examples{
minus(1,1)
minus(c(2,3),c(3,4))
minus(c(5,7),c(1,2))
}

