% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNN.R
\name{MNNIntegration}
\alias{MNNIntegration}
\title{Run classical MNN on Seurat's \link[SeuratObject]{Assay5} object through \code{\link[Seurat]{IntegrateLayers}}}
\usage{
MNNIntegration(
  object,
  orig = NULL,
  groups = NULL,
  layers = NULL,
  scale.layer = NULL,
  features = 2000,
  reconstructed.assay = "mnn.reconstructed",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{A \code{\link[SeuratObject]{Seurat}} object
(or an \code{\link[SeuratObject]{Assay5}} object if
not called by \code{\link[Seurat]{IntegrateLayers}})}

\item{orig}{\code{\link[SeuratObject]{DimReduc}} object. Not to be set
directly when called with \code{\link[Seurat]{IntegrateLayers}}, use
\code{orig.reduction} argument instead}

\item{groups}{Ignored}

\item{layers}{Name of the layers to use in the integration}

\item{scale.layer}{Name of the scaled layer in \code{Assay}}

\item{features}{Either a list of features to use when calculating batch
correction, or a number (2000 by default) of variable features to select.}

\item{reconstructed.assay}{Name for the assay containing the low-rank
reconstruction of the expression matrix.}

\item{verbose}{Print messages. Set to \code{FALSE} to disable}

\item{...}{Extra parameters passed to \code{\link[batchelor]{mnnCorrect}}}
}
\value{
A Seurat object merged from the objects in \code{object.list} and a
new DimReduc of name \code{reduction.name} (key set to \code{reduction.key})
with corrected embeddings matrix as well as the rotation matrix used for the
PCA stored in the feature loadings slot. Also returns an expression matrix
reconstructed from the low-rank approximation in the
\code{reconstructed.assay} assay; all other metadata info
\code{\link[batchelor]{mnnCorrect}} is stored in the \code{tool} slot,
accessible with \code{\link[Seurat]{Tool}}
}
\description{
A wrapper to run \code{\link[batchelor]{mnnCorrect}} on multi-layered Seurat V5 object
}
\note{
This function requires the
\href{https://bioconductor.org/packages/release/bioc/html/batchelor.html}{\pkg{batchelor}} package
to be installed
}
\examples{
\dontrun{
# Preprocessing
obj <- UpdateSeuratObject(SeuratData::LoadData("pbmcsca"))
obj[["RNA"]] <- split(obj[["RNA"]], f = obj$Method)
obj <- NormalizeData(obj)
obj <- FindVariableFeatures(obj)
obj <- ScaleData(obj)
obj <- RunPCA(obj)

# After preprocessing, we integrate layers:
obj <- IntegrateLayers(object = obj, method = MNNIntegration,
  new.reduction = 'integrated.mnn', verbose = FALSE)

# We can also add parameters specific to mnnCorrect.
# Here we set `k` to specify the number of nearest neighbors
# to use when identifying MNNs:
obj <- IntegrateLayers(object = obj, method = MNNIntegration,
  new.reduction = 'integrated.mnn', k = 15, verbose = FALSE)
}

}
\seealso{
\code{\link[batchelor]{mnnCorrect}} \code{\link[Seurat]{Tool}}
}
