% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{mom_gamma}
\alias{mom_gamma}
\title{먼저 제일 위쪽에 자리잡은 줄은 설명서 제목으로 사용된다.}
\usage{
mom_gamma(sample_g)
}
\arguments{
\item{:}{함수에 들어가는 입력값(parameter)을 명시해주고, 이에 대한 설명을 써주면 된다.}
}
\value{
: 함수의 결과값에 대한 설명을 써준다.
}
\description{
제목에서 한줄을 뗀 다음 줄은 설명서에서 함수의 description란에 위치하게 된다.
}
