% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operaciones.R
\name{multiplica.dos.numeros}
\alias{multiplica.dos.numeros}
\title{Multiplica dos números}
\usage{
multiplica.dos.numeros(x, y)
}
\arguments{
\item{x}{Primer valor numérico a multiplicar}

\item{y}{Primer valor numérico a multiplicar}
}
\value{
El producto de las dos entradas
}
\description{
Recibe dos números y devuelve su producto
}
\examples{
multiplica.dos.numeros(1,0)
multiplica.dos.numeros(-5,3)
}

