% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{multiply}
\alias{multiply}
\title{multiply two vectors.}
\usage{
multiply(x, y)
}
\arguments{
\item{x,y}{Numeric vectors}
}
\value{
multiplication of x and y
}
\description{
multiple two vectors or numbers
}
\examples{
multiply(2,3)
multiply (c(1,3),c(3,4))
}

