% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MUVR.R
\name{MUVR}
\alias{MUVR}
\title{MUVR: Wrapper for "Multivariate modelling with Unbiased Variable selection"}
\usage{
MUVR(X, Y, ID, nRep = 5, nOuter = 6, nInner, varRatio = 0.75,
  DA = FALSE, fitness = c("AUROC", "MISS", "RMSEP"), method = c("PLS",
  "RF"), nCompMax, methParam, ML = FALSE, modReturn = FALSE,
  newdata = NULL, logg = FALSE)
}
\arguments{
\item{X}{Independent variables. NB: Variables (columns) must have names/unique identifiers. NAs not allowed in data. For multilevel, only the positive half of the difference matrix is specified.}

\item{Y}{Response vector (Dependent variable). For classification, a factor (or character) variable should be used. For multilevel, Y is calculated automatically.}

\item{ID}{Subject identifier (for sampling by subject; Assumption of independence if not specified)}

\item{nRep}{Number of repetitions of double CV. (Defaults to 5)}

\item{nOuter}{Number of outer CV loop segments. (Defaults to 6)}

\item{nInner}{Number of inner CV loop segments. (Defaults to nOuter-1)}

\item{varRatio}{Ratio of variables to include in subsequent inner loop iteration. (Defaults to 0.75)}

\item{DA}{Logical for Classification (discriminant analysis) (Defaults do FALSE, i.e. regression). PLS is limited to two-class problems (see `Y` above).}

\item{fitness}{Fitness function for model tuning (choose either 'AUROC' or 'MISS' for classification; or 'RMSEP' (default) for regression.)}

\item{method}{Multivariate method. Supports 'PLS' and 'RF' (default)}

\item{nCompMax}{Option to choose max number of PLS components (default is 5)}

\item{methParam}{List with parameter settings for specified MV method (see function code for details)}

\item{ML}{Logical for multilevel analysis (defaults to FALSE)}

\item{modReturn}{Logical for returning outer segment models (defaults to FALSE)}

\item{newdata}{New data matrix ONLY for prediction NOT modelling (Not yet fully implemented)}

\item{logg}{Logical for whether to sink model progressions to `log.txt`}
}
\value{
An object containing stuff...
}
\description{
Repeated double cross validation with tuning of variables in the inner loop.
}

