% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_function.R
\name{my_function}
\alias{my_function}
\title{Title of your function}
\usage{
my_function(x)
}
\arguments{
\item{x}{Description of parameter x.}
}
\value{
Description of the return value.
}
\description{
Description of your function.
}
