% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-code.R
\name{my_power_function}
\alias{my_power_function}
\title{Raise a vector of values to a power}
\usage{
my_power_function(values, exponent)
}
\arguments{
\item{values}{A vector.}

\item{exponent}{A number.}
}
\value{
A numeric vector.
}
\description{
Raise a vector of values to a power
}
\examples{
my_power_function(seq(1,10), 2)
my_power_function(seq(1,10), 3)
}
