% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyFirstFunction.R
\name{MyFirstFunction}
\alias{MyFirstFunction}
\title{Print Number and Sentence}
\usage{
MyFirstFunction(x)
}
\value{
x and my sentence
}
\description{
Print Number and Sentence
}
\examples{
MyFirstFunction(1)
}
