% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearZeroVar.R
\name{nearZeroVar}
\alias{nearZeroVar}
\title{Identify variables with nea zero variance}
\usage{
nearZeroVar(x, freqCut = 95/5, uniqueCut = 10)
}
\arguments{
\item{x}{}

\item{freqCut}{}

\item{uniqueCut}{}
}
\value{
nzv object
}
\description{
Adapted and stripped down from mixOmics v 5.2.0 (https://cran.r-project.org/web/packages/mixOmics/)
}
