% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{new_dist}
\alias{new_dist}
\title{Distribution Class}
\usage{
new_dist(
  name = NULL,
  type = c("continous", "discrete"),
  ddist = NULL,
  supp = NA,
  cdist = NULL,
  qdist = NULL,
  hdist = NULL,
  ...,
  class = character()
)
}
\arguments{
\item{name}{A name.}

\item{type}{"continous" or "discrete".}

\item{ddist}{density function.}

\item{cdist}{cdf.}

\item{hdist}{hazard function.}

\item{support}{support.}
}
\value{
A dist class.
}
\description{
Create a distribution class:

OOP.
}
\examples{
normal_density <- function(x, par1 = 0, par2 = 1) { dnorm(x, mean = par1, sd = par2) }
normal_cdf <- function(x, par1 = 0, par2 = 1) { pnorm(x, mean = par1, sd = par2) }
normal_hazard <- function(x, par1 = 0, par2 = 1) { normal_density(x, par1, par2) / ( 1 - normal_cdf(x, par1, par2) ) }
x <- new_dist(name = "my.normal",  type = "continous", ddist = normal_density, supp = c(-Inf,Inf), cdist = normal_cdf, hdist = normal_hazard)
print(x)

y <- new_dist(ddist=dexp, type="continous", name="exponential", supp=c(0,Inf))
plot(y,type="b",pch=16)

z <- new_dist()
plot(y+z,type="l",col=2)
save(x,y,z, file="data/dist.rda")
}
