% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewtRaphXi.R
\name{NewtRaphXi}
\alias{NewtRaphXi}
\title{Newton-Raphson for Xi}
\usage{
NewtRaphXi(xiseed, theta, Y, X, tol = 10^(-5))
}
\arguments{
\item{xiseed}{Initial value}

\item{theta}{theta}

\item{Y}{data}

\item{X}{data}

\item{tol}{approximation tolerance}
}
\value{
list
}
\description{
Function for Newton-Raphson method for xi
}
\details{
Okay
}
\note{
%% ~~further notes~~
}
\examples{

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (xiseed, theta, Y, X, tol = 10^(-5)) 
{
    rho = exp(X \%*\% theta)
    n = length(Y)
    kiter = 0
    OK = FALSE
    xi0 = xiseed
    while (!OK) {
        kiter = kiter + 1
        g = sum((Y - rho)^2 - rho * (1 + (1 + rho)/xi0))
        gp = sum(rho * (1 + rho))/xi0^2
        xi1 = xi0 - g/gp
        if (abs(xi1 - xi0) < tol) {
            OK = TRUE
        }
        xi0 = xi1
        if (xi1 < 0) {
            OK = FALSE
            xiseed = xiseed/2
            xi0 = xiseed
            kiter = 0
        }
    }
    return(list(kiter = kiter, xi = xi1))
  }

}
\references{
%% ~put references to the literature/web site here ~
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\author{
E. A. Pena
}
