% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma_fit.R
\name{nma_fit}
\alias{nma_fit}
\title{Takes input data and a model plan and passes to the model engine specified. 
Current supported engines are the `gemtc` package (using mtc.model & mtc.run) or `rjags` (using jags and dic.samples functions)}
\usage{
nma_fit(data, plan)
}
\arguments{
\item{data}{input data which has been pre-processed}

\item{plan}{list object containing all input parameters}
}
\value{
model object of class `rjags` or `mtc.result`
}
\description{
Takes input data and a model plan and passes to the model engine specified. 
Current supported engines are the `gemtc` package (using mtc.model & mtc.run) or `rjags` (using jags and dic.samples functions)
}
