% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma_pre_proc.R
\name{nma_jags_inits}
\alias{nma_jags_inits}
\title{Create list of initial JAGS parameters}
\usage{
nma_jags_inits(dat)
}
\arguments{
\item{dat}{}
}
\value{
\code{list} of 4 elements containing the JAGS inits (see details)
}
\description{
Create list of initial JAGS parameters
}
\details{
The output list contains
\itemize{
 \item{"dat"}{The original data with labels for `arm` and `n_arm`}
 \item{"d_arms"}{inits for arms}
 \item{"d_std"}{inits for std}
 \item{"d_trts"}{inits for treatments}
}
}
