% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{NormaliseL2}
\alias{NormaliseL2}
\title{Normalise a matrix using L2 norm}
\usage{
NormaliseL2(mat, MARGIN = 1)
}
\arguments{
\item{mat}{a matrix (sparse or dense)}

\item{MARGIN}{one of 1 or 2, corresponding to normalisation per rows and
columns respectively}
}
\value{
the matrix with normalised rows or columns
}
\description{
Normalise the rows or columns of a matrix using L2 norm
}
\note{
Adapted from \href{https://github.com/satijalab/seurat/blob/1549dcb3075eaeac01c925c4b4bb73c73450fc50/R/utilities.R#L1941-L1956}{Seurat:::L2Norm}
}
