\name{nparts}
\Rdversion{1.1}

\alias{nparts}
\alias{length,GeoVector-method}

\title{
Number of parts
}

\description{
How many parts does this object have?
}

\usage{
nparts(x)
}

\arguments{
\item{x}{GeoVector object }
}

\value{Integer}

\author{
Robert J. Hijmans
}


\keyword{ spatial  }
