\name{npieces}
\Rdversion{1.1}
\alias{npieces}

\title{
Number of pieces
}

\description{
Number of pieces
}

\usage{
npieces(x)
}

\arguments{
  \item{x}{geoVector object}
}


\value{Integer}


\author{
Robert J. Hijmans
}


\seealso{ \code{\link{nparts}} }

\keyword{ spatial  }
