% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_lisi_class.R
\name{as.numeric,numeric_lisi-method}
\alias{as.numeric,numeric_lisi-method}
\alias{[,numeric_lisi,ANY,ANY,ANY-method}
\alias{rep,numeric_lisi-method}
\alias{[<-,numeric_lisi,ANY,ANY,ANY-method}
\title{Convert a numeric vector to a `numeric_lisi` vector}
\usage{
\S4method{as.numeric}{numeric_lisi}(x)

\S4method{[}{numeric_lisi,ANY,ANY,ANY}(x, i)

\S4method{rep}{numeric_lisi}(x, ...)

\S4method{[}{numeric_lisi,ANY,ANY,ANY}(x, i) <- value
}
\arguments{
\item{x}{a numeric_lisi vector}

\item{i}{indices specifying elements to extract or replace}

\item{...}{passed to rep.numeric}

\item{value}{value to replace with}
}
\description{
Convert a numeric vector to a `numeric_lisi` vector

Indexing method for `numeric_lisi` (keep `@N` slot)

Replicate method for numeric_lisi (keep `@N` slot)

Value replacement method for `numeric_lisi` (check `@N` slot compatibility)
}
\keyword{internal}
