% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairplot.R
\name{pairplot}
\alias{pairplot}
\title{Pairplot}
\usage{
pairplot(
  df,
  vars = NULL,
  method = "pearson",
  diag_fun = "densityDiag",
  lower_fun = "points",
  ...
)
}
\arguments{
\item{df}{Dataframe to be used.}

\item{vars}{Variables used from the df to create the pairplot}

\item{method}{Type of correlation: Pearson, Spearman, Kendall, Bicor}

\item{diag_fun}{Type of plot used for the diagonal of the paiplot}

\item{lower_fun}{Type of plot used for the lower triangular part of the pairplot.}
}
\value{
A list with the plot (and its information) and the correlation DF.
}
\description{
A pairplot is displayed for with the attributes defined by the user.
}
\examples{
result <- pairplot(iris, method = "kendall", diag_fun = "densityDiag", lower_fun = "points")
head(result$correlation_df)
result$plot$data
}
