% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_domain_input.R
\name{parse_domain_input}
\alias{parse_domain_input}
\title{Parse domain input file}
\usage{
parse_domain_input(seed, input_file, type)
}
\arguments{
\item{seed}{seed ID(s)}

\item{input_file}{name of input data (demo data, file name or path to folder)}

\item{type}{type of data (demo, file or folder)}
}
\value{
A dataframe containing protein domain info
}
\description{
Parse domain input file
}
\examples{
seed <- "OG_1009"
input_file <- system.file(
    "extdata", "domain_files/OG_1009.domains",
    package = "phyloprofile", mustWork = TRUE
)
type <- "file"
parse_domain_input(seed, input_file, type)
}
\seealso{
\code{\link{get_domain_online}}, \code{\link{get_domain_file}},
\code{\link{get_domain_folder}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
