% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_phylo_profile.R
\name{parse_info_profile}
\alias{parse_info_profile}
\title{Parsing info for phylogenetic profiles}
\usage{
parse_info_profile( input_df, sorted_input_taxa, var1_aggregate_by, 
    var2_aggregate_by)
}
\arguments{
\item{input_df}{input profiles in long format}

\item{sorted_input_taxa}{sorted taxonomy data for the input taxa
(check sort_input_taxa())}

\item{var1_aggregate_by}{aggregate method for VAR1 (min, max, mean or median)}

\item{var2_aggregate_by}{aggregate method for VAR2 (min, max, mean or median)}
}
\value{
Dataframe contains all info for input profiles (a full processed
profile)
}
\description{
Creating main dataframe for the input phylogenetic profiles with
the selected input taxonomy level (e.g. strain, species) and reference taxon.
The output contains the number of paralogs, percentage of species presence
in each supertaxon, and the max/min/mean/median of VAR1 and VAR2.
}
\examples{
data("main_long_raw", package="phyloprofile")
input_df <- main_long_raw
taxon_IDs <- get_input_taxa_id(input_df)
taxon_names <- get_input_taxa_name("class", taxon_IDs)
sorted_input_taxa <- sort_input_taxa(
    taxon_IDs, taxon_names, "class", "Mammalia", NULL
)
var1_aggregate_by <- "max"
var2_aggregate_by <- "mean"
parse_info_profile(
    input_df, sorted_input_taxa, var1_aggregate_by, var2_aggregate_by
)
}
\seealso{
\code{\link{create_long_matrix}}, \code{\link{sort_input_taxa}},
\code{\link{calc_pres_spec}}, \code{\link{main_long_raw}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
