\name{part}
\Rdversion{1.1}

\alias{part}
\alias{part,GeoVector,numeric-method}
\alias{part,GeoPolygons,numeric-method}

\title{
Get a part
}

\description{
Get a part
}

\usage{
part(x, i, ...)
}

\arguments{
\item{x}{GeoVector object }
\item{i}{ integer }
\item{\dots}{Additional arguments (none implemented) } 
}


\value{
GeoVector object
}

\author{
Robert J. Hijmans
}

\keyword{ spatial  }
