% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartialSum.R
\name{PartialSum}
\alias{PartialSum}
\title{Partial Sum}
\usage{
PartialSum(x)
}
\arguments{
\item{x}{data}
}
\value{
list
}
\description{
Function for data partial summation
}
\details{
Okay
}
\note{
%% ~~further notes~~
}
\examples{

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (x) 
{
    nlen = length(x)
    y = rep(0, nlen)
    for (i in 1:nlen) {
        y[i] = sum(x[1:i])
    }
    return(y)
  }

}
\references{
%% ~put references to the literature/web site here ~
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\author{
E. A. Pena
}
