% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{Pear_corr}
\alias{Pear_corr}
\title{Pearson correlation coefficient}
\usage{
Pear_corr(x, y)
}
\arguments{
\item{x}{is vector of variables}

\item{y}{is vector of variables}
}
\value{
a numerical value of length 1
}
\description{
Pearson correlation coefficient
}
\examples{
Pear_corr(x,y)
}
