% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_cor.R
\name{phi_cor}
\alias{phi_cor}
\title{Phi Correlation}
\usage{
phi_cor(data, asig = list())
}
\arguments{
\item{data}{Dataset to which the variables belong.}

\item{asig}{A list that can be defined by the user if they wish to specify a value for 0 and 1.}
}
\value{
Heatmap where the correlation between the categorical variables is shown.
}
\description{
Calculates the phi correlation between dichotomous variables and displays the results
with a heatmap.
}
\examples{
data <- data.frame(
smoke = c("Smokes", "Smokes", "Smokes", "Does not smoke", "Smokes"),
diabetes = c("No", "No", "Yes", "Yes", "No"),
hypertension = c("Yes", "No", "Yes", "No", "Yes")
)
asig <- list(smoke = c("Does not smoke", "Smokes"), diabetes = c("No", "Yes"))
phi_cor(data, asig)
}
