% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_leibniz.R
\name{pi.Leibniz}
\alias{pi.Leibniz}
\title{pi.Leibniz}
\usage{
pi.Leibniz(y)
}
\arguments{
\item{y}{integer}
}
\value{
the first y terms of the Leibniz formula for pi
}
\description{
pi.Leibniz
}
\examples{
pi.Leibniz(10000)
}

