\name{piece}
\Rdversion{1.1}

\alias{piece}
\alias{piece,GeoVector,numeric,numeric-method}
\alias{piece,GeoPolygons,numeric,numeric-method}


\title{
Piece
}
\description{
Get a piece
}

\usage{
piece(x, i, j, ...)
}

\arguments{
  \item{x}{GeoVector object}
  \item{i}{Integer (part) }
  \item{j}{Integer (piece) }
  \item{\dots}{Additional arguments (none implemented) } }



\value{
GeoVector object
}

\author{
Robert J. Hijmans
}

\keyword{ spatial  }
