% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boxplots.R
\name{plot_boxplots}
\alias{plot_boxplots}
\title{Boxplots}
\usage{
plot_boxplots(df, vars = NULL, ncols = 2)
}
\arguments{
\item{df}{Dataframe to be used.}

\item{vars}{Variables used from the df to create the boxplots}

\item{ncols}{Number of columns in which the boxplots are going to be displayed}
}
\description{
Boxplots are plotted for the given variables in a DF
}
\examples{
plot_boxplots(iris)
plot_boxplots(iris, "Sepal.Length")
plot_boxplots(iris, c("Petal.Length", "Petal.Width", "Sepal.Length"), ncols = 2)
}
