% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{plot_specificity}
\alias{plot_specificity}
\title{Plot specificity}
\usage{
plot_specificity(lr_result)
}
\arguments{
\item{lr_result}{A \code{list} value of class 'logistic_regression', found from running the 'get_beta_estimate' function in this package. The list contains the following objects:
\describe{
 \item{initial_beta}{The initial values of coefficients.}
 \item{beta_estimate}{The estimated coefficients of the logistic regression.}
 \item{response}{The \code{double} vector containing the response used for the estimation.}
 \item{predictors}{The \eqn{n \times p} \code{double} value of the matrix containing the values of the predictors used for the estimation.}
}}
}
\value{
A \code{list} value containing the following objects:
\describe{
 \item{df}{The data frame containing cut-off values and the corresponding specificity values.}
 \item{plot}{The ggplot object created for the specificity bar graph.}
}
}
\description{
This function plots the specificity for the found coefficients of logistic regression over different cut off values from 0.1 to 0.9.
}
\author{
Saksham Goel
}
