% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDP.R
\name{plotDP}
\alias{plotDP}
\title{Plotting RKHS smoothing mean and its DP version simultaneously}
\usage{
plotDP(DP, text.main = "", legend.cex = 1.4, legend.loc = "topleft",
  seg.lin = 0.75, text.font = 2, text.width = 0.15, xlab = "",
  ylab = "", extra_range = c(0, 0), lty.p = 3, lwd.p = 3,
  legend.text = c("sample functions", "RKHS mean", "DP RKHS mean"),
  col = c("grey", "green", "red"))
}
\arguments{
\item{DP}{output of \link{DP.RKHS} or \link{DP.Sim} list}
}
\description{
This function plots the original data, the RKHS smoothing mean and its differential private
version simply in a same plot
}

