% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top5.R
\name{plotfirst2}
\alias{plotfirst2}
\title{<U+FFFD>å…©<U+FFFD>‹æ<U+FFFD>ä½ç<U+FFFD><U+FFFD>«å<U+FFFD>}
\usage{
plotfirst2(mtx)
}
\arguments{
\item{mtx}{}
}
\description{
<U+FFFD>å…©<U+FFFD>‹æ<U+FFFD>ä½ç<U+FFFD><U+FFFD>«å<U+FFFD>
}

