% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMDP.R
\name{plotMDP}
\alias{plotMDP}
\title{Plotting 4 RKHS smoothing mean and its DP version simultaneously}
\usage{
plotMDP(MDP, text.main = "Default", legend.cex = "Default",
  legend.loc = "Default", seg.lin = 0.75, text.font = 2,
  text.width = 0.15, legend.size = 0.57, xlab = "", ylab = "",
  extra_range = "Default", lty.p = 3, lwd.p = 3, par = "Default")
}
\arguments{
\item{DP}{output of \link{DP.RKHS} or \link{DP.Sim} list}
}
\description{
This function plots 4 different figures which the original data, the RKHS smoothing mean and its differential private
version are simply including in each of them
}

