% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPred.R
\name{plotPred}
\alias{plotPred}
\title{Plot predictions}
\usage{
plotPred(Ytrue, Ypreds)
}
\arguments{
\item{Ytrue}{}

\item{Ypreds}{}
}
\description{
At present, this function only supports predictions for PLS regression type problems
}

