% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda.R
\name{plsda}
\alias{plsda}
\title{PLS-DA}
\usage{
plsda(X, Y, ncomp = 2, max.iter = 500, tol = 1e-06,
  near.zero.var = TRUE)
}
\arguments{
\item{X}{}

\item{Y}{}

\item{ncomp}{}

\item{max.iter}{}

\item{tol}{}

\item{near.zero.var}{}
}
\value{
plsda object
}
\description{
Adapted and stripped down from mixOmics v 5.2.0 (https://cran.r-project.org/web/packages/mixOmics/)
}
