% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{plus}
\alias{plus}
\title{add two vectors.}
\usage{
plus(x, y)
}
\arguments{
\item{x,y}{Numeric vectors}
}
\value{
sum of x and y
}
\description{
add two vectors or numbers
}
\examples{
plus(1,1)
plus(c(2,3),c(3,4))
}

