% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/plus.R
\name{plusFunc}
\alias{plusFunc}
\title{A Plus Function}
\usage{
plusFunc(x = 0, y = 0)
}
\arguments{
\item{x}{First Value. Defaults to 0.}

\item{y}{Second Value. Defaults to 0.}
}
\description{
This function allows you to add two numbers.
}
\examples{
plusFunc(4,5) gives you 9
}
\keyword{plus}

