\name{pointInPolygon}
\Rdversion{1.1}
\alias{pointInPolygon}

\title{point in polygon }

\description{
point in polygon
}

\usage{
pointInPolygon(points, polygons, fun = NULL)
}

\arguments{
 \item{points}{points }
  \item{polygons}{polygons}
  \item{fun}{function}
}

\value{
matrix
}

\author{
Robert J. Hijmans
}

\keyword{ spatial  }
