% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_cor.R
\name{poly_cor}
\alias{poly_cor}
\title{Polychoric Correlation}
\usage{
poly_cor(data)
}
\arguments{
\item{data}{Dataset to which the variables belong.}
}
\value{
Heatmap where the correlation between the categorical variables is shown.
}
\description{
Calculates the polychoric correlation between ordinal variables and displays the results
with a heatmap.
}
\examples{
data <- data.frame(
  taste = sample(1:100, 30, replace = TRUE),  # Scale from 1 to 5
  aroma = sample(1:100, 30, replace = TRUE),
  color = sample(1:100, 30, replace = TRUE),
  texture = sample(1:100, 30, replace = TRUE)
 )
example_data <- data.frame(lapply(example_data, as.factor))
poly_cor(example_data)
}
