% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pp_tree}
\alias{pp_tree}
\title{An example of a taxonomy tree in newick format.}
\format{A data frame with only one entry
\describe{
    \item{V1}{ tree in newick format}
}}
\usage{
pp_tree
}
\description{
An example of a taxonomy tree in newick format.
}
\keyword{datasets}
