% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.plsdaMUVR}
\alias{predict.plsdaMUVR}
\title{Predict plsda}
\usage{
\method{predict}{plsdaMUVR}(object, newdata, method = c("all", "max.dist",
  "centroids.dist", "mahalanobis.dist"), ...)
}
\arguments{
\item{object}{}

\item{newdata}{}

\item{method}{}

\item{...}{}
}
\value{
plsda predictions
}
\description{
Adapted and stripped down from mixOmics v 5.2.0 (https://cran.r-project.org/web/packages/mixOmics/)
}
