% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediantest.R
\name{print.mediantest}
\alias{print.mediantest}
\title{#' @export
mediantest.formula=function(formula,data=list(),...)
{
  mf <- model.frame(formula=formula, data=data)
  response <- attr(attr(mf, "terms"), "response")
  Fact<-factor(mf[[-response]])
  DATA <- setNames(split(mf[[response]], Fact), c("x", "y"))
  result <- wilcoxtest.default(DATA[[1]],DATA[[2]],...)
  return(result)
}}
\usage{
\method{print}{mediantest}(x, ...)
}
\description{
#' @export
mediantest.formula=function(formula,data=list(),...)
{
  mf <- model.frame(formula=formula, data=data)
  response <- attr(attr(mf, "terms"), "response")
  Fact<-factor(mf[[-response]])
  DATA <- setNames(split(mf[[response]], Fact), c("x", "y"))
  result <- wilcoxtest.default(DATA[[1]],DATA[[2]],...)
  return(result)
}
}
