% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Printing.R
\name{print.summary}
\alias{print.summary}
\title{Print summary of the vcf file}
\usage{
\method{print}{summary}(vcf, type = "both")
}
\arguments{
\item{x}{tidyvcfR object}
}
\value{
Outputs to console.
}
\description{
This function prints descriptive summary from a vcfR that is created
with vcfR2tidy() function from vcfR package
}
\details{
The function takes the vcfR file and type of the summary as an input.
Possible argumets for an input file are "variant", "structural", "both" and
"gt", argument "both" is a default value. Options variant and structural
print how many different (structural) variants are possible based on
the fixed file data from the tidyvcfR object. The "gt" option calls for
another function count.variants() to calculate how many times each variant type
occurs in the genotypes among all individuals in the dataset. This counting
function returns a list of the counts, which is then used to print the summary
numbers of the variants.
}
\section{Warning}{

Function requires tidyvcfR object as an input
}

