% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{print_summary}
\alias{print_summary}
\title{Print Summary}
\usage{
print_summary(lr_result)
}
\arguments{
\item{lr_result}{A \code{list} value of class 'logistic_regression', found from running the 'get_beta_estimate' function in this package. The list contains the following objects:
\describe{
 \item{initial_beta}{The initial values of coefficients.}
 \item{beta_estimate}{The estimated coefficients of the logistic regression.}
 \item{response}{The \code{double} vector containing the response used for the estimation.}
 \item{predictors}{The \eqn{n \times p} \code{double} value of the matrix containing the values of the predictors used for the estimation.}
}}
}
\description{
This function prints summary of the 'logistic_regression' class object created after running the logistic regression.
}
\author{
Saksham Goel
}
