% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDarch.R
\name{print_weight}
\alias{print_weight}
\title{Print out a DArch object's weight}
\usage{
print_weight(darch, layer)
}
\arguments{
\item{darch}{an DArch instance}

\item{layer}{the number of the layer to print}
}
\description{
Print function for DArch objects. The selected layer's weight is printed
in a heatmap
}

