% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qMUVR.R
\name{qMUVR}
\alias{qMUVR}
\title{qMUVR: Wrapper for speedy access to MUVR (autosetup of parallelization)}
\usage{
qMUVR(X, Y, ML = F, method = "RF", varRatio = 0.65, nCore,
  repMult = 1, nOuter = 5, ...)
}
\arguments{
\item{X}{X-data}

\item{Y}{Y-data}

\item{ML}{Boolean for multilevel}

\item{method}{'RF' (default) or 'PLS'}

\item{varRatio}{proportion of variables to keep in each loop of the recursive feature elimination}

\item{nCore}{Number of threads to use for calculation (defaults to detectCores()-1)}

\item{repMult}{Multiplier of cores -> nRep = repMult * nCore}

\item{nOuter}{Number of outer segments}

\item{...}{}
}
\value{
MUVR object
}
\description{
qMUVR: Wrapper for speedy access to MUVR (autosetup of parallelization)
}
