% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_phylo_profile.R
\name{reduce_profile}
\alias{reduce_profile}
\title{Reduce the full processed profile into supertaxon level}
\usage{
reduce_profile(full_profile)
}
\arguments{
\item{full_profile}{dataframe contains the full processed profiles}
}
\value{
A reduced dataframe contains only profiles for the selected rank.
This dataframe contains only supertaxa and their value (%present, mVar1 &
mVar2) for each gene.
}
\description{
Reduce data of the phylogenetic profiles from input taxonomy
rank into supertaxon level (e.g. from species to phylum)
}
\examples{
data("full_processed_profile", package="phyloprofile")
reduce_profile(full_processed_profile)
}
\seealso{
\code{\link{parse_info_profile}} for creating a full processed
profile dataframe, \code{\link{full_processed_profile}} for a demo full
processed profile dataframe
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
