% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_S4_helpers.R
\name{reloadCache}
\alias{reloadCache}
\title{Reload cache from disk}
\usage{
reloadCache()
}
\value{
a \linkS4class{CondaEnvManager} object (invisibly)
}
\description{
Reload cache from disk to update current list of conda environments in
package environment
}
