% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfInner.R
\name{rfInner}
\alias{rfInner}
\title{RF model in inner CV loop}
\usage{
rfInner(xTrain, yTrain, xVal, yVal, fitness)
}
\arguments{
\item{xTrain}{Training data (samples as rows; variables as columns)}

\item{yTrain}{Training response}

\item{xVal}{Validation data}

\item{yVal}{Validation response (for tuning)}

\item{fitness}{Fitness function ('misClass', 'AUROC' or 'RMSEP')}
}
\value{
An object containing:

(`miss`, `auc` or `rmsep`) A fitness metric

`vip` VIP rankings
}
\description{
Called from Wrapper
}

