% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfPred.R
\name{rfPred}
\alias{rfPred}
\title{RF predictions for outer segments and consensus models}
\usage{
rfPred(
  xTrain,
  yTrain,
  xTest,
  yTest,
  ntree = 500,
  DA,
  keep.forest = FALSE,
  method
)
}
\arguments{
\item{xTrain}{}

\item{yTrain}{}

\item{xTest}{}

\item{yTest}{}

\item{ntree}{}

\item{DA}{}

\item{keep.forest}{}

\item{method}{}
}
\value{

}
\description{
RF predictions for outer segments and consensus models
}
