% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_phyloprofile.R
\name{run_phyloprofile}
\alias{run_phyloprofile}
\title{Run PhyloProfile app}
\usage{
run_phyloprofile()
}
\value{
shiny application of PhyloProfile
}
\description{
Run PhyloProfile app
}
\examples{
?phyloprofile::run_phyloprofile
\dontrun{
run_phyloprofile()
}
}
