% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGeneEx.R
\name{runGeneEx}
\alias{runGeneEx}
\title{Run GeneEx - webapp for sRACIPE}
\usage{
runGeneEx()
}
\value{
Webapp interface
}
\description{
GeneEx enables interactive simulations, visualization, and data
analysis using sRACIPE.
}
