% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testR.R
\name{rw.Metropolis}
\alias{rw.Metropolis}
\title{a random walk Metropolis sampler using R}
\usage{
rw.Metropolis(sigma, x0, N)
}
\arguments{
\item{sigma}{the variances are used for the normal distribution}

\item{x0}{Initial value of x}

\item{N}{the number of samples}
}
\value{
a random sample of size N
}
\description{
a random walk Metropolis sampler using R
}
\examples{
\dontrun{
N=2000
sigma=c(.05,.5,2,16)
x0=25
rw1=rw.Metropolis(sigma[1],x0,N)
rw2=rw.Metropolis(sigma[2],x0,N)
rw3=rw.Metropolis(sigma[3],x0,N)
rw4=rw.Metropolis(sigma[4],x0,N)
}
}
