% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_S4_methods.R
\name{saveToCache}
\alias{saveToCache}
\title{Save a conda environment manager to cache}
\usage{
saveToCache(x, ..., verbose = getOption("verbose"))
}
\arguments{
\item{x}{a \linkS4class{CondaEnvManager} object}

\item{...}{ignored}

\item{verbose}{ignored}
}
\description{
Save a conda environment manager to cache (
\code{\link[tools:R_user_dir]{R_user_dir("SeuratIntegrate", which = "config")}})
as a data.frame in a compressed RDS file.
}
\seealso{
\linkS4class{CondaEnvManager}
}
