% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGRN_getNt.R
\name{scGRN_getNt}
\alias{scGRN_getNt}
\title{scGRN_getNt}
\usage{
scGRN_getNt(df, gexpr, df_gene_id = 'hgnc_symbol', gexpr_gene_id = 'hgnc_symbol',
           cutoff_by = 'quantile', cutoff_percentage = 0.9,
           cutoff_absolute = 0.1,scaleby = 'no',
           train_ratio = 0.7, num_cores = 2,
           mart = useMart(biomart="ENSEMBL_MART_ENSEMBL",
              dataset="hsapiens_gene_ensembl",
              host="uswest.ensembl.org"))
}
\arguments{
\item{df}{a data.table got from function scGRN_getTF}

\item{gexpr}{a data frame or large matrix. Each row represents one gene and each column represents one sample.}

\item{df_gene_id, gexpr_gene_id}{the type of gene_id, in df and gexpr, should be 'hgnc_symbol' or 'ensembl_gene_id'.}

\item{cutoff_by}{the type of cutoff criterion, should be 'quantile' or 'absolute'. If 'quantile' is specified, the cutoff_percentage
will be used to pick the coefficients whose absolute value are larger that the quantile. Otherwise, cutoff_absolute will be used.}

\item{cutoff_percentage}{a user defined numeric value to stop the cutoff specified as a proportion 0 to 1}

\item{cutoff_absolute}{a user defined numeric value to stop the cutoff, defaults to 0.9}

\item{train_ratio}{train size, should be between 0.0 and 1.0 and represent the proportion of the dataset to be included in the training set.}

\item{num_cores}{number of cores used to do parallel computing}

\item{mart}{a dataset in BioMart database}
}
\value{
a data frame containing TG, TF, promoter, enhancer and coef.
}
\description{
Get gene regulatory network
}
\seealso{
glmnet
}
