% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGRN_getTF.R
\name{scGRN_getTF}
\alias{scGRN_getTF}
\title{scGRN_getTF}
\usage{
scGRN_getTF(df, database = JASPAR2018, species_type = 9606, min_score = 0.9,
                   pwm_type = 'prob',num_cores = 2)
}
\arguments{
\item{df}{a data frame which can from the output of function scGRN_interation or available interation data.
The data frame should contain gene, gene_chr, promoter_start, promoter_end,}

\item{database}{a database of transcription factor binding profiles which can be searched by TFBSTools.}

\item{species_type}{defaults to 9606, which represents human.}

\item{min_score}{quantile set to find the potential TFs for a given region, defaults to 0.9.}

\item{num_cores}{number of cores used to do parallel computing, speeding up the match process.}

\item{pwm_tpye}{the type of PWM generated, should be one of "log2probratio" or "prob".
"log2probratio" will generate the PWM matrix in log-scale, while "prob" will give the PWM matrix in probability scale of 0 to 1.
Inherited from toPWM function.}
}
\value{
a data.table containing gene, promoter, enhancer, promoter_TF, enhancer_TF. The type of elements in promoter_TF and
enhancer_TF is list.
}
\description{
Find transcription factors which bind to promoters and enhancers
}
\seealso{
JASPAR2018, TFBSTools
}
