% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGRN_interaction.R
\name{scGRN_interaction}
\alias{scGRN_interaction}
\title{scGRN_interaction : find interaction from original data}
\usage{
scGRN_interaction(hic_interaction, enhancers, ref_promoters = 'all',up_stream = 2500,
down_stream = 2500, link_type = 'within' ,target_genes='all',gene_id_option = 'hgnc_symbol',
mart = useMart(biomart='ENSEMBL_MART_ENSEMBL',dataset='hsapiens_gene_ensembl',
host='uswest.ensembl.org'))
}
\arguments{
\item{hic_interaction}{a data frame containing the variables : chr1,start1,end1,chr2,start2,end2
or chr,start1,end1,start2,end2.
chr1,ch2 or chr should have the following format 'chrD'(D represents digit 1-22 or X Y)
start1,end1,start2,end2 should be integer type.}

\item{enhancers}{a data frame containing chr,start,end for enhancers.}

\item{ref_promoters}{an optional data frame containing chr,start,end for promoters. If it is not supplied,
promoters are from package - TxDb.Hsapiens.UCSC.hg19.knownGene. If it is supplied, ref_promoters are used and
linked to target genes by package - TxDb.Hsapiens.UCSC.hg19.knownGene.}

\item{up_stream, down_stream}{The up_stream and down_stream arguments define the number of nucleotides in the 5' and 3' direction,
respectively. Used for annatating promoters.}

\item{link_type}{"any", "start", "end" or "within" are used when ref_promoters is provided. Inherited from
function GenomicRanges:findOverlapPairs.}

\item{target_genes}{a vector of target genes which should be consistent with the gene_id_option.}

\item{gene_id_option}{'hgnc_symbol' or 'ensembl_gene_id'.}

\item{mart}{a dataset in BioMart database.}
}
\value{
a data frame containing gene, gene_chr, promoter_start, promoter_end, enh_chr,enh_start,enh_end.
}
\description{
Find enhancers and promoters for target genes.
}
\details{
it would be a good option to set a relavant large upstream and downstream and link the ref_promoters to genes by within the whole promoters we annotated.
}
\seealso{
GenomicRanges,TxDb.Hsapiens.UCSC.hg19.knownGene,GenomicInteractions
}
