% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_silhouette.R
\name{score-asw}
\alias{score-asw}
\alias{ScoreASW}
\alias{AddScoreASW}
\alias{ScoreASWBatch}
\alias{AddScoreASWBatch}
\title{Score an embedding or a count matrix with the average silhouette width}
\usage{
ScoreASW(
  object,
  cell.var,
  what,
  assay = NULL,
  metric = c("euclidean", "cosine", "angular", "manhattan", "hamming"),
  dist.package = c("distances", "Rfast", "parallelDist", "stats"),
  verbose = TRUE,
  ...
)

AddScoreASW(
  object,
  integration,
  cell.var,
  what,
  assay = NULL,
  metric = c("euclidean", "cosine", "angular", "manhattan", "hamming"),
  dist.package = c("distances", "Rfast", "parallelDist", "stats"),
  verbose = TRUE,
  ...
)

ScoreASWBatch(
  object,
  batch.var = NULL,
  cell.var = NULL,
  what,
  per.cell.var = TRUE,
  assay = NULL,
  metric = c("euclidean", "cosine", "angular", "manhattan", "hamming"),
  dist.package = c("distances", "Rfast", "parallelDist", "stats"),
  verbose = TRUE,
  ...
)

AddScoreASWBatch(
  object,
  integration,
  batch.var = NULL,
  cell.var = NULL,
  what,
  per.cell.var = TRUE,
  assay = NULL,
  metric = c("euclidean", "cosine", "angular", "manhattan", "hamming"),
  dist.package = c("distances", "Rfast", "parallelDist", "stats"),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{A Seurat object}

\item{cell.var}{The name of the column with cell type label variable
(must be in the object metadata).
Ignored by \code{ScoreASWBatch} when \code{per.cell.var = FALSE}}

\item{what}{the name of the dimension reduction or layer to score. Must be
in the Seurat object or obtainable after a
\code{\link[SeuratObject:JoinLayers]{JoinLayers()}} call.}

\item{assay}{name of the assay to use. The output of
\code{\link[SeuratObject:DefaultAssay]{DefaultAssay()}} is used by default}

\item{metric}{name of the distance metric to use. One of 'euclidean',
'cosine', 'angular', 'manhattan', 'hamming'. See \strong{Note} for details.}

\item{dist.package}{name of the package to compute distances with. One of
'distances', 'Rfast' ,'parallelDist', 'stats'. The latter is always available,
the others must be installed beforehand. They are ordered from fastest to
slowest. When the requested package is not installed, the fastest amongst the
available ones is picked. See \strong{Note} for details.}

\item{verbose}{Print messages. Set to \code{FALSE} to disable}

\item{...}{additional parameters to pass to the distance computation
functions}

\item{integration}{name of the integration to score}

\item{batch.var}{The name of the column with batch variable.
(must be in the object metadata). Required by \code{ScoreASWBatch}.}

\item{per.cell.var}{whether to compute silhouette coefficients with the batch
variable for each cell-type separately (default behaviour). Setting to
\code{FALSE} causes the silhouette coefficients to be computed on the whole
data directly.}
}
\value{
\code{ScoreASW} and \code{ScoreASWBatch}: a single float between 0
and 1, corresponding to the scaled average silhouette score.

\code{AddScoreASW} and \code{AddScoreASWBatch}:  the updated Seurat
\code{object} with the ASW score(s) set for the integration.
}
\description{
Compute scores based on the average silhouette width (ASW) metric.

\code{ScoreASW}: First, cell-to-cell distances are computed on the provided
embedding or layer matrix. Then, the silhouette score is calculated to
estimate the quality of the partition according to the variable with cell
type labels. Hence, this score measures to what extent cells with identical
label cluster together.

\code{ScoreASWBatch}: Similar, but with the batch variable. This score
provides an estimation of batch mixing.
}
\details{
\code{ScoreASW}: Given a matrix (reduction dimension or layer), the
cell-to-cell distance matrix \eqn{D} is computed. Then, the silhouette width
\eqn{s(i)} is calculated for each cell \eqn{i} with a label \eqn{c \in L}
(\eqn{L} is the set of possible cell type labels). Then, the mean of all
\eqn{s(i)} is computed (i.e. the ASW) and scaled between 0 and 1:
\deqn{\displaystyle ASW = \frac{1}{\left| L \right|} \times \sum_{i}{s(i)} \\[10pt]
score = \frac{ASW + 1}{2}}

\code{ScoreAWSBatch}: The default parameters (\code{per.cell.var = TRUE})
correspond to the original score from Luecken M.D. \emph{et al.}, 2022. It is
computed as follow: for each cell type label \eqn{c} among the set of all
labels \eqn{L}, the cell-to-cell matrix distance \eqn{D_c} is computed for
cells \eqn{i} with label \eqn{c}. Each cell's silhouette width \eqn{s(i)} is
calculated according to the batch variable and transformed to be as close to
1 as its absolute value is close to 0. An average silhouette width
\eqn{ASW_c} is then computed per cell type label \eqn{c \in L} and the mean
of those correspond to the final score:
\deqn{\displaystyle ASW_c = \frac{1}{\left| c \right|} \times \sum_{i \in c}{1 - \left| s(i) \right|} \\[10pt]
score = \frac{1}{\left| L \right|} \times \sum_{c \in L}{ASW_c}}

When \code{per.cell.var = FALSE}, \eqn{ASW} is computed for all cells at once
(just like for \code{ScoreASW} but on the batch variable), then scaled
and averaged similarly:
\deqn{\displaystyle score = ASW = \frac{1}{N} \times \sum_{i=1}^{N}{1 - \left| s(i) \right|}}
with \eqn{N} being the total number of cells
}
\note{
Those scores are an adaptation of the (cell-type) ASW and the batch ASW
as described in Luecken M.D. \emph{et al.}, 2022.

Hamming distance is only supported by the \pkg{parallelDist} package, while
\pkg{distances} can only compute euclidean and related distances (cosine and
angular). Angular distances are actually refereed to as 'cosine' in
\code{\link[Seurat:FindNeighbors]{FindNeighbors()}} (\code{annoy.metric}),
hence called 'angular' here. Actual cosine dissimilarity-derived distances
are returned when \code{metric = 'cosine'}. Internally, angular distances are
computed with euclidean distances on \eqn{L^2} norm. cosine distances are
further transformed with :
\deqn{\displaystyle D_{cosine} = \frac{D_{angular}^2}{2}}
}
\references{
Luecken, M. D., Büttner, M., Chaichoompu, K., Danese, A.,
Interlandi, M., Mueller, M. F., Strobl, D. C., Zappia, L., Dugas, M.,
Colomé-Tatché, M. & Theis, F. J. Benchmarking atlas-level data integration in
single-cell genomics. Nat Methods 19, 41–50 (2021).
\href{https://doi.org/10.1038/s41592-021-01336-8}{DOI}
}
\seealso{
\code{\link[Seurat]{FindNeighbors}}, \code{\link[cluster]{silhouette}}
to know more about the silhouette metric, \code{\link{GetNeighborsPerBatch}}
and \code{\link{GetPropInterBatch}}
}
