% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_pca.R
\name{ScoreRegressPC.CellCycle}
\alias{ScoreRegressPC.CellCycle}
\alias{AddScoreRegressPC.CellCycle}
\title{Score a corrected or uncorrected PCA to estimate the contribution of S and G2M
scores to variance}
\usage{
ScoreRegressPC.CellCycle(
  object,
  batch.var = NULL,
  what = NULL,
  dims.use = NULL,
  npcs = 50L,
  s.var = "S.Score",
  g2m.var = "G2M.Score",
  compute.cc = TRUE,
  s.features = NULL,
  g2m.features = NULL,
  assay = NULL,
  weight.by = c("var", "stdev"),
  adj.r2 = FALSE,
  approx = FALSE
)

AddScoreRegressPC.CellCycle(
  object,
  integration,
  batch.var = NULL,
  what = NULL,
  dims.use = NULL,
  npcs = 50L,
  s.var = "S.Score",
  g2m.var = "G2M.Score",
  compute.cc = TRUE,
  s.features = NULL,
  g2m.features = NULL,
  assay = NULL,
  weight.by = c("var", "stdev"),
  adj.r2 = FALSE,
  approx = FALSE
)
}
\arguments{
\item{object}{A Seurat object}

\item{batch.var}{The name of the batch variable (must be in the object metadata)}

\item{what}{the slot from Seurat to score. Can be a layer or a reduction.}

\item{dims.use}{The dimensions from \code{what} to consider. All dimensions
are used by default}

\item{npcs}{Total Number of PCs to compute and store (50 by default)}

\item{s.var}{The name of the S phase score variable (must be in the object
metadata)}

\item{g2m.var}{The name of the G2M phase score variable (must be in the
object metadata)}

\item{compute.cc}{whether to (re-)compute the cell cycle scores. Should be
\code{TRUE} (default), unless you have run
\code{\link{CellCycleScoringPerBatch}} beforehand because cell cycles scores
are expected to be computed per batch}

\item{s.features}{A vector of features associated with S phase}

\item{g2m.features}{A vector of features associated with G2M phase}

\item{assay}{assay to use. Passed to Seurat to automatically construct the
\code{batch.var} when not provided. Useless otherwise}

\item{weight.by}{one of 'var' (default) or 'stdev' (standing for variance and
standard deviation respectively). Use the variance or the standard deviation
explained by the principal components to weight the each PC's score.}

\item{adj.r2}{Whether to use the adjusted R2 instead of the raw R2}

\item{approx}{Use truncated singular value decomposition to approximate PCA}

\item{integration}{name of the integration to score}
}
\value{
\code{ScoreRegressPC.CellCycle}: A 2-columns data frame with the
batch variable in the first one and the corresponding score in the second
one. It has as many rows as batches.

\code{AddScoreRegressPC.CellCycle}: the updated Seurat \code{object} with the
cell cycle conservation score set for the integration.
}
\description{
Linearly regresses S and G2M scores to predict principal components. The
resulting R2 are then weighted by each dimension's contribution to variance.
Cell cycles scores and PCAs are computed for each batch independently.
}
\details{
The linear regression is
\deqn{PC_i = S_{score} + G2M_{score}}

The score is computed as follow :
\deqn{\sum_{i=1}^{p} \left ( R^2_i * V_i \right )}

For a PCA with p dimensions, \eqn{PC_i} is the principal component i,
\eqn{R^2_i} is the R squared coefficient of the linear regression for the
dimension i. \eqn{V_i} is the proportion of variance explained by the
\eqn{PC_i}.
}
\note{
This score is an adaptation of the principal component regression (PCR)
score from Luecken M.D. \emph{et al.}, 2022.
}
\examples{
\dontrun{
obj <- SeuratData::LoadData("pbmcsca")
obj[["RNA"]] <- split(obj[["RNA"]], f = obj$Method)
obj <- NormalizeData(obj)
obj <- FindVariableFeatures(obj)
obj <- ScaleData(obj)
obj <- RunPCA(obj)

score.cc.r2 <- ScoreRegressPC.CellCycle(obj, "Method", "pca", dim.use = 1:30)
score.cc.adj.r2 <- ScoreRegressPC.CellCycle(obj, "Method", "pca", dim.use = 1:30, adj.r2 = TRUE)

score.cc.r2        # ~ 0.0249
score.cc.adj.r2    # ~ 0.0249
}

}
\references{
Luecken, M. D., Büttner, M., Chaichoompu, K., Danese, A.,
Interlandi, M., Mueller, M. F., Strobl, D. C., Zappia, L., Dugas, M.,
Colomé-Tatché, M. & Theis, F. J. Benchmarking atlas-level data integration in
single-cell genomics. Nat Methods 19, 41–50 (2021).
\href{https://doi.org/10.1038/s41592-021-01336-8}{DOI}
}
\seealso{
\code{\link{CellCycleScoringPerBatch}} to compute cc scores per
batch. \code{\link{ScoreRegressPC}} to regresses PCs by batch.
}
