% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_kbet.R
\name{ScoreKBET}
\alias{ScoreKBET}
\alias{AddScoreKBET}
\title{Score an embedding or a knn graph with the kBET test}
\usage{
ScoreKBET(
  object,
  batch.var,
  cell.var,
  what,
  graph.type = c("distances", "connectivities"),
  seed.use = 42L,
  verbose = TRUE,
  assay = NULL,
  layer = NULL
)

AddScoreKBET(
  object,
  integration,
  batch.var,
  cell.var,
  what,
  graph.type = c("distances", "connectivities"),
  seed.use = 42L,
  verbose = TRUE,
  assay = NULL,
  layer = NULL
)
}
\arguments{
\item{object}{A Seurat object}

\item{batch.var}{The name of the batch variable (must be in the object
metadata)}

\item{cell.var}{The name of the column with cell type label variable
(must be in the object metadata). If \code{NULL} is passed, the kBET score is
computed for all cells at once.}

\item{what}{the name of the dimension reduction or Graph/Neighbor object to
score. Must be in the Seurat object.}

\item{graph.type}{one of 'distances' (default) or 'connectivities', to
indicate the type of graph when \code{what} is not a dimension reduction.}

\item{seed.use}{the value of the seed to obtain reproducible results.
\code{NULL} disables to use of a seed}

\item{verbose}{whether to print progress messages}

\item{assay}{assay to use. Passed to Seurat to automatically construct the
\code{batch.var} when not provided
(\code{\link[SeuratObject:DefaultAssay]{DefaultAssay()}} is used by default).
Useless otherwise}

\item{layer}{layer to use. Passed to Seurat to automatically construct the
\code{batch.var} when not provided. Useless otherwise}

\item{integration}{name of the integration to score}
}
\value{
\code{ScoreKBET}: a list with one element per \code{cell.var}. Each
is a named numeric vector of kBET score (floats between 0 and 1), one per
cell-label.

\code{AddScoreKBET}:  the updated Seurat \code{object} with the mean kBET
score(s) set for the integration.
}
\description{
Compute scores based on the k-nearest neighbour batch effect test (kBET).
Accepts a \code{\link[SeuratObject:DimReduc-class]{DimReduc}} instance or a
graph object (\code{\link[SeuratObject:Graph-class]{Graph}} or
\code{\link[SeuratObject:Neighbor-class]{Neighbor}}) provided that it
contains distances or connectivities.

If '\code{what}' is a \code{\link[SeuratObject:DimReduc-class]{DimReduc}}
object, a knn graph is computed for \code{k = 50}. Then, (or if '\code{what}'
is a \code{\link[SeuratObject:Graph-class]{Graph}} or a
\code{\link[SeuratObject:Neighbor-class]{Neighbor}} object), connectivities
are derives from 50 nearest neighbours distances. Finally, the kBET test is
computed for each cell-type label of all provided \code{cell.var}. kBET
measures if local batch label distribution is similar to the global one. The
result of kBET is the average test rejection rate (between 0 and 1). The
closest to zero, the less bias is attributable to the batch effect.
}
\references{
Büttner, M., Miao, Z., Wolf, F. A., Teichmann, S. A. & Theis, F.
J. A test metric for assessing single-cell RNA-seq batch correction.
Nat Methods 16, 43–49 (2018).
\href{https://doi.org/10.1038/s41592-018-0254-1}{DOI}

Luecken, M. D., Büttner, M., Chaichoompu, K., Danese, A.,
Interlandi, M., Mueller, M. F., Strobl, D. C., Zappia, L., Dugas, M.,
Colomé-Tatché, M. & Theis, F. J. Benchmarking atlas-level data integration in
single-cell genomics. Nat Methods 19, 41–50 (2021).
\href{https://doi.org/10.1038/s41592-021-01336-8}{DOI}
}
