% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_pca.R
\name{ScoreDensityPC}
\alias{ScoreDensityPC}
\title{Score a corrected or uncorrected PCA to estimate batch mixing}
\usage{
ScoreDensityPC(
  object,
  batch.var = NULL,
  reduction = "pca",
  dims = NULL,
  use.union = TRUE,
  bw.join = mean,
  bw.kernel = c("nrd0", "nrd", "ucv", "bcv", "sj"),
  ...
)
}
\arguments{
\item{object}{A Seurat object}

\item{batch.var}{The name of the batch variable (must be in the object metadata)}

\item{reduction}{The name of the reduction to score}

\item{dims}{The dimensions to consider. All dimensions are used by default}

\item{use.union}{Whether to use a union of densities (constructs a fake
density with pmax). The alternative is to use all density kernels and sum them}

\item{bw.join}{A function to pick a joint bandwidth. The default is to use the
\code{mean()}. You can also chose \code{median} for instance. If you want to
use a distinct bandwidth for each batch, just set \code{bw.join = c} (not recommended)}

\item{bw.kernel}{Bandwidth selector to use for Gaussian kernel density estimation.
One of 'nrd0', 'nrd', 'ucv','bcv' or 'sj' (see \link[stats:bw.nrd]{bandwidths}).
nrd0 is the default}

\item{...}{Additional parameters ('assay', 'search' and 'scale.layer') to pass
to Seurat to automatically construct the \code{batch.var} when not provided.}
}
\value{
A single float corresponding to the score of the given reduction
}
\description{
A batch-wise kernel density estimate is computed for each dimension of the PCA.
Then, the global area shared by all kernels is calculated and divided by a
reference area (see \strong{Arguments} and \strong{Details}) as a proxy to
batch mixing. Finally, the aforementioned proportions are weighted by each
dimension's contribution to variance.
}
\details{
The score is computed as follow :
\deqn{\sum_{i=1}^{p} \left ( \cfrac{ \bigcup_{b=1}^n D_{bi} }{ Ref_i } * V_i \right )}

For a PCA with p dimensions and n batchs. \eqn{D_{bi}} is the vector of density
of batch b on the dimension i. \eqn{V_i} is the proportion of variance
explained by the \eqn{PC_i}.

with \eqn{Ref = \bigcap_{i=1}^n D_{bi}} or \eqn{Ref = \sum_{i=1}^n D_{bi}}
when \code{use.union} is \code{TRUE} and \code{FALSE} respectively, for a
given dimension \eqn{i}.

\code{use.union = FALSE} tends to result in lower score since the
reference is the sum of kernel densities instead of maximum
}
\note{
This score is an adaptation of the principal component regression (PCR)
score from Luecken M.D. \emph{et al.}, 2022.
}
\examples{
\dontrun{
obj <- SeuratData::LoadData("pbmcsca")
obj[["RNA"]] <- split(obj[["RNA"]], f = obj$Method)
obj <- NormalizeData(obj)
obj <- FindVariableFeatures(obj)
obj <- ScaleData(obj)
obj <- RunPCA(obj)

score.union <- ScoreDensityPC(obj, "Method", "pca", dim = 1:30)
score.sum <- ScoreDensityPC(obj, "Method", "pca", dim = 1:30, use.union = FALSE)

score.union   # ~ 0.1511
score.sum     # ~ 0.0319
}

}
\references{
Luecken, M. D., Büttner, M., Chaichoompu, K., Danese, A.,
Interlandi, M., Mueller, M. F., Strobl, D. C., Zappia, L., Dugas, M.,
Colomé-Tatché, M. & Theis, F. J. Benchmarking atlas-level data integration in
single-cell genomics. Nat Methods 19, 41–50 (2021).
\href{https://doi.org/10.1038/s41592-021-01336-8}{DOI}
}
\seealso{
\code{\link[Seurat]{Layers}} and \code{\link[Seurat]{DefaultAssay}}
for \code{...} arguments and \code{\link{ScoreRegressPC}} to compute the PCR
score
}
