% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_pca.R
\name{ScoreRegressPC}
\alias{ScoreRegressPC}
\title{Score a corrected or uncorrected PCA to estimate batch mixing}
\usage{
ScoreRegressPC(
  object,
  batch.var = NULL,
  reduction = "pca",
  dims = NULL,
  all.at.once = FALSE,
  adj.r2 = FALSE,
  ...
)
}
\arguments{
\item{object}{A Seurat object}

\item{batch.var}{The name of the batch variable (must be in the object metadata)}

\item{reduction}{The name of the reduction to score}

\item{dims}{The dimensions to consider. All dimensions are used by default}

\item{all.at.once}{Whether to regress all the dimensions at once instead of
regressing them separately. Note that enabling this behaviour doesn't allow
to weight R2s by the contribution to variance}

\item{adj.r2}{Whether to use the adjusted R2 instead of the raw R2}

\item{...}{Additional parameters ('assay', 'search' and 'scale.layer') to pass
to Seurat to automatically construct the \code{batch.var} when not provided.}
}
\value{
A single float corresponding to the score of the given reduction
}
\description{
Linearly regresses principal components to predict batches as a proxy to
batch mixing. The resulting R2 are then weighted by each dimension's
contribution to variance.
}
\details{
The linear regression is
\deqn{Batch = PC_i} or
\deqn{Batch = \sum_{i=1}^{p}PC_i} when \code{all.at.once = TRUE}


The score is computed as follow :
\deqn{\sum_{i=1}^{p} \left ( R^2_i * V_i \right )} or just
\deqn{R^2} when \code{all.at.once = TRUE}

For a PCA with p dimensions, \eqn{PC_i} is the principal component i,
\eqn{R^2_i} is the R squared coefficient of the linear regression for the
dimension i. \eqn{V_i} is the proportion of variance explained by the
\eqn{PC_i}.
}
\note{
This score is an adaptation of the principal component regression (PCR)
score from Luecken M.D. \emph{et al.}, 2022.
}
\examples{
\dontrun{
obj <- SeuratData::LoadData("pbmcsca")
obj[["RNA"]] <- split(obj[["RNA"]], f = obj$Method)
obj <- NormalizeData(obj)
obj <- FindVariableFeatures(obj)
obj <- ScaleData(obj)
obj <- RunPCA(obj)

score.each <- ScoreRegressPC(obj, "Method", "pca", dim = 1:30)
score.all <- ScoreRegressPC(obj, "Method", "pca", dim = 1:30, all.at.once = TRUE)

score.each    # ~ 0.0257
score.all     # ~ 0.7553
}

}
\references{
Luecken, M. D., Büttner, M., Chaichoompu, K., Danese, A.,
Interlandi, M., Mueller, M. F., Strobl, D. C., Zappia, L., Dugas, M.,
Colomé-Tatché, M. & Theis, F. J. Benchmarking atlas-level data integration in
single-cell genomics. Nat Methods 19, 41–50 (2021).
\href{https://doi.org/10.1038/s41592-021-01336-8}{DOI}
}
\seealso{
\code{\link[Seurat]{Layers}} and \code{\link[Seurat]{DefaultAssay}}
for \code{...} arguments, \code{\link{ScoreDensityPC}} for an alternative and
\code{\link{ScoreRegressPC.CellCycle}} to regresses PCs by cell cycle scores.
}
