% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_pca_components.R
\name{select_pca_components}
\alias{select_pca_components}
\title{PCA Components Selection}
\usage{
select_pca_components(data, vars = NULL, perc = 80, eigenvalues = FALSE)
}
\arguments{
\item{data}{Data frame to be used.}

\item{vars}{Variables used from the data frame for the PCA}

\item{perc}{If 'eigenvalues==F', the principal components selected are the ones for which the at least perc percentage of the variability is explained.}

\item{eigenvalues}{If TRUE the principal components selected are the ones whose eigenvalues are greater than the mean of the eigenvalues.}
}
\value{
A list containing:
\describe{
\item{method}{A character indicating the method used for the component selectoin}
\item{num_components}{Number of components selected}
\item{explained_variance}{Total percentage of variance eplained by the selected components}
\item{pca_result}{Full PCA result from \code{FactoMineR::PCA}}
\item{selected_pca_result}{Subset of PCA results containing only selected components}
}
}
\description{
Performs a PCA on the numeric variables of a dataset and selects the number of
components to retain based on the cumulative explained variance or the eigenvalues
criterion
}
\examples{
select_pca_components(iris, perc=90)
select_pca_components(iris, vars=c("Sepal.Length", "Sepal.Width"), eigenvalues=TRUE)
}
