% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self_res_att.R
\docType{data}
\name{self_res_att}
\alias{self_res_att}
\title{Attractiveness of Self-Resembling Faces.}
\format{
A data frame with 108 rows and 16 variables:
\describe{
\item{col1}{col1 description}
\item{col2}{col2 description}
}
}
\source{
\url{https://osf.io/3c5s4/}
}
\usage{
self_res_att
}
\description{
Attractiveness judgements of same-sex and other-sex self-resembling faces from DeBruine (2004).
}
\keyword{datasets}
