% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_a_report.R
\name{server_analysis_a_report}
\alias{server_analysis_a_report}
\title{server_analysis_a_report}
\usage{
server_analysis_a_report(id = "analysis_a_report", server_input)
}
\description{
server_analysis_a_report
}
