% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{SetMiscScore}
\alias{SetMiscScore}
\title{Set the value of a score in the score tibble}
\usage{
SetMiscScore(object, integration, score.name, score.value, ...)
}
\arguments{
\item{object}{a Seurat object}

\item{integration}{the name of the integration for which the score was
computed.}

\item{score.name}{the name of the computed score}

\item{score.value}{the value of the score}

\item{...}{additional parameter to pass to \code{\link{AddMiscScores}}. Has
no effect if the score slot is already present in the tibble. Otherwise,
enable to pass the \code{class} argument to specify the class of the score
slot to create.}
}
\value{
the Seurat object with updated table of scores.
}
\description{
Set a score value for a given integration and a given score name. If they
don't exist yet, they are created though \code{\link{AddMiscIntegrations}}
and \code{\link{AddMiscScores}} respectively.
}
