% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_package.R
\name{setUpPackageComponents}
\alias{setUpPackageComponents}
\title{function to set up the R package components (licence, README, pkgdown website)}
\usage{
setUpPackageComponents()
}
\description{
function to set up the R package components (licence, README, pkgdown website)
}
\examples{
setUpPackageComponents()
}
