% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratIntegrate-package.R
\docType{package}
\name{SeuratIntegrate-package}
\alias{SeuratIntegrate}
\alias{SeuratIntegrate-package}
\title{SeuratIntegrate: Expands the set of integration methods available to Seurat}
\description{
R package gathering a set of wrappers to apply various integration methods to Seurat objects (and rate such methods). Intended to apply to Seurat V5 objects bearing multiple layers.
}
\author{
\strong{Maintainer}: Florian Specque \email{florian.specque@u-bordeaux.fr} (\href{https://orcid.org/0000-0002-4449-6381}{ORCID})

}
\keyword{internal}
