% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.R
\name{simple.mediation}
\alias{simple.mediation}
\title{fit a simple mediation model}
\usage{
simple.mediation(rxm = -2, rxy = -2, rym = -2, nobs = 0)
}
\arguments{
\item{rxm}{input of the correlation between X and M}

\item{rxy}{input of the correlation between X and Y}

\item{rym}{input of the correlation between Y and M}

\item{nobs}{input of the number of observations in your sample}
}
\value{
product of the two input numbers
}
\description{
give the parameter estimates of a simple mediation model
}
