% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{simulate_testing_data}
\alias{simulate_testing_data}
\title{Simulate Dummy Data}
\usage{
simulate_testing_data(n)
}
\arguments{
\item{n}{A \code{double} value of the size of dataset we want to create.}
}
\value{
A \code{list} value containing the following objects:
\describe{
 \item{response}{The \code{double} vector containing the response used for the estimation.}
 \item{predictors}{The \eqn{n \times p} \code{double} value of the matrix containing the values of the predictors used for the estimation.}
}
}
\description{
This function simulates dummy data that can be used to test the functionality of this package.
}
\author{
Saksham Goel
}
