% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulecdf.R
\name{simulecdf}
\alias{simulecdf}
\title{Simulate the test statistic for the robust independence test of Kolmogorov-Smirnov's type}
\usage{
simulecdf(n, N)
}
\arguments{
\item{n}{the size of the sample.}

\item{N}{the number of replications in the Monte-Carlo simulation.}
}
\value{
Returns the e.c.d.f. of the N Monte_Carlo simulations. The returned object
is a stepfun object obtained from the function ecdf.
}
\description{
For two continuous uniform variables on [0,1] compute the maximal distance
between the joint empirical cumulative distribution function and the product of the marginal
empirical cumulative distribution functions.
}
\details{
Let X and Y be two continuous uniform variables on [0,1].

For observations (x1,y1), ..., (x_n,y_n), the bivariate e.c.d.f.
(empirical cumulative distribution function) Fn is defined as

Fn(t1,t2)=#\code{{xi<=t1,yi<=t2}/n= sum_{i=1}^n Indicator(xi<=t1,yi<=t2)/n}

Let Fn(t1) and Fn(t2) be the marginals e.c.d.f. For the N Monte_Carlo simulations,
the function computes

\code{n^(1/2) sup_{t1,t2} |Fn(t1,t2)-Fn(t1)*Fn(t2)|}
}
\seealso{
Other test functions: 
\code{\link{ecdf2D}()},
\code{\link{stat_robustest}()}
}
\concept{test functions}
\keyword{test}
