% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_domain_plot.R
\name{sort_domains}
\alias{sort_domains}
\title{Sort one domain dataframe based on the other domain dataframe}
\usage{
sort_domains(seed_df, ortho_df)
}
\arguments{
\item{seed_df}{data of seed protein}

\item{ortho_df}{data of ortholog protein}
}
\value{
sorted domain list (dataframe)
}
\description{
Sort domain dataframe of one protein (either seed or ortholog)
based on the dataframe of the its paired protein, in order to bring the
common domain feature in the same order which make it easy for comparing.
}
\examples{
\dontrun{
# get domain data
domain_file <- system.file(
    "extdata", "domain_files/OG_1009.domains",
    package = "phyloprofile", mustWork = TRUE
)
domain_df <- parse_domain_input(seedID, domain_file, "file")
# get seed_df and ortho_df
sub_df <- domain_df[domain_df$seedID == "OG_1009#A.thaliana@3702@241736",]
ortho_df <- dplyr::filter(sub_df, orthoID == "A.thaliana@3702@241736")
seed_df <- dplyr::filter(sub_df, orthoID != "A.thaliana@3702@241736")
# sort
sort_domains(seed_df, ortho_df)
}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
