% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_phylo_profile.R
\name{sort_input_taxa}
\alias{sort_input_taxa}
\title{Sort list of (super)taxa based on a selected reference (super)taxon}
\usage{
sort_input_taxa(taxon_IDs, taxon_names, rank_name, ref_taxon, taxa_tree)
}
\arguments{
\item{taxon_IDs}{list of taxon IDs (e.g.: ncbi1234, ncbi9999, ...)}

\item{taxon_names}{list of taxon names and their corresponding taxonomy
ranks, IDs,...}

\item{rank_name}{working taxonomy rank (e.g. "species", "phylum",...)}

\item{ref_taxon}{selected reference taxon}

\item{taxa_tree}{input taxonomy tree (optional)}
}
\value{
Taxonomy matrix for the input taxa ordered by the selected reference
taxon. This matrix is sorted either based on the NCBI taxonomy info, or
based on an user-defined taxonomy tree (if provided).
}
\description{
Sort list of (super)taxa based on a selected reference (super)taxon
}
\examples{
taxon_IDs <- c(
    "ncbi272557", "ncbi176299", "ncbi3702", "ncbi876142", "ncbi9606"
)
taxon_names <- get_input_taxa_name("species", taxon_IDs)
sort_input_taxa(taxon_IDs, taxon_names, "species", "Homo sapiens", NULL)
}
\seealso{
\code{\link{get_name_list}}, \code{\link{get_taxonomy_matrix}},
\code{\link{create_rooted_tree}}, \code{\link{sort_taxa_from_tree}},
\code{\link{get_input_taxa_name}}, \code{\link{get_input_taxa_id}},
\code{\link{create_long_matrix}}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
