% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_phylotree.R
\name{sort_taxa_from_tree}
\alias{sort_taxa_from_tree}
\title{Get sorted supertaxon list based on a rooted taxonomy tree}
\usage{
sort_taxa_from_tree(tree)
}
\arguments{
\item{tree}{rooted taxonomy tree}
}
\value{
list of taxa sorted according to the taxonomy tree
}
\description{
Get sorted supertaxon list based on a rooted taxonomy tree
}
\examples{
data("pp_taxonomy_matrix", package = "phyloprofile")
# prepare matrix for calculating distances
distDf <- subset(pp_taxonomy_matrix, select = -c(ncbiID, fullName))
row.names(distDf) <- distDf$abbrName
distDf <- distDf[, -1]
# create taxonomy tree rooted by ncbi10090
rooted_tree <- create_rooted_tree(distDf, "ncbi10090")
# get taxon list sorted from tree
sort_taxa_from_tree(rooted_tree)
}
\seealso{
\code{\link{pp_taxonomy_matrix}} for a demo taxonomy matrix data
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
