% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialAnno.R
\name{SpatialAnno}
\alias{SpatialAnno}
\title{SpatialAnno.}
\usage{
SpatialAnno(
  X,
  Adj_sp,
  markers,
  initial = c("SCINA", "scSorter"),
  Unknown = TRUE,
  xi_int = 1.5,
  xi_grid = seq(0.1, 2.5, by = 0.2),
  mode = "full",
  q = 15
)
}
\arguments{
\item{X}{a normalized matrix. The row is for spots/cells, and the column is for genes.}

\item{Adj_sp}{a integer vector used in SC.MEB. The default is 2:10}

\item{markers}{a list that contain cell types and corresponding cell-type specific markers.}

\item{initial}{character specifying which method to generate initial value. In general, We recommend to use the method of SCINA for high-dimensional non-markers and the method of scSorter for low-dimensional non-marker. The default is SCINA.}

\item{Unknown}{logical value specifying whether or not include unknown}

\item{xi_int}{a initial value of smooth parameter}

\item{xi_grid}{a pre-defined vector of smooth parameter}

\item{mode}{a string specifying which mode to use. It has four modes: "full", "annotation", "spatial+annotation" and "annotation+factor". The "full" mode means SpatialAnno runs with all three components, marker component, spatial component, and factor component. The "annotation" mode means SpatialAnno runs with only marker component. The "spatial+annotation" means SpatialAnno runs with marker and spatial components. The "annotation+factor" mode means SpatialAnno runs with marker and non-marker components. The default is "full".}

\item{q}{the dimension of embedding generated by SpatialAnno. The default is 15}
}
\value{
a list.
}
\description{
SpatialAnno, an efficient and accurate annotation method for spatial transcriptomics datasets, with capability of effectively leveraging a large number of non-marker genes with “qualitative” information about marker genes, without using a reference dataset.
}
\details{
SpatialAnno, an efficient and accurate annotation method for spatial transcriptomics datasets, with capability of effectively leveraging a large number of non-marker genes with “qualitative” information about marker genes, without using a reference dataset.
}
