% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speed_and_performance.R
\name{speed_and_performance}
\alias{speed_and_performance}
\title{Speeding and performance test for different models}
\usage{
speed_and_performance(
  X,
  Y,
  ID,
  scale = TRUE,
  nRep = 5,
  nOuter = 6,
  nInner,
  varRatio = 0.75,
  DA = FALSE,
  fitness = c("AUROC", "MISS", "BER", "RMSEP"),
  method = c("PLS", " RF"),
  methParam,
  ML = FALSE,
  modReturn = FALSE,
  logg = FALSE,
  parallel = TRUE,
  repeatMUVR = 2
)
}
\arguments{
\item{varRatio}{= 0.75,}

\item{repeatMUVR}{This is to decide how many MUVR to repeat(since each MUVR may give a differnent value)}
}
\value{
time   average time for each MUVR

performance   average miss auc and rmsep
}
\description{
Speeding and performance test for different models
}
