% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoredProcedures.R
\name{SPFIInstrumentGet}
\alias{SPFIInstrumentGet}
\title{Internal Solvas|Capital Function 
Get financial instruments}
\usage{
SPFIInstrumentGet(connection_string, fs_id, effective_date, effective_period)
}
\arguments{
\item{connection_string}{The string representing the connection string...}

\item{fs_id}{The scenario ID}

\item{effective_date}{- The date to use for schedule data types}

\item{effective_period}{- The period to use for schedule data types (1=first period)}
}
\value{
data frame containing the financial instrument data
}
\description{
Internal Solvas|Capital Function 
Get financial instruments
}
\keyword{internal}

