% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoredProcedures.R
\name{SPFSAccountBalancePut}
\alias{SPFSAccountBalancePut}
\title{Internal Solvas|Capital Function 
Saves Solvas|Capital account balances}
\usage{
SPFSAccountBalancePut(connection_string, fs_id, df)
}
\arguments{
\item{connection_string}{- the string representing the connection string...}

\item{fs_id}{The scenario ID}

\item{df}{- data frame of account balances (must match SPFSInitialAccountBalanceGet)}
}
\value{
result of the SP
}
\description{
Internal Solvas|Capital Function 
Saves Solvas|Capital account balances
}
\keyword{internal}
