% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoredProcedures.R
\name{SPFSAssumptionsGet}
\alias{SPFSAssumptionsGet}
\title{Internal Solvas|Capital Function 
gets Solvas|Capital assumptions}
\usage{
SPFSAssumptionsGet(connection_string, fs_id, tm_desc, use_dates)
}
\arguments{
\item{connection_string}{- the string representing the connection string...}

\item{fs_id}{The scenario ID}

\item{tm_desc}{- description/name of transformation (i.e. '(apply to all)').  the first}

\item{use_dates}{- if true then matrix columns will be dates, else periods. default to false}
}
\value{
data frame containing the economic factors
}
\description{
Internal Solvas|Capital Function 
gets Solvas|Capital assumptions
}
\keyword{internal}
