% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoredProcedures.R
\name{SPFSInitialAccountBalanceGet}
\alias{SPFSInitialAccountBalanceGet}
\title{Internal Solvas|Capital Function 
Gets Solvas|Capital initial account balances}
\usage{
SPFSInitialAccountBalanceGet(connection_string, fs_id,
  use_account_name = TRUE)
}
\arguments{
\item{connection_string}{- the string representing the connection string...}

\item{fs_id}{The scenario ID}

\item{use_account_name}{if TRUE will use account_name, false will use account_number}
}
\value{
data frame containing the account balances
}
\description{
Internal Solvas|Capital Function 
Gets Solvas|Capital initial account balances
}
\keyword{internal}
