% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoredProcedures.R
\name{SPInstrumentPropertyLoad}
\alias{SPInstrumentPropertyLoad}
\title{Internal Solvas|Capital Function 
Load Property Values}
\usage{
SPInstrumentPropertyLoad(connection_string, fs_id)
}
\arguments{
\item{connection_string}{- The string representing the connection string...}

\item{fs_id}{- The scenario id}
}
\value{
A data frame containing the financial instrument data
}
\description{
Internal Solvas|Capital Function 
Load Property Values
}
\keyword{internal}

