% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoredProcedures.R
\name{SPPackageVersionCompatible}
\alias{SPPackageVersionCompatible}
\title{Internal Solvas|Capital Function 
Checks the passed package version with the expected version on the current server for compatibility}
\usage{
SPPackageVersionCompatible(connection_string)
}
\arguments{
\item{connection_string}{- the string representing the connection string...}
}
\value{
boolean - true if they are compatible
}
\description{
Internal Solvas|Capital Function 
Checks the passed package version with the expected version on the current server for compatibility
}
\keyword{internal}
