% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoredProcedures.R
\name{SPResultCheck}
\alias{SPResultCheck}
\title{Internal Solvas|Capital Function 
Capital convention used to return errors from SP calls. All SP should use this before
using the results of an SP}
\usage{
SPResultCheck(data)
}
\arguments{
\item{data}{- dataframe returned from a sql call}
}
\value{
throws status_message if status is exit
}
\description{
Checks the data returned from a SP for a r_status_code column - if it exists and the values is 'CERR' 
then throw an error
Example:
  ...
  data <- sqlQuery(cn, sp, errors=TRUE)
  SPResultCheck(data)
  ...
}
\keyword{internal}
