% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\docType{class}
\name{spverse-class}
\alias{spverse-class}
\title{The sp class}
\description{
The AnchorSet class is an intermediate data storage class that stores the anchors and other
related information needed for performing downstream analyses - namely data integration
(\code{\link{IntegrateData}}) and data transfer (\code{\link{TransferData}}).
}
\section{Slots}{

\describe{
\item{\code{rawdata}}{proteins identification}

\item{\code{dim}}{data.frame.}

\item{\code{description}}{data.frame.}

\item{\code{color}}{data.frame.}

\item{\code{cluster}}{numeric.}

\item{\code{dis}}{numeric.}
}}

