% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StoredProcedures.R
\name{sqlIsNullStr}
\alias{sqlIsNullStr}
\title{Internal Solvas|Capital Function 
return sinqle quoated passed string if not null, return single quoate 'replacement_value' otherwise}
\usage{
sqlIsNullStr(string_value, replacement_value = "NULL")
}
\arguments{
\item{fs_id}{- The scenario id}
}
\value{
retuns a the passed string surrounded by single quoates if not na or null - otherwise returns unquoted string
}
\description{
Internal Solvas|Capital Function 
return sinqle quoated passed string if not null, return single quoate 'replacement_value' otherwise
}
\keyword{internal}
