% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square_root.R
\name{square_root}
\alias{square_root}
\title{test roxygen2
Return the square root of a number}
\usage{
square_root(x)
}
\arguments{
\item{x}{A number}
}
\description{
test roxygen2
Return the square root of a number
}
