% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGeneEx.R
\name{sracipeGeneEx}
\alias{sracipeGeneEx}
\title{Run GeneEx - webapp for sRACIPE}
\usage{
sracipeGeneEx()
}
\value{
Webapp interface
}
\description{
GeneEx enables interactive simulations, visualization, and data
analysis using sRACIPE.
}
