% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{sracipeGetTS}
\alias{sracipeGetTS}
\alias{sracipeGetTS,RacipeSE-method}
\title{A method to extract the time series}
\usage{
sracipeGetTS(.object)

\S4method{sracipeGetTS}{RacipeSE}(.object)
}
\arguments{
\item{.object}{RacipeSE object}
}
\value{
list
}
\description{
If timeSeries option is used in sracipeSimulate function, this
method will return the simulated time series.
}
\examples{

data("demoCircuit")
sracipeCircuit(RacipeSet) <- demoCircuit
RacipeSet <- sracipeSimulate(demoCircuit, timeSeries = T, simulationTime = 5)
trajectories <- sracipeGetTS(RacipeSet)
rm(RacipeSet)
}
