% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fasta_seqs.R
\name{str_reverse}
\alias{str_reverse}
\title{Reverse string}
\usage{
str_reverse(x)
}
\arguments{
\item{x}{input string}
}
\value{
a reversed string
}
\description{
Reverse string
}
\examples{
\dontrun{
str_reverse("abc")
}
}
