% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{string_length}
\alias{string_length}
\title{Strings Length}
\usage{
string_length(strings)
}
\arguments{
\item{strings}{A vector of character strings.}
}
\description{
Returns the number of characters in strings.
}
