% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_nums_sumit.R
\name{sub_nums_sumit}
\alias{sub_nums_sumit}
\title{Substraction program}
\usage{
sub_nums_sumit(arg1, arg2)
}
\arguments{
\item{arg1}{first argument}

\item{arg2}{second argument}
}
\value{
This function will return the substraction of the first 	argument given to the second argument
}
\description{
Substraction of 2 numbers
}
\examples{
sub_nums_sumit(12,6)
sub_nums_sumit (50,70)
}
\references{
Not Applicable
}
