\name{sum}
\alias{sum}
\title{Adds 2 numbers}
\usage{
sum(a,b)
}
\arguments{
\item{a}{1st number}
\item{b}{2nd number}
}
\value{
sum
}
\description{
This function adds any 2 numbers
}
\note{
Maintainer: A. Clever Guy\email{clever.guy@gmail.com}
}
\examples{
sum(5,3)
}
}
\author{
A. Clever Guy
}
\references{

}
