% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function sum.R
\name{sum1}
\alias{sum1}
\title{Sum two numbers}
\usage{
sum1(a, b)
}
\arguments{
\item{a}{first number}

\item{b}{second number}
}
\value{
the sum of a and b
}
\description{
Sum two numbers
}
