% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_nums_sumit.R
\name{sum_nums_sumit}
\alias{sum_nums_sumit}
\title{addition program}
\usage{
sum_nums_sumit(arg1, arg2)
}
\arguments{
\item{arg1}{first argument}

\item{arg2}{second argument}
}
\value{
This function will return the substraction of the first 	argument given to the second argument
}
\description{
addition of 2 numbers
}
\examples{
sum_nums_sumit(12,6)
sum_nums_sumit (50,70)
}
\references{
Not Applicable
}
