% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma_cuadrados.R
\name{suma.cuadrados}
\alias{suma.cuadrados}
\title{Add the squares of two numbers}
\usage{
suma.cuadrados(x, y)
}
\arguments{
\item{x}{a number}

\item{y}{a number}
}
\value{
\code{x}^2 \code{y}^2
}
\description{
Add the squares of two numbers
}
\examples{
suma.cuadrados(3,6)
}

