% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operaciones.R
\name{suma.dos.numeros}
\alias{suma.dos.numeros}
\title{Suma dos números}
\usage{
suma.dos.numeros(x, y)
}
\arguments{
\item{x}{Primer valor numérico a sumar}

\item{y}{Primer valor numérico a sumar}
}
\value{
La suma de las dos entradas
}
\description{
Recibe dos números y devuelve su suma
}
\examples{
suma.dos.numeros(1,0)
suma.dos.numeros(-5,3)
}

