% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmInner.R
\name{svmInner}
\alias{svmInner}
\title{Support vector machine model in inner CV loop}
\usage{
svmInner(x, y, kernal, cost, gamma)
}
\arguments{
\item{cost}{cost}

\item{xTrain}{Training data (samples as rows; variables as columns)}

\item{yTrain}{Training response}

\item{xVal}{Validation data}

\item{yVal}{Validation response (for tuning)}

\item{DA}{Logical for discriminant analysis (classification)}

\item{fitness}{Fitness function ('MISS', 'AUROC' or 'RMSEP')}

\item{ntree}{See original function (`randomForest`). Passed from wrapper.}

\item{method}{Choice of Random Forest implementation (randomForest, ranger or Rborist). Passed from wrapper.}
}
\value{
An object containing:

(`miss`,`ber`,`auc` or `rmsep`) A fitness metric

`virank` variable importance rankings
}
\description{
Support vector machine model in inner CV loop
}
