% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmpredict.R
\name{svmpredict}
\alias{svmpredict}
\title{RF predictions for outer segments and consensus models}
\usage{
svmpredict(
  xTrain,
  yTrain,
  xTest,
  yTest,
  scale,
  kernelmethod = c("linear", "radical", "polynomial", "sigmoid")
)
}
\arguments{
\item{DA}{}
}
\value{
The predicted value of yTest
}
\description{
RF predictions for outer segments and consensus models
}
