% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SymmetrizeKnn}
\alias{SymmetrizeKnn}
\alias{SymmetrizeKnn,Seurat-method}
\alias{SymmetrizeKnn,Matrix-method}
\alias{SymmetrizeKnn,Graph-method}
\alias{SymmetrizeKnn,Neighbor-method}
\title{Symmetrize a nearest neighbours graph}
\usage{
SymmetrizeKnn(object, graph.name = "RNA_nn", use.max = TRUE, assay = NULL)

\S4method{SymmetrizeKnn}{Seurat}(object, graph.name = "RNA_nn", use.max = TRUE, assay = NULL)

\S4method{SymmetrizeKnn}{Matrix}(object, use.max = TRUE)

\S4method{SymmetrizeKnn}{Graph}(object, use.max = TRUE)

\S4method{SymmetrizeKnn}{Neighbor}(object, use.max = TRUE)
}
\arguments{
\item{object}{a \code{Seurat}, \code{Graph} or \code{Neighbor} object}

\item{graph.name}{name of a \code{Graph} or \code{Neighbor} instance stored in
the \code{Seurat object}.}

\item{use.max}{by default, use the maximum value in case of discrepancy
between m[i,j] and m[j,i]. Set to \code{FALSE} to use the minimum value.}

\item{assay}{name of the assay to store in the output \code{Graph}}
}
\value{
the Seurat object with a new \code{Graph} instance or a
\code{dgCMatrix} representing the \code{Graph} itself
}
\description{
The approximate nature of the nearest neighbour search algorithm used to
compute the knn graph makes the resulting adjacency matrix asymmetric.
Those functions symmetrize a \code{Graph} or a \code{Neighbor} object.
}
\seealso{
The classes \code{\link[SeuratObject:Graph]{Graph}} and
\code{\link[SeuratObject:Neighbor]{Neighbor}} and
\code{\link{symmetrize.pmax.sparse}}
}
