% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_taxa_groups.R
\name{taxa_select_gc}
\alias{taxa_select_gc}
\title{print list of available taxa}
\usage{
taxa_select_gc(rank_select_gc, input_taxonID)
}
\arguments{
\item{rank_select_gc}{rank selected for group compariosn}

\item{input_taxonID}{contains "seedID",  "orthoID", "feature", "start", "end"}
}
\value{
avilable taxa containing "ncbiID", "fullName", "rank", "parentID"
}
\description{
print list of available taxa
}
\author{
Carla Mölbert (carla.moelbert@gmx.de)
}
