% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.r
\name{test}
\alias{test}
\title{A test Function}
\usage{
test(love = TRUE)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to print test results.
}
\examples{
cat_function()
}
\keyword{haha}

