% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_card.R
\name{test_analysis_card}
\alias{test_analysis_card}
\title{test_analysis_card}
\usage{
test_analysis_card(
  test_name = "Analysis A",
  test_description = stringi::stri_rand_lipsum(1),
  test_link = "analysis_a",
  ignore_dl = TRUE
)
}
\description{
test_analysis_card
}
