% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDarch.R
\name{test_DArch}
\alias{test_DArch}
\title{Wirte and test my first R package}
\usage{
test_DArch(input, target, ...)
}
\arguments{
\item{input}{input variable}

\item{target}{target variable}
}
\description{
A test function for R package development
}

