% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Test data for z-scores}
\format{a numeric vector, length 20}
\usage{
test_data
}
\description{
Data drawn from a uniform distribution between 10 and 100
}
\keyword{datasets}
