% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_energy.R
\docType{data}
\name{test_energy}
\alias{test_energy}
\title{Test data for energy mixes}
\format{a data frame of 7 rows and 2 columns
\describe{
\item{energy_type}{Abbreviation of the name of the energy source}
\item{probability}{The probability of drawing each type of energy based on global potential for each source}
}}
\usage{
test_energy
}
\description{
A selection of energy types and associated probabilities based on global energy potential maps
}
\keyword{datasets}
