% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_func.R
\name{test_func}
\alias{test_func}
\title{SGSS test function}
\usage{
test_func(data)
}
\arguments{
\item{data}{data frame to print - purely to proove the package has been installed correctly}
}
\description{
SGSS test function
}
