% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testfx.r
\name{testfx}
\alias{testfx}
\title{A basic function}
\usage{
testfx(test = TRUE)
}
\arguments{
\item{test}{Tests if the function is working. Defaults to TRUE.}
}
\description{
This is a test function I wrote.
}
\examples{
test()
}

