% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testWrap.R
\name{testWrap}
\alias{testWrap}
\title{testWrap: Wrapper for testing new functions in "Multivariate modelling with Unbiased Variable selection"}
\usage{
testWrap(X, Y, ID, nRep = 5, nOuter = 6, nInner, varRatio = 0.75,
  DA = FALSE, fitness = c("AUROC", "MISS", "RMSEP"), method = c("PLS",
  "RF"), methParam, ML = FALSE, modReturn = FALSE, newdata = NULL,
  logg = FALSE)
}
\arguments{
\item{X}{Independent variables. NB: Variables (columns) must have names/unique identifiers. NAs not allowed in data.}

\item{Y}{Response vector (Dependent variable). For PLS-DA, values should be -1 vs 1 for the two classes.}

\item{ID}{Subject identifier (for sampling by subject; Assumption of independence if not specified)}

\item{nRep}{Number of repetitions of double CV..}

\item{nOuter}{Number of outer CV loop segments.}

\item{nInner}{Number of inner CV loop segments.}

\item{varRatio}{Ratio of variables to include in subsequent inner loop iteration.}

\item{DA}{Logical for Classification (discriminant analysis) (Defaults do FALSE, i.e. regression). PLS is limited to two-class problems (see `Y` above).}

\item{fitness}{Fitness function for model tuning (choose either 'AUROC' or 'MISS' for classification; or 'RMSEP' (default) for regression.)}

\item{method}{Multivariate method. Supports 'PLS' and 'RF' (default)}

\item{methParam}{List with parameter settings for specified MV method (defaults to ???)}

\item{ML}{Logical for multilevel analysis (defaults to FALSE)}

\item{modReturn}{Logical for returning outer segment models (defaults to FALSE)}

\item{newdata}{New data matrix ONLY for prediction NOT modelling}

\item{logg}{Logical for whether to sink model progressions to `log.txt`}
}
\value{
An object containing stuff...
}
\description{
Repeated double cross validation with tuning of variables in the inner loop.
}

