% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfreq.R
\name{tfreq}
\alias{tfreq}
\title{Frequency and Contingency Tables for a Variable}
\usage{
tfreq(data, var, vec = NULL)
}
\arguments{
\item{data}{Dataset to which the variables belong.}

\item{var}{Variable 1 for which the frequency table is created.}

\item{vec}{Vector of variables with which contingency tables are created.}
}
\value{
Frequency table for the given variable and contingency tables for other variables.
}
\description{
Creates frequency tables for one variable or others chosen by the user, if desired.
}
\examples{
tfreq(iris, "Species")$freq_table;
}
