% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_Publication.R
\name{theme_Publication}
\alias{theme_Publication}
\title{theme_Publication.}
\usage{
theme_Publication(base_size = 14, base_family = "Helvetica")
}
\arguments{
\item{base_size}{the base size of text.}

\item{base_family}{the base type of plot.}
}
\value{
a ggplot.
}
\description{
theme_Publication
}
\details{
theme_Publication
}
