% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiebreak.R
\name{tiebreak}
\alias{tiebreak}
\title{Break the ties in a given vector or between two vectors}
\usage{
tiebreak(x, y = NULL)
}
\arguments{
\item{x, y}{the variables containing ties.}
}
\description{
If the vector contains ties (either inside a single or between two vectors), the function breaks them using a random perturbation.
}
\details{
If \code{y=NULL} the function detects the ties in the vector \code{x}. A uniform variable with parameters [-e^(-5),e^(-5)] is added
to the value of all the ties but one in the vector \code{x}. If \code{y} is also provided, the function detects the ties between
\code{x} and \code{y} and break them (only in the \code{x} vector) by adding a uniform variable with parameters [-e^(-5),e^(-5)] to these values.
}
\examples{
x <- c(1,2,2,3,4,5,5,5,7)
xbreak=tiebreak(x)
xbreak #a uniform value has been added to the second, sixth and seventh value of x.
length(unique(xbreak))==length(x) #check if the breaking procedure has worked.
y <- c(4,9,12,11,2,10)
c(xbreak,ybreak)=tiebreak(x,y) #a uniform value has been added to the second, third and fifth value of x.
xbreak\%in\%ybreak #check that no values for xbreak can be found in ybreak
}
