% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_speed_of_randomForest_and_ranger.R
\name{timesummary}
\alias{timesummary}
\title{This is to test the speed of randomForest and ranger}
\usage{
timesummary(x, y, z, times = 100, mtry = 5, ntree = 500)
}
\arguments{
\item{x}{x is a data frame of many numeric variables}

\item{y}{y is vector}

\item{z}{is a non- numeric variables matrix}

\item{times}{In microbenchmark}

\item{mtry}{mtry in RF}

\item{ntree}{ntree in RF}
}
\value{
time. it is a list
}
\description{
This is to test the speed of randomForest and ranger
}
