% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jis.R
\name{to_sjis}
\alias{to_sjis}
\title{Given JIS X 0208 Byte Pairs Return Shift JIS Encoding}
\usage{
to_sjis(j12)
}
\arguments{
\item{j12}{two row matrix with JIS byte-pair codes}
}
\value{
two column data frame with S-JIS encodings in \code{\link[=as.hexmode]{as.hexmode()}}
}
\description{
Given JIS X 0208 Byte Pairs Return Shift JIS Encoding
}
