% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc.R
\name{trunc_utf8}
\alias{trunc_utf8}
\alias{trunc_multi}
\alias{trunc_speed}
\title{Test Truncation Functions}
\usage{
trunc_utf8(x)

trunc_multi(x)

trunc_speed(x, n, utf8 = TRUE)
}
\arguments{
\item{x}{scalar character}
}
\value{
a "trunc_test" object
}
\description{
\code{trunc_utf8} assumes UTF8 encoding.  \code{trunc_sjis} requires "native" encoding
and is only useful if that is a multibyte encoding.  Internally it assumes
this is the case.
}
\details{
Will attempt to truncate character from 0 bytes to n+1 bytes, where n is the
number of bytes in the input.
}
