% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conda_S4_helpers.R
\name{UpdateEnvCache}
\alias{UpdateEnvCache}
\title{Handy \code{CondaEnvManger} instance modifier}
\usage{
UpdateEnvCache(
  method = known.methods,
  conda.bin = "auto",
  conda.env = NULL,
  conda.env.is.path = "auto",
  separate.scvi.envs = FALSE,
  overwrite.env = FALSE,
  dry.run = FALSE
)
}
\arguments{
\item{method}{the name of the method to update. One of "combat", "harmony",
"mnn", "bbknn", "scvi", "scanvi", "scanorama"}

\item{conda.bin}{the path to the conda binary. If empty (\code{""}),
\code{NULL} or \code{"conda"}, \code{"auto"} is passed to
\link[reticulate]{conda_binary} to find path to conda binary in \code{PATH}}

\item{conda.env}{Either the name of a conda environment or the path to such
an environment. Must be reachable by provided \code{conda.bin}. \code{NULL}
enable to use default environment names. (see \strong{Details} section)}

\item{conda.env.is.path}{Whether the \code{conda.env} is a path (\code{TRUE}),
or a name (\code{FALSE}). The default ("\code{auto}") guesses. \strong{Be
careful not to make a mistake if you switch the argument value to a
non-default value}}

\item{separate.scvi.envs}{By default, SCVI and SCANVI share the same conda
environment, since they rely on the same python package. If you wish to have
a distinct environment for each, set it to \code{TRUE}. Ignored if
\code{method} is not SCVI nor SCANVI.}

\item{overwrite.env}{Turn it to \code{TRUE} to enable overwriting an existing
environment (same name or same path). When the provided \code{conda.env}
already exists, the default behaviour is to update the package's registry
with the existing environment as it is}

\item{dry.run}{when \code{TRUE}, the package's current cache is not updated.
But the new conda environment (if any) will be created. \code{FALSE} by default.}
}
\value{
a \linkS4class{CondaEnvManager} object
}
\description{
Wrapper to update the \linkS4class{CondaEnvManager} object sotred in cache
and used as configuration for conda environments for all known method
(typically stored in the package's cache). A single method can be
updated at the time. In most cases, this is the best way to modify the
package's conda environments registry.
}
\details{
The conda environments default names are:
\itemize{
  \item{\strong{bbknn}}: {None}
  \item{\strong{SCVI}}: {\code{SeuratIntegrate_scvi-tools}}
  \item{\strong{SCANVI}}: {\code{SeuratIntegrate_scvi-tools}}
  \item{\strong{scanorama}}: {\code{SeuratIntegrate_scanorama}}
}
}
\seealso{
\linkS4class{CondaEnvManager} \link[=CondaManager]{CondaManager()}
}
