% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_package.R
\name{updateReadMe}
\alias{updateReadMe}
\title{function to build and update Readme.md file}
\usage{
updateReadMe()
}
\description{
function to build and update Readme.md file
}
\examples{
updateReadMe()
}
