% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateToday.R
\name{UpdateToday}
\alias{UpdateToday}
\title{Update today}
\usage{
UpdateToday()
}
\value{
list
}
\description{
function for date updating
}
\details{
Okay
}
\note{
function for data updating
}
\examples{

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function () 
{
    library(utils)
    library(httr)
    GET("https://opendata.ecdc.europa.eu/covid19/casedistribution/csv", 
        authenticate(":", ":", type = "ntlm"), write_disk(tf <- tempfile(fileext = ".csv")))
    covid19 <- read.csv(tf)
    countrydata = ByCountryData(covid19)
    return(countrydata)
  }

}
\references{
No particular reference
}
\seealso{
###
}
\author{
E. A. Pena
}
