% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful.R
\name{usefulfun}
\alias{usefulfun}
\title{usefulfun is a function of test package created for fun}
\usage{
usefulfun(x)
}
\arguments{
\item{x}{numeric variable}
}
\description{
usefulfun is a function of test package created for fun
}
