% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_min_max.R
\name{utility_ad_local_min_max}
\alias{utility_ad_local_min_max}
\title{Add local points to charts}
\usage{
utility_ad_local_min_max(df, number_of_days = 20)
}
\arguments{
\item{df}{dataframe cointains financela data}

\item{number_of_days}{how many days to check for min or max}
}
\description{
Add local points to charts
}
