% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_min_max.R
\name{utility_local_point}
\alias{utility_local_point}
\title{Check if a price is a local point}
\usage{
utility_local_point(current_price, surrended_price, searchtype = "min")
}
\arguments{
\item{current_price}{actual price}

\item{surrended_price}{the price before and after}

\item{searchtype}{Are you looking for min or max}
}
\description{
Check if a price is a local point
}
