% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradingview.R
\name{utility_tradingview_data_from_json_string}
\alias{utility_tradingview_data_from_json_string}
\title{Download data from the tradingView site with the passed json string}
\usage{
utility_tradingview_data_from_json_string(json_string)
}
\arguments{
\item{my_json_string}{json string from inspect view at the bottom}
}
\description{
Download data from the tradingView site with the passed json string
}
