% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varClass.R
\name{varClass}
\alias{varClass}
\title{Report variables belonging to different classes}
\usage{
varClass(MVObject)
}
\arguments{
\item{MVObject}{A MUVR object}
}
\value{
A list with names and numbers of variables: all as well as Optimal ('min' model), Redundant (from 'min' up to 'max') and Noisy (the rest)
}
\description{
Reports names and numbers of variables: all as well as Optimal (min model), redundant (from min up to max) and noisy (the rest)
}

