% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_selection_gee.R
\name{variable_selection_gee}
\alias{variable_selection_gee}
\title{Variable Selection for gee Logistic Regression Model Based on AIC...}
\arguments{
\item{data}{A data frame containing the variables for the model.}

\item{response}{The name of the response variable.}

\item{id}{The name of the ID variable.}

\item{family}{The family of the distribution (default is binomial).}

\item{maxit}{The maximum number of iterations for the GEE algorithm (default is 25).}
}
\value{
The best GEE model according to AIC.
}
\description{
This function performs variable selection for a gee logistic regression model using AIC as the selection criteria.
}
\examples{
library(gee)
library(MASS)
data <- data.frame(y = rbinom(100, 1, 0.5), x1 = rnorm(100), x2 = rnorm(100), id = rep(1:10, each = 10))
variable_selection_gee(data, "y", "id")
}
