% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectSamp.R
\name{vectSamp}
\alias{vectSamp}
\title{Sampling of a vector}
\usage{
vectSamp(vect, n = 4, sampLen)
}
\arguments{
\item{vect}{A vector to be sampled into groups}

\item{n}{Number of groups}

\item{sampLen}{A vector with custom number of samples per group. Is calculated if missing (best choice).}
}
\value{
a list with n groups containing sub sampled `vect`
}
\description{
Random sampling of a vector into n groups
Bug: Sampling is faulty when length(vect) < n
}
