\name{voronoi}
\Rdversion{1.1}
\alias{voronoi}

\title{
Voronoi polygons
}

\description{
Voronoi polygons
}

\usage{
voronoi(xy, sp = TRUE, dataframe = FALSE)
}

\arguments{
 \item{xy}{points}
  \item{sp}{Logical}
  \item{dataframe}{Logical}
}

\value{
polygons
}

\author{
Robert J. Hijmans. Adapted from code by Carson Farmer: \url{http://www.carsonfarmer.com/?p=455}
}

\keyword{ spatial  }
