% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_weights.R
\name{weight.bisquare}
\alias{weight.bisquare}
\title{Bisquare weights function}
\usage{
weight.bisquare(dist, bandwidth)
}
\arguments{
\item{dist}{distance vector between observations and fit point}

\item{bandwidth}{cutoff distance (i.e. distance after which weight=0)}
}
\description{
Computes a vector of weights using a bisquare kernel
}
