% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_weights.R
\name{weight.boxcar}
\alias{weight.boxcar}
\title{Box-car weights function}
\usage{
weight.boxcar(dist, bandwidth)
}
\arguments{
\item{dist}{distance vector between observations and fit point}

\item{bandwidth}{bandwidth}
}
\description{
Computes a vector of weights using a boxcar kernel
}
