% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_weights.R
\name{weight.gaussian}
\alias{weight.gaussian}
\alias{weight.gaussian}
\title{Gaussian weights function}
\usage{
weight.gaussian(dist, bandwidth)

weight.gaussian(dist, bandwidth)
}
\arguments{
\item{dist}{distance vector between observations and fit point}

\item{bandwidth}{bandwidth}

\item{dist}{distance vector between observations and fit point}

\item{bandwidth}{bandwidth}
}
\description{
Computes a vector of weights using a bisquare kernel

Computes a vector of weights using a boxcar kernel
}
