% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_main_input.R
\name{wide_to_long}
\alias{wide_to_long}
\title{Transform input file in wide matrix into long matrix format}
\usage{
wide_to_long(input_file)
}
\arguments{
\item{input_file}{input file in wide matrix format}
}
\value{
A data frame containing input data in long-format
}
\description{
Transform input file in wide matrix into long matrix format
}
\examples{
\dontrun{
input_file <- system.file(
    "extdata", "test.main.wide", package = "phyloprofile", mustWork = TRUE
)
wide_to_long(input_file)
}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
