% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{wiki_graph}
\alias{wiki_graph}
\title{Data}
\format{
## dataframe six nodes and eighteen edges and the distancies between them
\describe{
 \item{v1, v2}{nodes of the edges}
 \item{w}{weights of the edges}
}
}
\source{
https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm
}
\usage{
wiki_graph
}
\description{
Example graph with six nodes
}
