% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialAnno.R
\name{wpca}
\alias{wpca}
\title{wpca.}
\usage{
wpca(X, q, weighted = TRUE)
}
\arguments{
\item{X}{a normalized gene expression matrix.}

\item{q}{the number of dimension for low-dimensional embedding}

\item{weighted}{a logical value specifying whether or not to be weighted}
}
\value{
a list.
}
\description{
wpca weighted PCA
}
\details{
weighted PCA
}
