% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{x}
\alias{x}
\title{My test function}
\format{An object of class \code{numeric} of length 100.}
\usage{
x
}
\arguments{
\item{x}{= a random variable}
}
\value{
the output from\code{link{hello.r}}returns a cells(column) by protein (rows) matrix that can be added to the normalized data slot of a S4(e.g. Seurat or SingleCellExperiment) object
}
\description{
My test function
}
\keyword{datasets}
