% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_main_input.R
\name{xml_parser}
\alias{xml_parser}
\title{Parse orthoXML input file}
\usage{
xml_parser(input_file)
}
\arguments{
\item{input_file}{input file in xml format}
}
\value{
A data frame containing input data in long-format
}
\description{
Parse orthoXML input file
}
\examples{
\dontrun{
input_file <- system.file(
    "extdata", "test.main.xml", package = "phyloprofile", mustWork = TRUE
)
xml_parser(input_file)
}
}
\author{
Vinh Tran {tran@bio.uni-frankfurt.de}
}
