% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yy.R
\name{yy}
\alias{yy}
\title{FUNCTION_TITLE}
\usage{
yy(a = 4, b = 2)
}
\arguments{
\item{a}{numeric, set the head to trim from random unif Default: 4}

\item{b}{PARAM_DESCRIPTION, Default: 2}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[utils]{head}}
 \code{\link[stats]{runif}}
}
\author{
Jonathan Sidi
}
