% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore.R
\name{z_score}
\alias{z_score}
\title{This function calculates a z-score based on a single vector}
\usage{
z_score(vec)
}
\arguments{
\item{vec}{a vector of numbers}
}
\value{
z the z-score for each number in the vector
}
\description{
This function calculates a z-score based on a single vector
}
