% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz.R
\name{zz}
\alias{zz}
\title{FUNCTION_TITLE}
\usage{
zz(v = 10, a = 8)
}
\arguments{
\item{v}{PARAM_DESCRIPTION, Default: 10}

\item{a}{PARAM_DESCRIPTION, Default: 8}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[utils]{head}}
 \code{\link[stats]{runif}}
}
\author{
Jonathan Sidi
}
