\name{add}
\alias{add}
\title{
No
}
\description{
No
}
\usage{
add(x, y)
}
\arguments{
  \item{x}{
The first number.
}
  \item{y}{
The second number.
}
}
\details{
No
}
\value{
  \item{jg }{The add value}
  
}
\references{
No
}
\author{
Xiangyong Tan
}
\note{
No
}


\seealso{
No
}
\examples{
\dontrun{
     ## fun(1,2)
  }
}
