% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAccess.R
\name{DataAccess.fi_instrument_get}
\alias{DataAccess.fi_instrument_get}
\title{Get fincancial instruments for the context}
\usage{
DataAccess.fi_instrument_get(da_obj, effective_sched_date = NULL)
}
\arguments{
\item{da_obj}{- The Solvas dataAccess object}

\item{effective_sched_date}{The effective_sched_date to use for schedule data types}
}
\value{
Data frame containing the financial instrument data
}
\description{
Takes a DataAccess da_obj and effective_sched_date and returns a datatable
with the instruments.  Properties that are schedules are coalesced 
to a single value based on the effective_sched_date
}

