% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAccess.R
\name{DataAccess.FiInstrumentGet}
\alias{DataAccess.FiInstrumentGet}
\title{Get financial instruments}
\usage{
DataAccess.FiInstrumentGet(da_obj, effective_date = NULL,
  effective_period = NULL)
}
\arguments{
\item{da_obj}{- Current instance of Solvas|Capital's DataAccess class.}

\item{effective_date}{- The date to use for schedule data types}

\item{effective_period}{- The period to use for schedule data types (1=first period)}
}
\value{
Data frame containing the financial instrument data
}
\description{
Takes a DataAccess da_obj and effective_date or effective_period and returns a DataTable
with the instruments.  Properties that are schedules are coalesced 
to a single value based on the effective_date or effective_period. 
NOTE: EITHER effective_date or effective_period must be populated, the other one must be NULL
}
