% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAccess.R
\name{DataAccess.FSAssumptionsGet}
\alias{DataAccess.FSAssumptionsGet}
\title{Get economic assumptions}
\usage{
DataAccess.FSAssumptionsGet(da_obj, tm_desc = NULL, use_dates = FALSE)
}
\arguments{
\item{da_obj}{- Current instance of Solvas|Capital's DataAccess class.}

\item{tm_desc}{- description of transformation (i.e. '(apply to all)').  the first
transformation by sequence order will be used if NULL, if two or more transformations have the
same description an error message is returned}

\item{use_dates}{- if TRUE then matrix columns will be dates, else periods. default to false}
}
\description{
Get economic assumptions
}

