% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAccess.R
\name{DataAccess.FsInitalAccountBalanceGet}
\alias{DataAccess.FsInitalAccountBalanceGet}
\title{Get initial account balances for the scenario (based on the scenarios entity)
Return a dataframe of all the initial account balances. The dataframe will contain all accounts 
and all dates for the reporting period.  NA will be used when no account balance exists.}
\usage{
DataAccess.FsInitalAccountBalanceGet(da_obj, use_account_name = TRUE)
}
\arguments{
\item{da_obj}{- Current instance of Solvas|Capital's DataAccess class.}

\item{use_account_name}{if TRUE will use account_name, false will use account_number}
}
\description{
Get initial account balances for the scenario (based on the scenarios entity)
Return a dataframe of all the initial account balances. The dataframe will contain all accounts 
and all dates for the reporting period.  NA will be used when no account balance exists.
}
