% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAccess.R
\name{DataAccess}
\alias{DataAccess}
\title{DataAccess: Object to connect and interact with the Capital database}
\usage{
DataAccess(connection_param = "", fs_id_param = NULL)
}
\arguments{
\item{connection_param}{The SQL Server connection string}

\item{fs_id_param}{The scenario ID}
}
\examples{
da <- DataAccess(connection_param=connectionString, fs_id_param=1)
DataAccess.fi_get_instrument(da)
}

