% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdf2D.R
\name{ecdf2D}
\alias{ecdf2D}
\title{Bidimensional Empirical Cumulative Distribution Function}
\usage{
ecdf2D(x, y)
}
\arguments{
\item{x, y}{the two continuous variables. Must be of same length.}
}
\value{
The result is returned as a matrix of dimension (n*n) where the entry (i,j) corresponds to Fn(xi,yj), i=1, ...,n,
j=1, ...,n.
}
\description{
Compute the empirical cumulative distribution function for a bivariate continuous distribution.
}
\details{
The bidimensional e.c.d.f. (empirical cumulative distribution function) Fn is a step function with jumps i/n at observation values, where i is
the number of tied observations at that value.

For observations (x1,y1), ..., (x_n,y_n), Fn is defined as

Fn(t1,t2) = #\code{{xi<=t1,yi<=t2}/n = 1/n sum_{i=1}^n Indicator(xi<=t1,yi<=t2)}
}
\note{
Missing values are removed such that if a value of \code{x} (resp. \code{y}) is missing then the corresponding
values of both \code{x} and \code{y} are removed. The bidimensional e.c.d.f. is then computed on the remaining elements.
}
\examples{
#Simulated data #1
x<-c(0.2, 0.3, 0.1, 0.4)
y<-c(0.5, 0.4, 0.05, 0.2)
ecdf2D(x,y)

#Simulated data #2
n<-40
x<-rnorm(n)
y<-x^2+0.3*rnorm(n)
plot(x,y)
ecdf2D(x,y)
}
\seealso{
\code{\link{robustest}}; the \code{\link{bivariate}} package also provides plots for the
bidimensional e.c.d.f.
}
