% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{f}
\alias{f}
\title{function f is used to find minimum and maximum
of a numeric variable}
\usage{
f(x)
}
\arguments{
\item{x}{numeric variable}
}
\description{
function f is used to find minimum and maximum
of a numeric variable
}
