% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{get-score}
\alias{get-score}
\alias{GetMiscIntegrations}
\alias{GetMiscScores}
\title{Get integration or score name(s) stored in a Seurat object}
\usage{
GetMiscIntegrations(object, search = NULL)

GetMiscScores(object, search = NULL)
}
\arguments{
\item{object}{a Seurat object}

\item{search}{a character vector of names to search for through
case-insensitive exact match. \code{NULL} disables the search (default)}
}
\value{
a character vector of integration or score names, or
\code{NULL} when the object doesn't have 'si_scores' in Misc or when search
did not return any result.
}
\description{
Get (or search for) integration or score names stored in a Seurat object
(in a tibble object named 'si_scores' stored in Misc).
}
