% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwr.sel.par.R
\name{gwr.sel.par}
\alias{gwr.sel.par}
\title{GWR bandwidth selection}
\usage{
gwr.sel.par(formula, data = list(), coords, adapt = FALSE,
  kernel = "gaussian", method = "cv", verbose = TRUE, longlat = NULL,
  RMSE = FALSE, weights, interval_dist = 100, show.error.messages = TRUE,
  ncores = 1, min_dist = NULL, max_dist = NULL)
}
\arguments{
\item{formula}{Formula of the GWR}

\item{data}{Dataset (either data.frame of SpatialDataframe object)}

\item{coords}{A two-columns matrix with the coordinates as X-Y if data is a data.frame}

\item{adapt}{Logical. TRUE if Adaptative bandwith, FALSE if fixed}

\item{kernel}{Character string. Weight kernel. Either "gaussian" or "bisquare"}

\item{method}{Validation method. So far, only the cross-validation approach is implemented}

\item{longlat}{TRUE if coordinates are longitude-latitude in decimal degrees, in which case, distances are measured in kilometers}

\item{interval_dist}{Minimum distance between two separate runs of the optimize}

\item{min_dist}{Minimum bandwith}

\item{max_dist}{Maximum bandwith}
}
\value{
A bandwidth
}
\description{
This function computes the optimal bandwidth for a GWR (Geographically Weighted Regression). It spans the computation across several nodes in a cluster.
}
