% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{First line is title}
\usage{
hello()
}
\value{
The value returned by this function.
}
\description{
This is a brief description.
}
\details{
Anything else after the description goes to the Details section.

You can write several paragraphs.
}
\examples{
hello()



}
\references{
\url{https://github.com/yihui/rmini}
}
\seealso{
\code{\link[tools]{file_ext}}, \code{\link[tools]{file_path_sans_ext}}
}
\author{
Who are you?
}
