% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matrix-symmetrize}
\alias{matrix-symmetrize}
\alias{symmetrize.pmax.sparse}
\alias{symmetrize.pmin.sparse}
\title{Matrix symmetrizing}
\usage{
symmetrize.pmax.sparse(i, j, x, height)

symmetrize.pmin.sparse(i, j, x, height)
}
\arguments{
\item{i}{row indices (1-based)}

\item{j}{column indices (1-based)}

\item{x}{values such as \code{m[i,j] = x}}

\item{height}{height of the matrix (number of rows)}

\item{by}{matrix to sort by. If \code{NULL}, sort \code{mat} by itself (default)}
}
\value{
a symmetric sparse dgCMatrix of size height x height
}
\description{
Create a sparse matrix using i, j and x (row indices, column indicies and
values respectively) then symmetrizes it with low memory footprint\cr
\cr
\code{symmetrize.pmax.sparse} symmetrize matrix with max(m[i,j], m[j,i])\cr
\code{symmetrize.pmin.sparse} symmetrize matrix with min(m[i,j], m[j,i])
}
\note{
If \code{mat} and \code{by} have different dimensions, results might be
incorrect
}
