% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutationPlot.R
\name{permutationPlot}
\alias{permutationPlot}
\title{Plot permutation analysis using actual model and permutations()}
\usage{
permutationPlot(MVObj, permMatrix, model = "Mid", type = c("t", "non"),
  side = c("greater", "smaller"), pos, xlab = NULL, xlim, ylim = NULL,
  breaks = "Sturges", main = NULL)
}
\arguments{
\item{MVObj}{A MUVR object}

\item{permMatrix}{A permutation fitness outcome from MUVR::permutations()}

\item{model}{'Min', 'Mid', or 'Max'}

\item{type}{'t' (default; for Student's t) or 'non' for "non-parametric" (i.e. rank) studen'ts}

\item{side}{'smaller' for actual lower than H0 or 'greater' for actual larger than H0 (automatically selected if not specified)}

\item{pos}{which side of actual to put p-value on}

\item{xlab}{optional xlabel}

\item{xlim}{optional x-range}

\item{ylim}{otional y-range}

\item{breaks}{optional custom histogram breaks (defaults to 'sturges')}

\item{main}{optional plot title (or TRUE for autoname)}
}
\value{
A permutation plot
}
\description{
This is asically a wrapper for the MUVR::plotPerm() function using model objects to make coding nicer and cleaner
}
\examples{
library(MUVR)
library(doParallel)
nCore=detectCores()-1
cl=makeCluster(nCore)
registerDoParallel(cl)
nRep=2*nCore
varRatio=.75
nOuter=6
nPerm=50
R12ML=MUVR(X=mlr12,ML=T,nRep=nRep,nOuter=nOuter,varRatio=varRatio,method='RF')
permR12=permutations(R12ML)
stopCluster(cl)
permutationPlot(R12ML,permR12)
}
