% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutations.R
\name{permutations}
\alias{permutations}
\title{Make permutations with data and default settings from an actual MUVR object}
\usage{
permutations(MVObj, nPerm = 50, nRep, nOuter, varRatio, parallel)
}
\arguments{
\item{MVObj}{A MUVR obvject}

\item{nPerm}{number of permutations to run}

\item{nRep}{number of repetitions for each permutation (defaults to value of actual model)}

\item{nOuter}{number of outer validation segments for each permutation (defaults to value of actual model)}

\item{varRatio}{varRatio for each permutation (defaults to value of actual model)}

\item{parallel}{whether to run calculations using parallel processing - which requires registered backend (defaults to value of actual model)}
}
\value{
A permutation matrix with permuted fitness statistics (nrow=nPerm and ncol=3 for min/mid/max)
}
\description{
Make permutations with data and default settings from an actual MUVR object
}
\examples{
library(MUVR)
library(doParallel)
nCore=detectCores()-1
cl=makeCluster(nCore)
registerDoParallel(cl)
nRep=2*nCore
varRatio=.75
nOuter=6
nPerm=50
R12ML=MUVR(X=mlr12,ML=T,nRep=nRep,nOuter=nOuter,varRatio=varRatio,method='RF')
permR12=permutations(R12ML)
stopCluster(cl)
permutationPlot(R12ML,permR12)
}
