% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMV.R
\name{plotMV}
\alias{plotMV}
\title{Plot MV object}
\usage{
plotMV(MVObj, model = "mid")
}
\arguments{
\item{MVObj}{An `MVobject` obtained from the MVWrap function}

\item{model}{What type of model to plot ('min', 'mid' or 'max'). Defaults to 'mid'.}
}
\value{
A pdf with plots of results from multivariate predictions
}
\description{
Plot MV object
}
