% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{PlotScores}
\alias{PlotScores}
\title{Visualise and compare the performances of integration algorithms}
\usage{
PlotScores(
  object,
  plot.type = c("dot", "radar", "lollipop"),
  split.by.score.type = TRUE,
  order.by = c("score", "name", "asis"),
  include.integration = NULL,
  exclude.integration = NULL,
  include.score = NULL,
  exclude.score = NULL,
  recompute.overall.scores = TRUE,
  rescale = TRUE,
  batch.coeff = 0.4,
  bio.coeff = 0.6,
  point.max.size = 20L,
  use.ggforce = is_installed("ggforce")
)
}
\arguments{
\item{object}{a Seurat object}

\item{plot.type}{one of 'table' (default), 'radar' or 'lollipop'. Type of
desired plot}

\item{split.by.score.type}{whether to split scores by type (bio-conservation,
batch correction and overall scores). When set to \code{FALSE}, all scores
are mixed in a single figure.}

\item{order.by}{one of 'score' (default), 'name' or 'asis'. Determines the
order of integrations in the legend (and on the y-axis for lolliplop and
table plots). Scores are ordered by decreasing overall score by default,
by name or by row-order when setting 'name' and 'asis' respectively.}

\item{include.integration}{name of the integration(s) to include. The
default value (\code{NULL}) enable to include them all.}

\item{exclude.integration}{name of the integration(s) to exclude. The default
value (\code{NULL}) enable to include them all.}

\item{include.score}{name of the score(s) to include. The default value
(\code{NULL}) enable to include them all.}

\item{exclude.score}{name of the score(s) to exclude. The default value
(\code{NULL}) enable to include them all.}

\item{recompute.overall.scores}{whether to recompute overall scores. Useful
when there is a restriction on scores to plot. When \code{FALSE},
coefficient parameters have no impact.}

\item{rescale}{whether to rescale each score between 0 and 1 using min-max
normalisation before computing overall scores. This ensures that each metric
equally contributes to the overall scores. Has no effect when
\code{recompute.overall.scores = FALSE}. \code{TRUE by default}}

\item{batch.coeff}{the weight of batch correction performance evaluation
scores in the overall score.}

\item{bio.coeff}{the weight of bio-conservation performance evaluation scores
in the overall score.}

\item{point.max.size}{inoperative unless \code{plot.type = 'table'} and
\code{use.ggforce = FALSE}. Determine the maximum size of the points
(only achieved for a score of 1) to fit the plotting area (handled
automaticaly when ggforce is used).}

\item{use.ggforce}{for \code{plot.type = 'table'}, enable or disable the use
of \pkg{ggforce} to draw the circles. Used by default when the package is
installed}
}
\value{
a ggplot object
}
\description{
Plot the scaled integration scores to compare the obtained integrations
}
