% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVAL.R
\name{plotVAL}
\alias{plotVAL}
\title{Plot validation metric}
\usage{
plotVAL(MUVRclassObject)
}
\arguments{
\item{MUVRclassObject}{An object of class `MUVR`}
}
\value{
A plot
}
\description{
Produces a plot of validation metric vs number of variables in model (inner segment)
}
