% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsda.R
\name{plsda}
\alias{plsda}
\title{PLS-DA}
\usage{
plsda(
  x,
  y,
  ncomp = 2,
  max.iter = 500,
  tol = 1e-06,
  near.zero.var = TRUE,
  scale = TRUE
)
}
\arguments{
\item{x}{}

\item{y}{}

\item{ncomp}{}

\item{max.iter}{}

\item{tol}{}

\item{near.zero.var}{}

\item{scale}{}
}
\value{
plsda object
}
\description{
Adapted and stripped down from mixOmics v 5.2.0 (https://cran.r-project.org/web/packages/mixOmics/)
}
