% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus.R
\name{plus}
\alias{plus}
\title{Easy function to plus two numbers}
\usage{
plus(x, y)
}
\arguments{
\item{x}{- first element to be added \code{x}}

\item{y}{- second element to be added}
}
\value{
the result of summation
}
\description{
Easy function to plus two numbers
}
\examples{
plus(1,2)
}
