% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.plsdaMUVR}
\alias{predict.plsdaMUVR}
\title{Predict plsda}
\usage{
\method{predict}{plsdaMUVR}(object, newdata, onlyPred = FALSE, scale = TRUE,
  ...)
}
\arguments{
\item{object}{a plsdaMUVR object}

\item{newdata}{new data}

\item{onlyPred}{Boolean for whether to report back predictions only (defaults to FALSE)}

\item{...}{}
}
\value{
plsda predictions
}
\description{
Adapted and stripped down from mixOmics v 5.2.0 (https://cran.r-project.org/web/packages/mixOmics/)
}
