% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.plsMUVR}
\alias{predict.plsMUVR}
\title{Predict pls}
\usage{
\method{predict}{plsMUVR}(object, newdata, ...)
}
\arguments{
\item{object}{}

\item{newdata}{}

\item{...}{}
}
\value{
pls prediction
}
\description{
Adapted and stripped down from mixOmics v 5.2.0 (https://cran.r-project.org/web/packages/mixOmics/)
}
