% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdCVnet_getVar.R
\name{rdCVnet_getVar}
\alias{rdCVnet_getVar}
\title{the function to get the variables selected}
\usage{
rdCVnet_getVar(
  rdCVnetObject,
  option = c("quantile", "smoothcurve"),
  span = 1,
  outlier = T,
  percent_smoothcurve = 0.05,
  percent_quantile = 0.25
)
}
\arguments{
\item{rdCVnetObject}{a object from rdCVnet}

\item{option}{quantile or smoothcurve}

\item{span}{for smooth curve:  how smooth the curve need to be}

\item{outlier}{if remove ourlier variables or not}

\item{percent_smoothcurve}{If select_variables_by smoothcurve, then it is robust}

\item{percent_quantile}{range from 0 to 0.5. When select_variables_by quantile, this value represent the first quantile.}
}
\value{
a rdCVnet object
}
\description{
the function to get the variables selected
}
