% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustest.R
\name{robustest}
\alias{robustest}
\title{Robust independence test for two continuous variables of Kolmogorov-Smirnov's type}
\usage{
robustest(x, y, N = 50000, simu = FALSE)
}
\arguments{
\item{x, y}{the two continuous variables. Must be of same length.}
}
\value{
Returns the result of the test with its corresponding p-value and the value of the test statistic.
}
\description{
Test the independence between two continuous variables based on the maximum distance
between the joint empirical cumulative distribution function and the product of the marginal
empirical cumulative distribution functions.
}
\details{
For two continuous variables, robustest tests H0 X and Y are independent
against H1 X and Y are not independent.

For observations (x1,y1), ..., (x_n,y_n), the bivariate e.c.d.f.
(empirical cumulative distribution function) Fn is defined as

Fn(t1,t2)=#\code{{xi<=t1,yi<=t2}/n= sum_{i=1}^n Indicator(xi<=t1,yi<=t2)/n}

Let Fn(t1) and Fn(t2) be the marginals e.c.d.f. The test statistics is defined as:

\code{n^(1/2) sup_{t1,t2} |Fn(t1,t2)-Fn(t1)*Fn(t2)|}

Under H0 the distribution of the test statistic is free and is equivalent to
the same test statistic computed for two continuous uniform variables on [0,1],
where the supremum is taken for t1,t2 on [0,1]. Using this result, the distribution of the test
statistic is obtained using Monte-Carlo simulations.
}
\note{
Only a two sided alternative is possible with this test.
}
\examples{
#Simulated data 1
x<-c(0.2, 0.3, 0.1, 0.4)
y<-c(0.5, 0.4, 0.05, 0.2)
stat_robustest(x,y)
#Simulated data 2
n<-40
x<-rnorm(n)
y<-x^2+0.3*rnorm(n)
plot(x,y)
stat_robustest(x,y)
#Real data
}
\seealso{
\code{\link{cortest}}, \code{\link{vartest}}, \code{\link{mediantest}}, \code{\link{wilcoxtest}}.
}
\keyword{test}
