% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{ScaleScores}
\alias{ScaleScores}
\title{Scale the scores in the score tibble to plot them}
\usage{
ScaleScores(object, ref = "Unintegrated", batch.coeff = 0.4, bio.coeff = 0.6)
}
\arguments{
\item{object}{a Seurat object}

\item{ref}{the name of the integration to use as a reference for scaling.
Useful for PCA regression (and density) and cell cycle conservation scores.}

\item{batch.coeff}{the weight of batch correction performance evaluation
scores in the overall score.}

\item{bio.coeff}{the weight of bio-conservation performance evaluation scores
in the overall score.}

\item{integration}{the name of the integration for which the score was
computed.}
}
\description{
Once the scores of interest have been computed and saved in the score tibble,
they can be scaled to make them comparable by bounding them between 0 and 1
and harmonise their direction (0 and 1 always mean bad and good performance
respectively). This is also a prerequisite for plotting.
}
