% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_pca.R
\name{ScoreRegressPC}
\alias{ScoreRegressPC}
\title{Score a corrected or uncorrected PCA to estimate batch mixing}
\usage{
ScoreRegressPC(
  object,
  batch.var = NULL,
  reduction = "pca",
  dims = NULL,
  adj.r2 = FALSE,
  ...
)
}
\arguments{
\item{object}{A Seurat object}

\item{batch.var}{The name of the batch variable (must be in the object metadata)}

\item{reduction}{The name of the reduction to score}

\item{dims}{The dimensions to consider. All dimensions are used by default}

\item{adj.r2}{Whether to use the adjusted R2 instead of the raw R2}

\item{...}{Additional parameters ('assay', 'search' and 'scale.layer') to pass
to Seurat to automatically construct the \code{batch.var} when not provided.}
}
\value{
A single float corresponding to the score of the given reduction
}
\description{
Linearly regresses principal components with batch variable as a proxy to
estimate batch mixing. The resulting R2 are then weighted by each dimension's
contribution to variance.
}
\details{
The linear regression is
\deqn{PC_i = Batch}


The score is computed as follow :
\deqn{\sum_{i=1}^{p} \left ( R^2_i * V_i \right )}

For a PCA with p dimensions, \eqn{PC_i} is the principal component i,
\eqn{R^2_i} is the R squared coefficient of the linear regression for the
dimension i. \eqn{V_i} is the proportion of variance explained by the
\eqn{PC_i}.
}
\note{
This score is an adaptation of the principal component regression (PCR)
score from Luecken M.D. \emph{et al.}, 2022.
}
\examples{
\dontrun{
obj <- SeuratData::LoadData("pbmcsca")
obj[["RNA"]] <- split(obj[["RNA"]], f = obj$Method)
obj <- NormalizeData(obj)
obj <- FindVariableFeatures(obj)
obj <- ScaleData(obj)
obj <- RunPCA(obj)

score.r2 <- ScoreRegressPC(obj, "Method", "pca", dim = 1:30)
score.adj.r2 <- ScoreRegressPC(obj, "Method", "pca", dim = 1:30, adj.r2 = TRUE)

score.r2    # ~ 0.1147
score.adj.r2     # ~ 0.1145
}

}
\references{
Luecken, M. D., Büttner, M., Chaichoompu, K., Danese, A.,
Interlandi, M., Mueller, M. F., Strobl, D. C., Zappia, L., Dugas, M.,
Colomé-Tatché, M. & Theis, F. J. Benchmarking atlas-level data integration in
single-cell genomics. Nat Methods 19, 41–50 (2021).
\href{https://doi.org/10.1038/s41592-021-01336-8}{DOI}
}
\seealso{
\code{\link[Seurat]{Layers}} and \code{\link[Seurat]{DefaultAssay}}
for \code{...} arguments, \code{\link{ScoreDensityPC}} for an alternative and
\code{\link{ScoreRegressPC.CellCycle}} to regresses PCs by cell cycle scores.
}
