% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speed_and_performance.R
\name{speed_and_performance}
\alias{speed_and_performance}
\title{Speeding and performance test for different models}
\usage{
speed_and_performance(
  X,
  Y,
  ID,
  scale = TRUE,
  nRep = 5,
  nOuter = 6,
  nInner,
  varRatio = 0.75,
  DA = FALSE,
  fitness = c("AUROC", "MISS", "BER", "RMSEP"),
  method = c("PLS", " RF"),
  methParam,
  ML = FALSE,
  modReturn = T,
  logg = FALSE,
  parallel = TRUE,
  repeatMUVR = 2,
  MUVRversion = c("MUVR", "MUVRoriginal")
)
}
\arguments{
\item{X}{X in MUVR}

\item{Y}{Y in MUVR}

\item{ID}{ID in MUVR}

\item{scale}{scale in MUVR}

\item{nRep}{nRep in MUVR}

\item{nOuter}{nOuter in MUVR}

\item{nInner}{nInner in MUVR}

\item{varRatio}{= 0.75 varRatio in MUVR}

\item{DA}{DA in MUVR}

\item{fitness}{fitness in MUVR}

\item{method}{method in MUVR}

\item{methParam}{methParam in MUVR}

\item{ML}{ML in MUVR}

\item{modReturn}{modReturn in MUVR}

\item{logg}{logg in MUVR}

\item{parallel}{parallel X in MUVR}

\item{repeatMUVR}{This is to decide how many MUVR to repeat(since each MUVR may give a differnent value)}

\item{MUVRversion}{"MUVR" or "MUVRoriginal"}
}
\value{
time   average time for each MUVR

performance   average miss auc and rmsep
}
\description{
Speeding and performance test for different models
}
