% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma_dos_numeros.R
\name{suma.dos.numeros}
\alias{suma.dos.numeros}
\title{Add two numbers}
\usage{
suma.dos.numeros(x, y)
}
\arguments{
\item{x}{a number}

\item{y}{a number}
}
\value{
\code{x} + \code{y}
}
\description{
Add two numbers
}
\examples{
suma.dos.numeros(1,1)
suma.dos.numeros(10,1)
}

