% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-package.R
\docType{package}
\name{test-package}
\alias{test}
\alias{test-package}
\title{test: I would edit if my computer worked}
\description{
Fun stats stuff
}
\author{
\strong{Maintainer}: Braeden Falzarano \email{first.last@example.com}

}
\keyword{internal}
