\name{test-package}
\alias{test-package}
\alias{test}
\docType{package}
\title{
\packageTitle{test}
}
\description{
\packageDescription{test}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{test}
\packageIndices{test}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{test}

Maintainer: \packageMaintainer{test}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
