\name{test-package}
\alias{test-package}
\alias{test}
\docType{package}
\title{
\packageTitle{test}
}
\description{
\packageDescription{test}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{test}

Maintainer: \packageMaintainer{test}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = test)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{test}}
%\section{Documentation Index}{\packageIndices{test}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
