% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-package.R
\docType{package}
\name{test-package}
\alias{test}
\alias{test-package}
\title{test: Euclidean and Dijkstra Algorithms in R}
\description{
A package that allows users to use a version of the Euclidean and Dijkstra algorithms to find the greatest common divisor of two numbers or the shortest distances between two vectors in a graph.
}
\keyword{internal}
