\name{Test-package}
\alias{Test-package}
\alias{Test}
\docType{package}
\title{
\packageTitle{Test}
}
\description{
\packageDescription{Test}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Test}
\packageIndices{Test}
~~ An overview of how to use the package, ~~
~~ including the most important functions ~~
}
\author{
\packageAuthor{Test}

Maintainer: \packageMaintainer{Test}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per ~~
~~ line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
