\name{test-package}
\alias{test-package}
\alias{test}
\docType{package}
\title{test


}
\description{
ciao
}
\details{
\tabular{ll}{
Package: \tab test\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-10-04\cr
License: \tab What license is it under?\cr
}

}
\author{
Who wrote it


}
\references{

}

\keyword{ package }
\seealso{

}
\examples{

}
