\name{test-package}
\Rdversion{1.1}

\alias{test-package}

\alias{test}

\docType{package}
\title{What the package does (short line)}

\description{
More about what it does (maybe more than one line)
}

\details{
\tabular{ll}{
Package: \tab test\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-03-06\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}

}
\author{
Rober J. Hijmans

Maintainer: <r.hijmans@gmail.com>
}

\keyword{ package }
\keyword{ spatial }

\seealso{
\code{\link[<sp>:<sp>-package]{<sp>}}
}

