% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-package.R
\docType{package}
\name{test-package}
\alias{test}
\alias{test-package}
\title{test: This Is A Test Package}
\description{
This package will include raw data from a paper describing women with temporomandibular disorders. This was an untargeted proton nuclear magnetic resonance spectroscopy-based metabolomics study. The group used saliva as their sample.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stanciua19/test}
  \item Report bugs at \url{https://github.com/stanciua19/test/issues}
}

}
\author{
\strong{Maintainer}: Ana Stanciu \email{stanciua@mcmaster.ca}

}
\keyword{internal}
