% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-package.R
\docType{package}
\name{test-package}
\alias{test}
\alias{test-package}
\title{test: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: YU LUO \email{first.last@example.com} (YOUR-ORCID-ID)

}
\keyword{internal}
