% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Wiki_graph}
\format{
A dataframe with 3 variables and 18 rows:
\describe{
\item{v1}{Node 1 of al edges}
\item{V2}{Node 2 of all edges}
\item{w}{Node 3 of all edges}
}
}
\usage{
wiki_graph
}
\description{
A dataset containg nodes and edge details of the a pgraph
}
\keyword{datasets}
